% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signals.R
\name{abort_on_invalid_user_action}
\alias{abort_on_invalid_user_action}
\title{Constructs a helpful message and signals an error when the user takes an
invalid action. Can include an optional hint to fix the error.}
\usage{
abort_on_invalid_user_action(
  cause_string,
  hint_string = NULL,
  call = rlang::caller_env(),
  ...
)
}
\arguments{
\item{cause_string}{Why the action was invalid.}

\item{hint_string}{An optional hint to fix the error.}

\item{call}{The function environment that is most relevant from the user's
perspective. Default is \code{rlang::caller_env()}, which gives the environment
attached to the function that called this signal function.}

\item{...}{Passed to \code{simChef::abort()}.}
}
\description{
Constructs a helpful message and signals an error when the user takes an
invalid action. Can include an optional hint to fix the error.
}
\keyword{internal}
