% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-lib-utils.R
\name{add_all_group}
\alias{add_all_group}
\title{Add ".all" group to data}
\usage{
add_all_group(data, group_cols)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{group_cols}{A character string or vector identifying the
column(s) to group observations by before evaluating metrics.
This is useful for assessing within-group prediction errors.}
}
\value{
A \code{data.frame} with twice the number of rows as \code{data}.
In the second half of this \code{data.frame}, the columns specified by
\code{group_cols} have been set to ".all" (or some variant thereof if
".all" already appears in the column).
}
\description{
A helper function to append ".all" group to data in order to
conveniently compute (unstratified) global metrics on the full dataset in
addition to within-group metrics, specified by \code{group_cols}.
}
\keyword{internal}
