% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{add_funs}
\alias{add_funs}
\alias{add_dgp}
\alias{add_method}
\alias{add_evaluator}
\alias{add_visualizer}
\title{Helper functions for adding components to an \code{Experiment}.}
\usage{
add_dgp(experiment, dgp, name = NULL, ...)

add_method(experiment, method, name = NULL, ...)

add_evaluator(experiment, evaluator, name = NULL, ...)

add_visualizer(experiment, visualizer, name = NULL, ...)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{dgp}{A \code{DGP} object.}

\item{name}{A name to identify the object to be added.}

\item{...}{Not used.}

\item{method}{A \code{Method} object.}

\item{evaluator}{An \code{Evaluator} object.}

\item{visualizer}{A \code{Visualizer} object.}
}
\value{
The original \code{Experiment} object passed to \code{add_*}.
}
\description{
Helper functions for adding \code{DGPs}, \code{Methods},
\code{Evaluators}, and \code{Visualizers} to an \code{Experiment}.
}
