% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-lib-errors.R
\name{ar1_errors}
\alias{ar1_errors}
\title{Generate autoregressive Gaussian errors.}
\usage{
ar1_errors(n, rho)
}
\arguments{
\item{n}{Number of observations.}

\item{rho}{Correlation.}
}
\value{
A vector of simulated errors of length n.
}
\description{
Generate correlated Gaussian errors based on an
autoregressive(1) covariance structure.
}
