% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer-lib-ggthemes.R
\name{blank_ggplot_themes}
\alias{blank_ggplot_themes}
\alias{blank_x_theme}
\alias{blank_y_theme}
\alias{blank_xy_theme}
\title{Blank axis themes for ggplot objects.}
\usage{
blank_x_theme()

blank_y_theme()

blank_xy_theme()
}
\value{
A ggplot theme object.
}
\description{
Blank axis themes for ggplot objects.
}
\examples{
require(ggplot2)

## blank x-axis theme
ggplot(iris) +
  aes(x = Sepal.Length, fill = Species) +
  geom_density() +
  blank_x_theme()
  
## blank y-axis theme
ggplot(iris) +
  aes(x = Sepal.Length, fill = Species) +
  geom_density() +
  blank_y_theme()
  
## blank x- and y-axes theme
ggplot(iris) +
  aes(x = Sepal.Length, fill = Species) +
  geom_density() +
  blank_xy_theme()
  
}
