% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer-lib-ggthemes.R
\name{blank_x_theme}
\alias{blank_x_theme}
\title{Blank x-axis theme for ggplot objects.}
\usage{
blank_x_theme()
}
\value{
A ggplot theme object.
}
\description{
Blank x-axis theme for ggplot objects.
}
\examples{
require(ggplot2)
ggplot(iris) +
  aes(x = Sepal.Length, fill = Species) +
  geom_density() +
  blank_x_theme()
  
}
