% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-lib-errors.R
\name{block_errors}
\alias{block_errors}
\title{Generate block-correlated Gaussian errors.}
\usage{
block_errors(n, rho = 0.8)
}
\arguments{
\item{n}{Number of observations.}
}
\value{
A vector of simulated errors of length n.
}
\description{
Generate correlated Gaussian errors based on a block-dependence
covariance structure (with 3 blocks).
}
