% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-lib-utils.R
\name{class_metric_vec_constructor}
\alias{class_metric_vec_constructor}
\title{Helper function for constructing \code{metric_vec} methods for
custom \code{yardstick} metrics.}
\usage{
class_metric_vec_constructor(
  name,
  fun,
  truth,
  estimate,
  estimator,
  na_rm,
  case_weights,
  event_level,
  ...
)
}
\description{
Helper function for constructing \code{metric_vec} methods for
custom \code{yardstick} metrics.
}
\keyword{internal}
