% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-utils.R
\name{clean_up_worker_env}
\alias{clean_up_worker_env}
\title{Clean up \code{future} worker-local environments on exit.}
\usage{
clean_up_worker_env(what = c("future", "dgp", "method"), env = parent.frame())
}
\description{
Clean up \code{future} worker-local environments on exit.
}
\keyword{internal}
