% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear cached results from \code{Experiment}.}
\usage{
clear_cache(experiment)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}
}
\value{
The original \code{Experiment} object with cache cleared.
}
\description{
Clear (or delete) cached results from an \code{Experiment} to
start the experiment fresh/from scratch.
}
