% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_tibble_rows}
\alias{compare_tibble_rows}
\title{Compare two tibbles.}
\usage{
compare_tibble_rows(x, y, op = c("equal", "contained_in"))
}
\arguments{
\item{x}{A tibble with unique rows.}

\item{y}{A tibble with unique rows.}

\item{op}{Name of opertaion.}
}
\value{
If \code{op == "equal"}, returns a boolean indicating if \code{x} and
\code{y} have the same rows, ignoring the row order. If
\code{op == "contained_in"}, returns a boolean indicating if all rows in
\code{x} are contained in the rows of \code{y}.
}
\description{
Compare two tibbles for equality or containment.
}
\keyword{internal}
