% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-utils.R
\name{compute_dgp}
\alias{compute_dgp}
\title{Distribute simulation computation by DGPs.}
\usage{
compute_dgp(reps, future.globals, future.packages, future.seed, ...)
}
\description{
Distribute simulation computation by DGPs.
}
\keyword{internal}
