% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-utils.R
\name{compute_dgp_method}
\alias{compute_dgp_method}
\title{Doubly nested distributed simulation computation nested by DGPs and Methods.}
\usage{
compute_dgp_method(n_reps, future.globals, future.packages, future.seed, ...)
}
\description{
Doubly nested distributed simulation computation nested by DGPs and Methods.
}
\keyword{internal}
