% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-utils.R
\name{compute_dgp_rep}
\alias{compute_dgp_rep}
\title{Doubly nested distributed simulation computation nested by DGPs and reps.}
\usage{
compute_dgp_rep(n_reps, future.globals, future.packages, future.seed, ...)
}
\description{
Doubly nested distributed simulation computation nested by DGPs and reps.
}
\keyword{internal}
