% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-utils.R
\name{compute_method}
\alias{compute_method}
\title{Distribute simulation computation by Methods.}
\usage{
compute_method(n_reps, future.globals, future.packages, future.seed, ...)
}
\description{
Distribute simulation computation by Methods.
}
\keyword{internal}
