% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-utils.R
\name{compute_method_rep}
\alias{compute_method_rep}
\title{Doubly nested distributed simulation computation nested by Methods and reps.}
\usage{
compute_method_rep(n_reps, future.globals, future.packages, future.seed, ...)
}
\description{
Doubly nested distributed simulation computation nested by Methods and reps.
}
\keyword{internal}
