% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-utils.R
\name{compute_rep}
\alias{compute_rep}
\title{Distribute simulation computation by replicates.}
\usage{
compute_rep(reps, future.globals, future.packages, future.seed, ...)
}
\description{
Distribute simulation computation by replicates.
}
\keyword{internal}
