% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-lib.R
\name{correlated_linear_gaussian_dgp}
\alias{correlated_linear_gaussian_dgp}
\title{Generate correlated Gaussian covariates and linear response data.}
\usage{
correlated_linear_gaussian_dgp(
  n,
  p_uncorr,
  p_corr,
  s_uncorr = p_uncorr,
  s_corr = p_corr,
  corr,
  betas_uncorr = NULL,
  betas_corr = NULL,
  betas_uncorr_sd = 1,
  betas_corr_sd = 1,
  err = NULL,
  data_split = FALSE,
  train_prop = 0.5,
  return_values = c("X", "y", "support"),
  ...
)
}
\arguments{
\item{n}{Number of samples.}

\item{p_uncorr}{Number of uncorrelated features.}

\item{p_corr}{Number of features in correlated group.}

\item{s_uncorr}{Number of features in uncorrelated group with non-zero coef.}

\item{s_corr}{Number of features in correlated group with non-zero coef.}

\item{corr}{Correlation between features in correlated group.}

\item{betas_uncorr}{Coefficient vector for uncorrelated features.}

\item{betas_corr}{Coefficient vector for correlated features.}

\item{betas_uncorr_sd}{(Optional) SD of normal distribution from which to
draw \code{betas_uncorr}. Only used if \code{betas_uncorr} argument is
\code{NULL}.}

\item{betas_corr_sd}{(Optional) SD of normal distribution from which to draw
\code{betas_corr}. Only used if \code{betas_corr} argument is
\code{NULL}.}

\item{err}{Function from which to generate simulate error vector. Default
is \code{NULL} which adds no error to the DGP.}

\item{data_split}{Logical; if \code{TRUE}, splits data into training and test
sets according to \code{train_prop}.}

\item{train_prop}{Proportion of data in training set if
\code{data_split = TRUE}.}

\item{return_values}{Character vector indicating what objects to return in
list. Elements in vector must be one of "X", "y", "support".}

\item{...}{Other arguments to pass to err() to generate the error vector.}
}
\value{
A list of the named objects that were requested in
\code{return_values}. Note that if \code{data_split = TRUE} and "X", "y"
are in \code{return_values}, then the returned list also contains slots for
"Xtest" and "ytest".
}
\description{
Generate normally-distributed covariates that are potentially
correlated and linear response data with a specified error distribution.
}
