% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-lib.R
\name{correlated_lss_gaussian_dgp}
\alias{correlated_lss_gaussian_dgp}
\title{Generate correlated Gaussian covariates and LSS response data.}
\usage{
correlated_lss_gaussian_dgp(
  n,
  p_uncorr,
  p_corr,
  s_uncorr = p_uncorr,
  s_corr = p_corr,
  corr,
  k,
  thresholds = 0,
  signs = 1,
  betas = 1,
  intercept = 0,
  overlap = FALSE,
  err = NULL,
  data_split = FALSE,
  train_prop = 0.5,
  return_values = c("X", "y", "support"),
  ...
)
}
\arguments{
\item{n}{Number of samples.}

\item{p_uncorr}{Number of uncorrelated features.}

\item{p_corr}{Number of features in correlated group.}

\item{s_uncorr}{Number of features in uncorrelated group in support.}

\item{s_corr}{Number of features in correlated group in support.}

\item{corr}{Correlation between features in correlated group.}

\item{k}{Order of the interactions.}

\item{thresholds}{A scalar or a s x k matrix of the thresholds for each term
in the LSS model.}

\item{signs}{A scalar or a s x k matrix of the sign of each interaction
(1 means > while -1 means <).}

\item{betas}{Scalar or parameter vector for interaction terms.}

\item{intercept}{Scalar intercept term.}

\item{overlap}{If TRUE, simulate support indices with replacement; if FALSE,
simualte support indices without replacement (so no overlap).}

\item{err}{Function from which to generate simulate error vector. Default
is \code{NULL} which adds no error to the DGP.}

\item{data_split}{Logical; if \code{TRUE}, splits data into training and test
sets according to \code{train_prop}.}

\item{train_prop}{Proportion of data in training set if
\code{data_split = TRUE}.}

\item{return_values}{Character vector indicating what objects to return in
list. Elements in vector must be one of "X", "y", "support".}

\item{...}{Other arguments to pass to err() to generate the error vector.}
}
\description{
Generate normally-distributed covariates that are potentially
correlated and LSS response data with a specified error distribution.
}
