% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-templates.R
\name{create_assign_str}
\alias{create_assign_str}
\title{Create an assignment string and print to console.}
\usage{
create_assign_str(name, value, throw_error = FALSE)
}
\description{
Create an assignment string and print to console.
}
\keyword{internal}
