% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{create_doc_template}
\alias{create_doc_template}
\title{Initialize documentation template for the R Markdown results report.}
\usage{
create_doc_template(experiment, save_dir)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{create_doc_template()} was renamed to \code{init_docs()} to create a more
consistent API.
}
\keyword{internal}
