% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{create_experiment}
\alias{create_experiment}
\title{Create a new \code{Experiment}.}
\usage{
create_experiment(...)
}
\arguments{
\item{...}{Passed to \code{Experiment$new()}.}
}
\value{
A new instance of \code{Experiment}.
}
\description{
Create a new \code{Experiment}.
}
