% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-templates.R
\name{create_fun_str}
\alias{create_fun_str}
\title{Create a function call string and print to console.}
\usage{
create_fun_str(name, fun, args)
}
\description{
Create a function call string and print to console.
}
\keyword{internal}
