% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{create_method}
\alias{create_method}
\title{Create a new \code{Method}.}
\usage{
create_method(.method_fun, .name = NULL, ...)
}
\arguments{
\item{.method_fun}{The method function.}

\item{.name}{(Optional) The name of the \code{Method}.}

\item{...}{Arguments to pass into \code{.method_fun()}.}
}
\value{
A new instance of \code{Method}.
}
\description{
Create a new \code{Method}.
}
\examples{
# create an example Method function
lm_fun <- function(X, y, cols = c("X1", "X2")) {
  lm_fit <- lm(y ~ X)
  pvals <- summary(lm_fit)$coefficients[cols, "Pr(>|t|)"] \%>\%
    setNames(paste(names(.), "p-value"))
  return(pvals)
}

# create Method with default arguments
lm_method <- create_method(.method_fun = lm_fun, .name = "OLS")

# create Method with non-default arguments
lm_method_x1 <- create_method(.method_fun = lm_fun, .name = "OLS X1",
                              # additional named parameters to pass to lm_fun()
                              cols = "X1")
                  
}
