% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{create_plotter}
\alias{create_plotter}
\title{Create a new \code{Plotter}.}
\usage{
create_plotter(plot_fun, ...)
}
\arguments{
\item{plot_fun}{The plotting function.}

\item{...}{Arguments to pass into \code{plot_fun()}.}
}
\value{
A new instance of \code{Plotter}.
}
\description{
Create a new \code{Plotter}.
}
