% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-dir.R
\name{create_sim}
\alias{create_sim}
\title{Create a simulation project}
\usage{
create_sim(path, init_git = TRUE, init_renv = FALSE, tests = TRUE, hpc = FALSE)
}
\arguments{
\item{path}{A \code{character} specifying the path for your simulation
directory.}

\item{init_git}{A \code{logical} indicating whether to intialize your
simulation directory as a git repository.}

\item{init_renv}{A \code{logical} stating whether to initialize
\code{\link[renv::init]{renv}}. Defaults to \code{FALSE}.}

\item{tests}{A \code{logical} indicating whether to generate sub-directories
for organizing unit tests. Defaults to \code{TRUE}.}

\item{hpc}{A \code{logical} indicating whether to create sub-directories for
organizing files related to high-power computing environments. Defaults to
\code{FALSE}.}
}
\description{
\code{create_sim()} initializes a directory for your simulation
study. It wraps around \code{\link[usethis]{create_project}()} and
\code{\link[usethis]{use_git}()}.
}
