% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cross}
\alias{cross}
\title{Produce all combinations of list elements}
\usage{
cross(.l)
}
\arguments{
\item{.l}{A list of lists or atomic vectors. Alternatively, a data frame.}
}
\value{
List of all combinations of list elements
}
\description{
Copy of \code{purrr::cross()} which was deprecated in \verb{purrr 1.0.0}.
}
\keyword{internal}
