% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{do_call_handler}
\alias{do_call_handler}
\title{Call a function with given parameters and capture errors, warnings, or
messages that occur during evaluation.}
\usage{
do_call_handler(
  name,
  fun,
  params = list(),
  verbose = 1,
  call = rlang::caller_env()
)
}
\arguments{
\item{name}{a name to give some context for the call}

\item{fun}{a function to call}

\item{params}{arguments to pass to fun}

\item{verbose}{Verbosity level. If greater than 1, then handle warnings and
messages in addition to errors.}
}
\value{
the results of calling fun with params
}
\description{
Call a function with given parameters and capture errors, warnings, or
messages that occur during evaluation.
}
\keyword{internal}
