% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-lib-utils.R
\name{dots_to_fun_args}
\alias{dots_to_fun_args}
\title{Helper function to process ... args to pass to multiple functions.}
\usage{
dots_to_fun_args(fun_prefix = c("x", "y", "err", "betas"), ...)
}
\arguments{
\item{...}{Named arguments to process.}

\item{prefix}{Character vector of prefixes for the \code{*_fun} functions
used by the function that calls \code{dots_to_fun_args()}; e.g., if
\code{x_fun} is one of the functions, then this vector should include
\code{"x"}. Then, this function will look for an argument with name
prefixed by \code{".x_"} in \code{...}.}
}
\value{
A list with some or all of the following:
\describe{
\item{.x_args}{Args to pass to \code{x_fun}.}
\item{.y_args}{Args to pass to \code{y_fun}.}
\item{.err_args}{Args to pass to \code{err_fun}.}
\item{.betas_args}{Args to pass to \code{betas}.}
\item{.optional_args}{Args to pass optionally.}
}
}
\description{
Helper function to process ... args to pass to multiple functions.
}
\keyword{internal}
