% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-lib-prediction-metrics.R
\name{eval_auc_curve}
\alias{eval_auc_curve}
\title{Evaluate ROC or PR curves.}
\usage{
eval_auc_curve(y, yhat, metric = c("ROC", "PR"))
}
\arguments{
\item{y}{Vector of the true/observed response values.}

\item{yhat}{Vector of the predicted response values. Must be of the same
length as \code{y}.}

\item{metric}{A character string. Either "ROC" or "PR".}
}
\value{
A list of two:
\describe{
\item{AUC}{Area under the curve.}
\item{curve_df}{Data frame with the x and y values to plot curve. If
\code{metric = "ROC"}, the data frame columns give the true positive rate
(TPR) and the false positive rate (FPR). If \code{metric = "PR"}, the data
frame columns give the Precision and Recall.}
}
}
\description{
Compute the ROC or PR curves and its corresponding AUC value for
the given observed and predicted response values.
}
