% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-lib-utils.R
\name{eval_constructor}
\alias{eval_constructor}
\title{Developer function to construct basic Evaluator.}
\usage{
eval_constructor(
  fit_results,
  vary_params = NULL,
  fun,
  nested_cols = NULL,
  ...,
  group_cols = NULL,
  fun_options = NULL,
  na_rm = FALSE
)
}
\arguments{
\item{fit_results}{A tibble, as returned by the \code{fit} method.}

\item{vary_params}{A vector of parameter names that are varied across in the
\code{Experiment}.}

\item{fun}{Function to compute for each row in \code{fit_results}. This
function can take in the following optional arguments: (1) \code{data} =
\code{fit_results[i, ]}; (2) na_rm; (3) arguments specified via \code{...}
and \code{fun_options}.}

\item{nested_cols}{(Optional) A character string or vector specifying the
name of the column(s) in \code{fit_results} that need to be
unnested before evaluating results. Default is \code{NULL}, meaning no
columns in \code{fit_results} need to be unnested prior to computation.}

\item{...}{Named arguments, containing names of columns to pass to
\code{fun}.}

\item{group_cols}{(Optional) A character string or vector specifying the
column(s) to group rows by before evaluating metrics.
This is useful for assessing within-group metrics.}

\item{fun_options}{Named list of additional arguments to pass to \code{fun}.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA} values
should be stripped before the computation proceeds.}
}
\description{
Helper function for developing a new \code{Evaluator}
that evaluates some function (e.g., metric) for each row in
\code{fit_results}.
}
