% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-lib-feature-selection.R
\name{eval_feature_importances}
\alias{eval_feature_importances}
\title{Show feature importance scores.}
\usage{
eval_feature_importances(fit_results, vary_params = NULL, data_col)
}
\arguments{
\item{fit_results}{A tibble, as returned by the \code{fit} method.}

\item{vary_params}{A vector of parameter names that are varied across in the
\code{Experiment}.}

\item{data_col}{Character string or vector specifying the name of column(s)
in \code{fit_results} with the feature importance score information.}
}
\value{
A \code{tibble} with the columns \code{rep}, \code{dgp_name}, and
\code{method_name} in addition to the columns specified by
\code{vary_params} and \code{data_col}, where the \code{data_col} column(s)
has been unnested.
}
\description{
Show feature importance scores across all repetitions, methods,
and DGPs in an \code{Experiment}.
}
\seealso{
Other feature_selection_funs: 
\code{\link{eval_feature_recovery_curve}()},
\code{\link{eval_feature_recovery}()},
\code{\link{plot_feature_imp}()},
\code{\link{plot_feature_recovery_curve}()},
\code{\link{plot_feature_recovery}()},
\code{\link{summarize_feature_importances}()},
\code{\link{summarize_feature_recovery_curve}()},
\code{\link{summarize_feature_recovery}()}
}
\concept{feature_selection_funs}
