% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-lib-feature-selection.R
\name{eval_feature_recovery}
\alias{eval_feature_recovery}
\title{Evaluate feature support recovery.}
\usage{
eval_feature_recovery(
  s,
  shat,
  metrics = c("TP", "FP", "TPR", "FPR", "FDR", "P", "N", "AUROC", "AUPRC"),
  custom_metrics = NULL
)
}
\arguments{
\item{s}{Named vector of the true feature support set, where the names are
the feature names and the vector values are non-zero for the true
support set and zero for the true non-support.}

\item{shat}{Named vector of the estimated feature support set, where the
names are the feature names and the vector values are non-zero for the
estimated support with larger values indicating greater feature importance
and zero for the estimated non-support.}

\item{metrics}{Character vector of the support recovery metrics to compute.
Elements of the vector must be one of "TP", "FP", "TPR", "FPR", "FDR",
"P", "N", "AUROC", "AUPRC"}

\item{custom_metrics}{Named list of custom metric functions to compute. Names
in the list should correspond to the name of the metric. Values in the
list should be a function that takes in the arguments \code{s} and
\code{shat} only and returns the evaluated metric value.}
}
\value{
A (long-shaped) data frame with the following columns:
\describe{
\item{metric}{Name of feature support recovery metric.}
\item{value}{Feature support recovery value for the given metric.}
}
}
\description{
Evaluate various support recovery metrics, given the true
feature support set \code{s} and the estimated feature support set
\code{shat}.
}
\details{
TODO: explain different metrics
}
