% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-lib-feature-selection.R
\name{eval_feature_recovery_curve}
\alias{eval_feature_recovery_curve}
\title{Evaluate ROC or PR curves for feature support recovery.}
\usage{
eval_feature_recovery_curve(
  data,
  truth,
  estimate,
  metric = c("ROC", "PR"),
  options = list(),
  na_rm = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame} containing the \code{truth} and
\code{estimate} columns. Each row in this \code{data.frame} typically
corresponds to a feature or variable.}

\item{truth}{A character string identifying the column with the true
support. The column should be binary, where \code{TRUE} or \code{1} means
the feature (corresponding to that row) is in the support and \code{FALSE}
or \code{0} means the feature is not in the support.}

\item{estimate}{A character string identifying the column with the estimated
support, where a non-zero entry means that the feature (corresponding to
that row) is part of the estimated support and a zero entry means that
feature is not part of the estimated support. Moreover, a higher magnitude
indicates a more important feature.}

\item{metric}{Either "ROC" or "PR" indicating whether to evaluate the ROC or
Precision-Recall curve.}

\item{options}{A \code{list} of named options to pass to \code{\link[pROC:roc]{pROC::roc()}}
such as \code{smooth}. These options should not include \code{response},
\code{predictor}, \code{levels}, \code{quiet}, or \code{direction}.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA} values
should be stripped before the computation proceeds.}
}
\value{
If \code{metric = "ROC"}, returns a \code{tibble} with the columns
\code{.threshold}, \code{FPR}, and \code{TPR} for the threshold, false
positive rate, and true positive rate, respectively. If
\code{metric = "PR"}, returns a \code{tibble} with columns
\code{.threshold}, \code{recall}, and \code{precision}.
}
\description{
Evaluate the ROC or PR curves for feature support recovery and
return a tibble with the results.
}
\seealso{
Other feature_selection_funs: 
\code{\link{eval_feature_importances}()},
\code{\link{eval_feature_recovery}()},
\code{\link{plot_feature_imp}()},
\code{\link{plot_feature_recovery_curve}()},
\code{\link{plot_feature_recovery}()},
\code{\link{summarize_feature_importances}()},
\code{\link{summarize_feature_recovery_curve}()},
\code{\link{summarize_feature_recovery}()}
}
\concept{feature_selection_funs}
