% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-lib-inference.R
\name{eval_pval_ranking_curve}
\alias{eval_pval_ranking_curve}
\title{Evaluate ROC or PR curves for feature rankings from p-values.}
\usage{
eval_pval_ranking_curve(
  data,
  truth,
  p_value,
  metric = c("ROC", "PR"),
  options = list(),
  na_rm = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame} containing the \code{truth} and
\code{p_value} columns. Each row in this \code{data.frame} typically
corresponds to a feature or variable.}

\item{truth}{A character string identifying the column with the true
support. The column should be binary, where \code{TRUE} or \code{1} means
the feature (corresponding to that row) is in the support and \code{FALSE}
or \code{0} means the feature is not in the support.}

\item{p_value}{A character string identifying the column with the estimated
p-values.}

\item{metric}{Either "ROC" or "PR" indicating whether to evaluate the ROC or
Precision-Recall curve.}

\item{options}{A \code{list} of named options to pass to \code{\link[pROC:roc]{pROC::roc()}}
such as \code{smooth}. These options should not include \code{response},
\code{predictor}, \code{levels}, \code{quiet}, or \code{direction}.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA} values
should be stripped before the computation proceeds.}
}
\value{
If \code{metric = "ROC"}, returns a \code{tibble} with the columns
\code{.threshold}, \code{FPR}, and \code{TPR} for the threshold, false
positive rate, and true positive rate, respectively. If
\code{metric = "PR"}, returns a \code{tibble} with columns
\code{.threshold}, \code{recall}, and \code{precision}.
}
\description{
Evaluate the ROC or PR curves for the feature rankings
determined by the p-values and return a tibble with the results.
}
\seealso{
Other inference_funs: 
\code{\link{eval_reject_prob}()},
\code{\link{eval_testing_err}()},
\code{\link{plot_pval_ranking_curve}()},
\code{\link{plot_reject_prob}()},
\code{\link{plot_testing_err}()},
\code{\link{summarize_pval_ranking_curve}()},
\code{\link{summarize_testing_err}()}
}
\concept{inference_funs}
