% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-lib-utils.R
\name{eval_summary_constructor}
\alias{eval_summary_constructor}
\title{Developer function for summarizing evaluation results.}
\usage{
eval_summary_constructor(
  eval_results,
  id,
  value = "value",
  summary_funs = c("mean", "median", "min", "max", "sd", "raw"),
  custom_summary_funs = NULL,
  na_rm = FALSE
)
}
\arguments{
\item{eval_results}{A list of result tibbles, as returned by the
\code{evaluate} method.}

\item{id}{Character string. ID to be used as a suffix when naming result
columns.}

\item{value}{Character string. Name of column in \code{eval_results} to
summarize.}

\item{summary_funs}{Character vector specifying how to summarize
evaluation metrics. Must choose from a built-in library of summary
functions - elements of the vector must be one of "mean", "median",
"min", "max", "sd", "raw".}

\item{custom_summary_funs}{Named list of custom functions to summarize
results. Names in the list should correspond to the name of the summary
function. Values in the list should be a function that takes in one
argument, that being the values of the evaluated metrics.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA} values
should be stripped before the computation proceeds.}
}
\value{
A \code{tibble} containing the summarized results aggregated
over experimental replicates. These columns correspond to the requested
statistics in \code{summary_funs} and \code{custom_summary_funs} and end
with the suffix specified by \code{id}. Note that if \code{eval_results}
is a grouped \code{data.frame}, then the group IDs are also retained in the
returned \code{tibble}.
}
\description{
A helper function for developing new \code{Evaluator} functions
that summarize results over multiple experimental replicates.
}
