% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-lib-inference.R
\name{eval_testing_err}
\alias{eval_testing_err}
\title{Evaluate error metrics when conducting multiple tests.}
\usage{
eval_testing_err(
  data,
  truth,
  p_value,
  metrics = NULL,
  alphas = 0.05,
  na_rm = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame} containing the \code{truth} and
\code{p_value} columns. Each row in this \code{data.frame} typically
corresponds to a feature or variable.}

\item{truth}{A character string identifying the column with the true
support. The column should be binary, where \code{TRUE} or \code{1} means
the feature (corresponding to that row) is in the support and \code{FALSE}
or \code{0} means the feature is not in the support.}

\item{p_value}{A character string identifying the column with the estimated
p-values.}

\item{metrics}{A \code{metric_set} object indicating the metrics to evaluate.
See \link[yardstick:metric_set]{yardstick::metric_set} for more details. Default \code{NULL} will
evaluate the following: number of true positives (\code{tp}), number of
false positives (\code{fp}), sensitivity (\code{sens}), specificity
(\code{spec}), positive predictive value (\code{ppv}), number of features
in the estimated support (\code{pos}), number of features not in the
estimated support (\code{neg}), AUROC (\code{roc_auc}), and AUPRC
(\code{pr_auc}).}

\item{alphas}{(Optional) Vector of significance levels at which to evaluate
the various metrics. By default, \code{alphas = 0.05}.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA} values
should be stripped before the computation proceeds.}
}
\value{
A two column \code{tibble} with the following columns:
\describe{
\item{.alpha}{Level of significance}
\item{.metric}{Name of error metric.}
\item{.estimate}{Error value for the given metric.}
}
}
\description{
Evaluate various error metrics, given the true
feature support and the estimated p-values from multiple tests at
pre-specified significance level thresholds.
}
\seealso{
Other inference_funs: 
\code{\link{eval_pval_ranking_curve}()},
\code{\link{eval_reject_prob}()},
\code{\link{plot_pval_ranking_curve}()},
\code{\link{plot_reject_prob}()},
\code{\link{plot_testing_err}()},
\code{\link{summarize_pval_ranking_curve}()},
\code{\link{summarize_testing_err}()}
}
\concept{inference_funs}
