% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{evaluate_experiment}
\alias{evaluate_experiment}
\alias{generate_data}
\title{Evaluate an \code{Experiment}}
\usage{
generate_data(experiment, ...)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{...}{Passed to \code{experiment$evaluate()}.}
}
\value{
A list of the simulation experiment's evaluation results.
}
\description{
Evaluate an \code{Experiment}
}
