% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator.R
\docType{class}
\name{Evaluator}
\alias{Evaluator}
\title{\code{R6} class representing an evaluator.}
\description{
An evaluator (or evaluation function) to \strong{evaluate} the
performance of methods in the \code{Experiment}.
}
\section{Public fields}{


\describe{
\item{\code{name}}{The name of the \code{Evaluator}.}
\item{\code{eval_fun}}{The evaluation function.}
\item{\code{eval_params}}{(Named) list of parameters to input into the evaluation function.}
\item{\code{doc_options}}{List of options to control the aesthetics of the
displayed \code{Evaluator}'s results table in the knitted R Markdown
report. See \code{\link[vthemes:pretty_DT]{vthemes::pretty_DT()}} for possible options.}
\item{\code{doc_show}}{If \code{TRUE} (default), show \code{Evaluator}'s results as
a table in the R Markdown report; if \code{FALSE}, hide output in the
R Markdown report.}
}
}

\section{Public methods}{


\itemize{
\item{\code{Evaluator$new()}} See \link{create_evaluator}.
\item{\code{Evaluator$evaluate()}} See \link{evaluate_evaluator}.
}
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Evaluator-new}{\code{Evaluator$new()}}
\item \href{#method-Evaluator-evaluate}{\code{Evaluator$evaluate()}}
\item \href{#method-Evaluator-print}{\code{Evaluator$print()}}
\item \href{#method-Evaluator-clone}{\code{Evaluator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Evaluator-new"></a>}}
\if{latex}{\out{\hypertarget{method-Evaluator-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evaluator$new(
  .eval_fun,
  .name = NULL,
  .doc_options = list(),
  .doc_show = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Evaluator-evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-Evaluator-evaluate}{}}}
\subsection{Method \code{evaluate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evaluator$evaluate(fit_results, vary_params = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Evaluator-print"></a>}}
\if{latex}{\out{\hypertarget{method-Evaluator-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evaluator$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Evaluator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Evaluator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evaluator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
