% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-lib-utils.R
\name{event_col}
\alias{event_col}
\title{Logic for \code{event_level} in custom \code{yardstick} metrics.}
\usage{
event_col(xtab, event_level)
}
\arguments{
\item{xtab}{Frequency table from \code{table()}}

\item{event_level}{A single string. Either \code{"first"} or \code{"second"} to specify
which level of \code{truth} to consider as the "event". This argument is only
applicable when \code{estimator = "binary"}. The default uses an
internal helper that generally defaults to \code{"first"}, however, if the
deprecated global option \code{yardstick.event_first} is set, that will be
used instead with a warning.}
}
\value{
Name of factor level to use as the "event" when computing evaluation
metrics.
}
\description{
Logic for \code{event_level} in custom \code{yardstick} metrics.
}
\keyword{internal}
