% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{Experiment}
\alias{Experiment}
\title{A computational experiment.}
\description{
A computational experiment.

A computational experiment.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Experiment$new()}}
\item \href{#method-run}{\code{Experiment$run()}}
\item \href{#method-evaluate}{\code{Experiment$evaluate()}}
\item \href{#method-add_dgp}{\code{Experiment$add_dgp()}}
\item \href{#method-update_dgp}{\code{Experiment$update_dgp()}}
\item \href{#method-add_method}{\code{Experiment$add_method()}}
\item \href{#method-update_method}{\code{Experiment$update_method()}}
\item \href{#method-add_evaluator}{\code{Experiment$add_evaluator()}}
\item \href{#method-update_evaluator}{\code{Experiment$update_evaluator()}}
\item \href{#method-clone}{\code{Experiment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$new(
  n_obs,
  n_reps,
  dgp_list = list(),
  method_list = list(),
  evaluator_list = list(),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$run(trial_run = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-evaluate}{}}}
\subsection{Method \code{evaluate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$evaluate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_dgp"></a>}}
\if{latex}{\out{\hypertarget{method-add_dgp}{}}}
\subsection{Method \code{add_dgp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$add_dgp(dgp, name = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_dgp"></a>}}
\if{latex}{\out{\hypertarget{method-update_dgp}{}}}
\subsection{Method \code{update_dgp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$update_dgp(dgp, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_method"></a>}}
\if{latex}{\out{\hypertarget{method-add_method}{}}}
\subsection{Method \code{add_method()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$add_method(method, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_method"></a>}}
\if{latex}{\out{\hypertarget{method-update_method}{}}}
\subsection{Method \code{update_method()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$update_method(method, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_evaluator"></a>}}
\if{latex}{\out{\hypertarget{method-add_evaluator}{}}}
\subsection{Method \code{add_evaluator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$add_evaluator(evaluator, dgp_name = NULL, method_name = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_evaluator"></a>}}
\if{latex}{\out{\hypertarget{method-update_evaluator}{}}}
\subsection{Method \code{update_evaluator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$update_evaluator(
  evaluator,
  name = NULL,
  dgp_name = NULL,
  method_name = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
