% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-helpers.R
\name{export_visualizers}
\alias{export_visualizers}
\title{Export cached \code{Visualizer} results to image.}
\usage{
export_visualizers(experiment, device = "png", ...)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{device}{See \code{device} argument of \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.}

\item{...}{Additional arguments to pass to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} to be used
for all visualizations. If not provided, the \code{export_options} from
each \code{Visualizer} will be used.}
}
\value{
The original \code{Experiment} object passed to
\code{export_visualizers}.
}
\description{
Export all cached \code{Visualizer} results from an \code{Experiment} to
images in the \verb{viz_results/} directory under the \code{Experiment}'s results
directory (see \code{\link[=get_save_dir]{get_save_dir()}}).
}
\examples{
\dontrun{
export_visualizers(experiment)}

}
