% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{fit_experiment}
\alias{fit_experiment}
\title{Fit an \code{Experiment}}
\usage{
fit_experiment(experiment, n_reps = 1, ...)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{n_reps}{The number of replicates of the \code{Experiment} for this run.}

\item{...}{Passed to \code{experiment$fit()}.}
}
\value{
A list of the simulation experiment's evaluation results.
}
\description{
Fit an \code{Experiment}
}
