% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_duplicate_param_names}
\alias{fix_duplicate_param_names}
\title{Fix duplicate \code{vary_across} parameter names.}
\usage{
fix_duplicate_param_names(
  dgp_params,
  method_params,
  duplicate_param_names = NULL
)
}
\arguments{
\item{dgp_params}{A named list of the DGP parameters.}

\item{method_params}{A named list of the Method parameters.}

\item{duplicate_param_names}{A vector of parameter names that are varied
across in both a DGP and a Method}
}
\value{
A named list of the DGP and Method parameters with no duplicate
names.
}
\description{
Add "_dgp" or "_method" suffixes to parameter names that are
found in both the DGP and Method \code{vary_across} components. This is to
avoid errors that occur from duplicate column names when trying to create a
tibble.
}
\keyword{internal}
