% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-lib-utils.R
\name{generate_coef}
\alias{generate_coef}
\title{Helper function to generate a coefficient vector.}
\usage{
generate_coef(betas = NULL, p, s = p, sd = 1, betas_name = "betas")
}
\arguments{
\item{betas}{Coefficient vector. If a scalar is provided, the coefficient
vector is a constant vector. If \code{NULL} (default), entries in the
coefficient vector are drawn iid from N(0, \code{sd}^2).}

\item{p}{Number of features.}

\item{s}{Sparsity level. Coefficients corresponding to features after the
\code{s}th position (i.e., positions i = \code{s} + 1, ..., \code{p}) are
set to 0.}

\item{sd}{(Optional) SD of normal distribution from which to draw
\code{betas}. Only used if \code{betas} argument is \code{NULL}.}

\item{betas_name}{Name of coefficient variable to use in error message.}
}
\value{
A vector of length \code{p}.
}
\description{
Helper function to generate a coefficient vector.
}
\keyword{internal}
