% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-helpers.R
\name{generate_data}
\alias{generate_data}
\title{Generate data from each \code{DGP} in the \code{Experiment}.}
\usage{
generate_data(experiment, n_reps = 1, ...)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{n_reps}{The number of datasets to generate per \code{DGP}.}

\item{...}{Not used.}
}
\value{
A tibble with one row for each of the requested \code{n_reps}
datasets, generated from each \code{DGP} in the \code{Experiment} and
varying across the \code{vary_across} parameters(s) if necessary. Columns
in this tibble include \code{.rep} (experimental replicate ID),
\code{.dgp_name} (name of the DGP), additional columns for the
\code{vary_across} parameters, and \code{.data_output} (data from the
specified \code{DGP} configuration).
}
\description{
Generate data from each \code{DGP} in the \code{Experiment}.
}
