% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-lib-errors.R
\name{generate_errors}
\alias{generate_errors}
\title{Helper function to generate simulated errors.}
\usage{
generate_errors(err = NULL, n, X, ...)
}
\arguments{
\item{err}{Function from which to generate simulate error vector. Default
is \code{NULL} which returns an error vector of all zeros.}

\item{n}{Number of samples.}

\item{X}{Data matrix or data frame. Used to determine n if n is missing.}

\item{...}{Other arguments to pass to err() to generate the error vector.}
}
\value{
A vector of simulated errors with length \code{n}.
}
\description{
Generate simulated errors from the specified error function and
passed arguments.
}
\details{
The function \code{err} must accept an argument named \code{n}
(i.e., the sample size).
}
