% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-lib-X.R
\name{generate_X_gaussian}
\alias{generate_X_gaussian}
\title{Generate a normal random matrix of covariates/features.}
\usage{
generate_X_gaussian(n, p, mean = 0, sd = 1, corr = 0, Sigma = NULL)
}
\arguments{
\item{n}{Number of samples.}

\item{p}{Number of features.}

\item{mean}{Mean of normal distribution from which to generate data. Can be
either a scalar value or vector of length p. Default is 0.}

\item{sd}{Standard deviation of normal distribution from which to generate
data. Default is 1.}

\item{corr}{Correlation between all pairs of features. Default is 0 for no
correlation.}

\item{Sigma}{(Optional) p x p covariance matrix of the normal distribution
from which to generate data. Default is NULL (not used). If provided, the
arguments \code{corr} and \code{sd} are ignored.}
}
\value{
A normal random matrix of size n x p.
}
\description{
Generate an n x p normal random matrix with the specified mean
and covariance structure.
}
