% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-lib-X.R
\name{generate_X_rwd}
\alias{generate_X_rwd}
\title{Generate a design matrix X by sampling from a real-world data matrix.}
\usage{
generate_X_rwd(X, n = nrow(X), p = ncol(X), clusters = NULL, replace = FALSE)
}
\arguments{
\item{X}{A data frame.}

\item{n}{Number of samples if \code{clusters} is not \code{NULL}. If
\code{clusters = NULL}, this is the number of clusters.}

\item{p}{Number of features. If \code{p < ncol(X)}, the \code{p} features
are sampled uniformly at random from the full feature set.}

\item{clusters}{(Optional) Vector of cluster IDs. If provided, block or
clustered sampling will be performed according to these clusters so that
each cluster will be entirely in or entirely out of the retrieved sample.}

\item{replace}{Logical. If TRUE, sample observations with replacement; if
FALSE, sample observations without replacement}
}
\value{
A matrix of size n x p.
}
\description{
Generate a design matrix X by sampling from a real-world data
matrix under the specified sampling scheme.
}
\details{
Only one of the \code{strata} and \code{clusters} arguments should
be provided.
}
