% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer-lib-utils.R
\name{get_aesthetics}
\alias{get_aesthetics}
\title{Get list of aesthetics to add to ggplot}
\usage{
get_aesthetics(
  x_str = NULL,
  y_str = NULL,
  color_str = NULL,
  fill_str = NULL,
  group_str = NULL
)
}
\value{
Named list of non-NULL inputs.
}
\description{
Helper funciton to ignore NULL inputs when adding aesthetics to
a ggplot
}
