% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_helpers.R
\name{get_cache}
\alias{get_cache}
\title{Retrieve cached results from previously saved \code{Experiment}.}
\usage{
get_cache(experiment)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}
}
\value{
The DGP, Method, Evaluator, and Visualizer IDs in this
\code{Experiment} that have been cached (i.e., saved to disk).
}
\description{
Retrieve cached results from previously saved \code{Experiment}.
}
