% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-helpers.R
\name{get_cached_results}
\alias{get_cached_results}
\title{Retrieve cached results from previously saved \code{Experiment}.}
\usage{
get_cached_results(experiment, results_type, verbose = 0)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{results_type}{Character string indicating the type of results to read
in. Must be one of "experiment", "experiment_cached_params", "fit", "eval",
or "viz".}

\item{verbose}{Level of verbosity. Default is 1, which prints out messages
after major checkpoints in the experiment. If 2, prints additional
debugging information for warnings and messages from user-defined functions
(in addition to error debugging information). If 0, no messages are printed
other than user-defined function error debugging information.}
}
\value{
The cached results, specifically the cached \code{Experiment} object
if \code{results_type = "experiment"}, the cached fit results if
\code{results_type = "fit"}, the cached evaluation results if
\code{results_type = "eval"}, the cached visualization results if
\code{results_type = "viz"}, and the experiment parameters used in
the cache if \code{results_type = "experiment_cached_params"}.
}
\description{
Read in cached results from disk from a previously saved
\code{Experiment}.
}
