% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer-lib-utils.R
\name{get_dot_args}
\alias{get_dot_args}
\title{Get dot (...) arguments.}
\usage{
get_dot_args(user_args, default_args)
}
\arguments{
\item{user_args}{List of user-specified dot (...) arguments.}

\item{default_args}{List of default dot (...) arguments.}
}
\value{
A named list of arguments that includes all arguments from the user
and the defaults, but with the user-specified arguments overwriting the
defaults.
}
\description{
Helper function to merge default and user-specified dot (...)
arguments such that default arguments are overwritten by the user-specified
arguments.
}
\examples{
arg_list <- get_dot_args(user_args = list(a = 1, b = 2, c = 3), 
                         default_args = list(a = "a", d = "d"))

}
