% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer-lib-utils.R
\name{get_eval_tibble}
\alias{get_eval_tibble}
\title{Get the summarized evaluation results tibble for plotting.}
\usage{
get_eval_tibble(
  fit_results,
  eval_results,
  evaluator_name = NULL,
  eval_id,
  eval_fun,
  vary_params = NULL,
  show,
  y_str = NULL,
  ...
)
}
\arguments{
\item{fit_results}{A tibble, as returned by the \code{fit} method.}

\item{eval_results}{A list of result tibbles, as returned by the
\code{evaluate} method.}

\item{evaluator_name}{Name of \code{Evaluator} containing results to plot.
To compute the evaluation summary results from scratch or if the evaluation
summary results have not yet been evaluated, set to \code{NULL}.}

\item{eval_id}{Character string. ID used as the suffix for naming columns in
\code{eval_summary_constructor()}. Should be the same as the \code{id}
argument in \code{eval_summary_constructor()}.}

\item{eval_fun}{Function used to compute evaluation results summary. This
function is only used (and required) if necessary results have not already
been computed in the evaluation summary results.}

\item{vary_params}{A vector of parameter names that are varied across in the
\code{Experiment}.}

\item{show}{Character vector with elements being one of "boxplot", "point",
"line", "bar", "errorbar", "ribbon" indicating what plot layer(s) to
construct.}

\item{y_str}{(Optional) Name of column in evaluation summary results
\code{tibble} to plot on the y-axis if \code{show} is anything but
"boxplot".}

\item{...}{Additional arguments to pass to \code{eval_fun()}. This is only
used if necessary results have not already been computed in
\code{eval_results[[evaluator_name]]}.}
}
\description{
Helper function to get the summarized evaluation results
\code{tibble} for plotting. This function will compute the summarized
evaluation results if they have not been computed previously. Otherwise,
it will read in the previously computed results and compute and append any
new results necessary to construct the specified plot.
}
