% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{get_funs}
\alias{get_funs}
\alias{get_dgps}
\alias{get_methods}
\alias{get_evaluators}
\alias{get_visualizers}
\title{Helper functions for getting components in an \code{Experiment}.}
\usage{
get_dgps(experiment, ...)

get_methods(experiment, ...)

get_evaluators(experiment, ...)

get_visualizers(experiment, ...)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{...}{Not used.}
}
\value{
The original \code{Experiment} object passed to \code{get_*}.
}
\description{
Helper functions for getting or retrieving \code{DGPs},
\code{Methods}, \code{Evaluators}, and \code{Visualizers} from an
\code{Experiment}.
}
