% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_matching_rows}
\alias{get_matching_rows}
\title{Get matching rows from x based on id}
\usage{
get_matching_rows(id, x)
}
\arguments{
\item{id}{A tibble with distinct rows.}

\item{x}{A tibble.}
}
\value{
A tibble, containing the subset of rows from \code{x} that match
id rows from \code{id}.
}
\description{
Get rows in \code{x} tibble that match the id rows specified
by the \code{id} tibble. This function is an alternative to the usual
\code{dplyr::inner_join} function. The difference is that this function
ignores the source bytecode of functions when looking for matching rows,
while \code{dplyr::inner_join} treats functions with different sources
as different. This function also requires that the \code{id} tibble have
distinct rows while \code{dplyr::inner_join} does not. This function
enables caching when functions are used as parameters in DGPs and Methods.
}
\keyword{internal}
