% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer-lib-utils.R
\name{get_plot_data}
\alias{get_plot_data}
\title{Developer function to get data for plotting in Visualizers.}
\usage{
get_plot_data(
  fit_results = NULL,
  eval_results = NULL,
  eval_name = NULL,
  eval_fun = NULL,
  eval_fun_options = NULL
)
}
\arguments{
\item{fit_results}{A tibble, as returned by the \code{fit} method.}

\item{eval_results}{A list of result tibbles, as returned by the
\code{evaluate} method.}

\item{eval_name}{Name of \code{Evaluator} containing results to plot. If
\code{NULL}, the data used for plotting is computed from scratch via
\code{eval_fun}.}

\item{eval_fun}{Character string, specifying the function used to compute
the data used for plotting if \code{eval_name = NULL}. If \code{eval_name}
is not \code{NULL}, this argument is ignored.}

\item{eval_fun_options}{List of named arguments to pass to \code{eval_fun}.}
}
\value{
If \code{eval_name} is not \code{NULL}, returns the list
component in \code{eval_results} named \code{eval_name}. Otherwise,
returns the result of \code{eval_fun()}.
}
\description{
A helper function to retrieve data for plotting in
\code{Visualizers}. It first looks for the data in \code{eval_results} by
name (\code{eval_name}). If this is not provided, then it uses the
function + arguments specified by \code{eval_fun} and
\code{eval_fun_options} to compute the plotting data from
\code{fit_results}.
}
