% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_helpers.R
\name{get_save_dir}
\alias{get_save_dir}
\title{Get results directory for an \code{Experiment}.}
\usage{
get_save_dir(experiment)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}
}
\value{
The relative path to where the \code{Experiment}'s results and
visualizations are saved.
}
\description{
Get the directory in which the \code{Experiment}'s results and
visualizations are saved.
}
