% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-lib-utils.R
\name{indicator}
\alias{indicator}
\title{Helper function to compute indicator function in LSS model.}
\usage{
indicator(X, thresh, sgn)
}
\arguments{
\item{X}{Data matrix or data frame.}

\item{thresh}{thresholds}

\item{sgn}{sign of the interaction}
}
\description{
Helper function to compute indicator function in LSS model.
}
\keyword{internal}
