% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{init_docs}
\alias{init_docs}
\title{Initialize documentation template for the R Markdown results report.}
\usage{
init_docs(experiment, save_dir)
}
\arguments{
\item{experiment}{An \code{Experiment} object. If provided, documentation is
created for all previously saved \code{Experiments} that are found in the
directory given by \code{experiment$get_save_dir()}. If no
\code{Experiments} have been previously saved under this directory, then
the current \code{experiment} is saved to disk and used to create the
documentation template.}

\item{save_dir}{An optional directory in which to find previously saved
\code{Experiment} objects. Documentation is created for these found
\code{Experiments}. Not used if \code{experiment} is provided.}
}
\value{
The original \code{Experiment} object if provided. Otherwise,
returns \code{NULL}.
}
\description{
Create documentation template (a series of .md files) to
fill out for the R Markdown results report. If the \code{experiment} is
provided, the documentation files can be found in the \code{Experiment}'s
results directory (see \code{Experiment$get_save_dir()}) under docs/.
Otherwise, the documentation files can be found in the specified
\code{save_dir} directory under docs/. The documentation files generated
include objectives.md and .md files corresponding to \code{DGPs},
\code{Methods}, \code{Evaluators}, and \code{Visualizers} in the
\code{Experiment}.
}
\examples{
\dontrun{
# create documentation template from an experiment (of class `Experiment`)
init_docs(experiment)

# or alternatively, create documentation template from a specific directory
init_docs(save_dir = experiment$get_save_dir())}

}
