% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-dir.R
\name{init_sim_project}
\alias{init_sim_project}
\title{Create a simulation project}
\usage{
init_sim_project(
  path,
  init_git = TRUE,
  init_renv = FALSE,
  tests = TRUE,
  hpc = FALSE
)
}
\arguments{
\item{path}{Character string, specifying the path for your simulation
directory.}

\item{init_git}{Logical, indicating whether to intialize your
simulation directory as a git repository.}

\item{init_renv}{Logical, stating whether to initialize \code{renv} with
\code{init()} from the \code{renv} package. Defaults to \code{FALSE}.}

\item{tests}{Logical, indicating whether to generate sub-directories
for organizing unit tests. Defaults to \code{TRUE}.}

\item{hpc}{Logical, indicating whether to create sub-directories for
organizing files related to high-performance computing environments. Defaults to
\code{FALSE}.}
}
\description{
\code{init_sim_project()} initializes a directory for your simulation
study. It wraps around \code{usethis::create_project()}, as well as
\code{usethis::use_git()} and \code{renv::init()}.
}
\examples{
\dontrun{
# create template directory for simulation project
init_sim_project("path/to/sim")}

}
