% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_templates.R
\name{library_templates}
\alias{library_templates}
\alias{use_prediction_template}
\alias{use_feature_selection_template}
\alias{use_inference_template}
\title{Functions to create boilerplate code for specific types of experiments.}
\usage{
use_prediction_template(
  experiment_name = "Prediction Experiment",
  type = c("regression", "classification"),
  support = FALSE,
  include_dgp_example = FALSE,
  include_method_example = FALSE
)

use_feature_selection_template(
  experiment_name = "Feature Selection Experiment",
  include_dgp_example = FALSE,
  include_method_example = FALSE
)

use_inference_template(
  experiment_name = "Inference Experiment",
  include_dgp_example = FALSE,
  include_method_example = FALSE
)
}
\arguments{
\item{experiment_name}{Name of the experiment.}

\item{type}{Either "regression" or "classification" specifying the type of
prediction problem.}

\item{support}{Logical. If \code{TRUE}, include code to evaluate the
estimated feature support.}

\item{include_dgp_example}{Logical. If \code{TRUE}, include a completed
DGP example, rather than a fill-in-the-blank template.}

\item{include_method_example}{Logical. If \code{TRUE}, include a completed
Method example, rather than a fill-in-the-blank template.}
}
\value{
Invisible \code{NULL} but code is printed to the console.
}
\description{
These functions make suggestions for code when performing a few
common types of experiments (i.e., for prediction, feature selection, and
inference). They print out code to the console that could be considered
minimal syntax.
}
