% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-lib.R
\name{linear_gaussian_dgp}
\alias{linear_gaussian_dgp}
\title{Generate independent Gaussian covariates and linear response data.}
\usage{
linear_gaussian_dgp(
  n,
  p_obs = 0,
  p_unobs = 0,
  s_obs = p_obs,
  s_unobs = p_unobs,
  betas = NULL,
  betas_unobs = NULL,
  betas_sd = 1,
  betas_unobs_sd = 1,
  err = NULL,
  data_split = FALSE,
  train_prop = 0.5,
  return_values = c("X", "y", "support"),
  ...
)
}
\arguments{
\item{n}{Number of samples.}

\item{p_obs}{Number of observed features.}

\item{p_unobs}{Number of unobserved (omitted) features.}

\item{s_obs}{Number of observed features with non-zero coefficients.}

\item{s_unobs}{Number of unobserved (omitted) features with non-zero
coefficients.}

\item{betas}{Coefficient vector for observed design matrix.}

\item{betas_unobs}{Coefficient vector for unobserved design matrix.}

\item{betas_sd}{(Optional) SD of normal distribution from which to draw
\code{betas}. Only used if \code{betas} argument is \code{NULL}.}

\item{betas_unobs_sd}{(Optional) SD of normal distribution from which to draw
\code{betas_unobs}. Only used if \code{betas_unobs} argument is
\code{NULL}.}

\item{err}{Function from which to generate simulate error vector. Default
is \code{NULL} which adds no error to the DGP.}

\item{data_split}{Logical; if \code{TRUE}, splits data into training and test
sets according to \code{train_prop}.}

\item{train_prop}{Proportion of data in training set if
\code{data_split = TRUE}.}

\item{return_values}{Character vector indicating what objects to return in
list. Elements in vector must be one of "X", "y", "support".}

\item{...}{Other arguments to pass to err() to generate the error vector.}
}
\value{
A list of the named objects that were requested in
\code{return_values}. Note that if \code{data_split = TRUE} and "X", "y"
are in \code{return_values}, then the returned list also contains slots for
"Xtest" and "ytest".
}
\description{
Generate independent normally-distributed covariates (including
potentially omitted variables) and linear response data with a specified
error distribution.
}
\details{
TODO: Support is the first s_obs and s_unobs features in the
generated X and U data.
}
