% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_to_tibble}
\alias{list_to_tibble}
\title{Coerce list into a tibble.}
\usage{
list_to_tibble(ls)
}
\arguments{
\item{ls}{List to convert into a tibble.}
}
\value{
A tibble.
}
\description{
Coerce a list into a tibble. Default is to coerce using
\code{\link[tibble:as_tibble]{tibble::as_tibble()}}, but if this fails, then coerce list into a tibble,
where each non-scalar column in the tibble is of type list.
}
\keyword{internal}
