% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_to_tibble_row}
\alias{list_to_tibble_row}
\title{Coerce list into a tibble row.}
\usage{
list_to_tibble_row(lst, simplify = FALSE)
}
\arguments{
\item{lst}{List to convert into a tibble row.}
}
\value{
A tibble with one row.
}
\description{
Coerce a list into a single row in a tibble. Default is to
coerce using \code{\link[tibble:as_tibble]{tibble::as_tibble_row()}}, but if this fails, then coerce list
into a tibble row, where each column in the tibble is of type list.
}
\keyword{internal}
