% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-tests.R
\name{load_all}
\alias{load_all}
\title{Load All Simulation Functions in R/}
\usage{
load_all()
}
\description{
\code{load_all()} is simulation study counterpart to
\code{\link[devtools:load_all]{devtools::load_all()}} it loads all of
the functions in the \code{R/dgp}, \code{R/method}, \code{R/eval} and
\code{R/viz} directories of the current simulation study.
}
\examples{
\dontrun{
# load all functions in R/dgp, R/method, R/eval, and R/viz directories
load_all()}

}
