% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_initialize_arg_list}
\alias{make_initialize_arg_list}
\title{The point of this method is to undo partial argument matching on the formals
of the method that calls make_initialize_arg_list().}
\usage{
make_initialize_arg_list(obj_fun, ..., which = -1)
}
\arguments{
\item{obj_fun}{a required function}

\item{...}{other objects needed to construct the args list}

\item{which}{see argument of the same name in \code{\link{sys.call}}}
}
\value{
a named list of formal args in position order followed by dots args
}
\description{
The point of this method is to undo partial argument matching on the formals
of the method that calls make_initialize_arg_list().
}
\keyword{internal}
