% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer-lib-utils.R
\name{maybe_get_vary_params_col}
\alias{maybe_get_vary_params_col}
\title{Convert \code{.vary_params} keyword to plotting variable in data}
\usage{
maybe_get_vary_params_col(var_str, vary_params)
}
\arguments{
\item{var_str}{String to check for \code{.vary_params} keyword.}

\item{vary_params}{Vector of parameter names being varied.}
}
\description{
Convert \code{.vary_params} keyword to plotting variable in data
}
\keyword{internal}
