% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-lib-errors.R
\name{norm_errors}
\alias{norm_errors}
\title{Generate heteroskedastic Gaussian errors based on the norm of X.}
\usage{
norm_errors(X)
}
\arguments{
\item{X}{Data design matrix}
}
\value{
A vector of simulated errors of length n.
}
\description{
Generate independent Gaussian errors with variance based on norm
of row observations in the data matrix X.
}
