% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-utils.R
\name{obj_size}
\alias{obj_size}
\title{Get the size of an object, including environments.}
\usage{
obj_size(obj = rlang::caller_env())
}
\arguments{
\item{obj}{The object to measure. Default is the calling environment.}
}
\description{
Get the size of an object, including environments.
}
\keyword{internal}
