% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-lib.R
\name{omitted_var_dgp}
\alias{omitted_var_dgp}
\title{Generate data from a model with omitted variable bias.}
\usage{
omitted_var_dgp(dgp_fun, omitted_vars = 1, ...)
}
\arguments{
\item{dgp_fun}{A function that generates data according to some known
data-generating process. This function should return an object of the same
format as the output of \code{return_DGP_output()}.}

\item{omitted_vars}{A vector of indices or column names corresponding to
columns in X that should be omitted.}

\item{...}{Additional arguments to pass to \code{dgp_fun()}.}
}
\value{
The returned object has the same format as the output of
\code{dgp_fun()}, except that specified variables, given by \code{omitted_vars}, have
been omitted from the \code{X} component and the \code{support} (if applicable).
}
\description{
Takes in a data-generating process (DGP), and induces some bias
due to omitted variable(s). In other words, this function will generate
a design matrix \code{X} and response vector \code{y} according to the inputted
DGP function, but will return a partially missing design matrix, where
some variable/feature columns have been omitted.
}
\examples{
# generate data from a linear gaussian DGP with the first variable missing
dgp_out <- omitted_var_dgp(dgp_fun = linear_gaussian_dgp,
                           n = 100, p_obs = 10, s_obs = 2,
                           omitted_vars = 1)
# or equivalently, (minus the difference in column names)
dgp_out <- linear_gaussian_dgp(n = 10, p_obs = 9, p_unobs = 1,
                               s_obs = 1, s_unobs = 1)

}
