% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rmd.R
\name{paste_md}
\alias{paste_md}
\title{Read Markdown file into R Markdown.}
\usage{
paste_md(filename)
}
\arguments{
\item{filename}{filepath of markdown file to read into R Markdown}
}
\description{
Read contents of Markdown file into R Markdown.
}
\references{
Adapted from \url{https://stackoverflow.com/questions/56328581/how-to-read-markdown-code-from-a-file-from-an-r-markdown-document}.
}
\keyword{internal}
