% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-templates.R
\name{pipe_value}
\alias{pipe_value}
\title{Find last non-comment line, add a \verb{\%>\%} to the end, then add another line.}
\usage{
pipe_value(base, value, expr_width = 80)
}
\description{
This function is adapted from the function of the same name in
the \code{usemodels} package.
}
\keyword{internal}
