% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer-lib-prediction-metrics.R
\name{plot_auc_curve}
\alias{plot_auc_curve}
\title{Plot ROC/PR curves.}
\usage{
plot_auc_curve(
  fit_results,
  eval_results,
  evaluator_name = NULL,
  vary_params = NULL,
  metric = c("ROC", "PR"),
  show = c("boxplot", "point", "line", "bar", "errorbar", "ribbon"),
  ...
)
}
\arguments{
\item{fit_results}{A tibble, as returned by the \code{Experiment$fit()}
method.}

\item{eval_results}{A list of result tibbles, as returned by the
\code{Experiment$evaluate()} method.}

\item{evaluator_name}{(Optional) Name of evaluator containing results to
plot. If not provided, evaluation results will be automatically
computed by calling \code{eval_fun()}. If the evaluation results have
already been computed previously, should provide \code{evaluator_name} to
avoid duplicate computations.}

\item{vary_params}{Vector of parameter names that are varied in the
\code{Experiment}'s run.}

\item{metric}{A character string. Either "ROC" or "PR".}

\item{show}{Character vector with elements being one of "boxplot", "point",
"line", "bar", "errorbar", "ribbon" indicating what plot(s) to construct.}

\item{...}{Additional arguments to pass to \code{plot_eval_summary()}.}
}
\value{
A ggplot object or list of ggplot objects.
}
\description{
Plot ROC/PR curves or some summary thereof across experimental
repetitions.
}
