% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer-lib-feature-selection.R
\name{plot_feature_importance}
\alias{plot_feature_importance}
\title{Plot feature importances.}
\usage{
plot_feature_importance(
  fit_results = NULL,
  eval_results = NULL,
  eval_name = NULL,
  eval_fun = "summarize_feature_importance",
  eval_fun_options = NULL,
  vary_params = NULL,
  feature_col,
  show_max_features = NULL,
  show = c("errorbar", "bar"),
  ...
)
}
\arguments{
\item{fit_results}{A tibble, as returned by the \code{fit} method.}

\item{eval_results}{A list of result tibbles, as returned by the
\code{evaluate} method.}

\item{eval_name}{Name of \code{Evaluator} containing results to plot. If
\code{NULL}, the data used for plotting is computed from scratch via
\code{eval_fun}.}

\item{eval_fun}{Character string, specifying the function used to compute
the data used for plotting if \code{eval_name = NULL}. If \code{eval_name}
is not \code{NULL}, this argument is ignored.}

\item{eval_fun_options}{List of named arguments to pass to \code{eval_fun}.}

\item{vary_params}{A vector of parameter names that are varied across in the
\code{Experiment}.}

\item{feature_col}{A character string identifying the column in
\code{fit_results} with the feature names or IDs.}

\item{show_max_features}{Maximum number of features to plot.}

\item{show}{Character vector with elements being one of "boxplot", "point",
"line", "bar", "errorbar", "ribbon" indicating what plot layer(s) to
construct.}

\item{...}{Not used.}
}
\value{
If \code{interactive = TRUE}, returns a \code{plotly} object if
\code{plot_by} is \code{NULL} and a list of \code{plotly} objects if
\code{plot_by} is not \code{NULL}. If \code{interactive = FALSE}, returns
a \code{ggplot} object if \code{plot_by} is \code{NULL} and a list of
\code{ggplot} objects if \code{plot_by} is not \code{NULL}.
}
\description{
Plot the raw or summarized feature importances as a boxplot,
scatter plot, line plot, or bar plot with or without 1 SD error bars.
}
\examples{
# generate example fit_results data
fit_results <- tibble::tibble(
  .rep = rep(1:2, times = 2),
  .dgp_name = c("DGP1", "DGP1", "DGP2", "DGP2"),
  .method_name = c("Method"),
  feature_info = lapply(
    1:4,
    FUN = function(i) {
      tibble::tibble(
        # feature names
        feature = c("featureA", "featureB", "featureC"),
        # estimated feature importance scores
        est_importance = c(10, runif(2, min = -2, max = 2))  
      )
    }
  )
)

# generate example eval_results data
eval_results <- list(
  `Feature Importance` = summarize_feature_importance(
    fit_results,
    nested_cols = "feature_info",
    feature_col = "feature",
    imp_col = "est_importance"
  )
)

# create bar plot using pre-computed evaluation results
plt <- plot_feature_importance(eval_results = eval_results,
                               eval_name = "Feature Importance",
                               feature_col = "feature")
# or alternatively, create the same plot directly from fit results
plt <- plot_feature_importance(fit_results = fit_results,
                               feature_col = "feature",
                               eval_fun_options = list(
                                 nested_cols = "feature_info",
                                 imp_col = "est_importance"
                               ))

# can customize plot (see plot_eval_constructor() for possible arguments)
plt <- plot_feature_importance(eval_results = eval_results,
                               eval_name = "Feature Importance",
                               feature_col = "feature",
                               errorbar_args = list(width = .5, position = "dodge"),
                               bar_args = list(width = .5))

}
\seealso{
Other feature_selection_funs: 
\code{\link{eval_feature_importance_funs}},
\code{\link{eval_feature_selection_curve_funs}},
\code{\link{eval_feature_selection_err_funs}},
\code{\link{plot_feature_selection_curve}()},
\code{\link{plot_feature_selection_err}()}
}
\concept{feature_selection_funs}
