% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer-lib-feature-selection.R
\name{plot_feature_recovery}
\alias{plot_feature_recovery}
\title{Plot feature recovery error according to various metrics}
\usage{
plot_feature_recovery(
  fit_results,
  eval_results,
  evaluator_name = NULL,
  vary_params = NULL,
  metrics = c("TPR", "FPR", "FDR", "AUROC", "AUPRC"),
  show = c("boxplot", "point", "line", "bar", "errorbar"),
  ...
)
}
\arguments{
\item{fit_results}{A tibble, as returned by the \code{Experiment$fit()}
method.}

\item{eval_results}{A list of result tibbles, as returned by the
\code{Experiment$evaluate()} method.}

\item{evaluator_name}{(Optional) Name of evaluator containing results to
plot. If not provided, evaluation results will be automatically
computed by calling \code{eval_fun()}. If the evaluation results have
already been computed previously, should provide \code{evaluator_name} to
avoid duplicate computations.}

\item{vary_params}{Vector of parameter names that are varied in the
\code{Experiment}'s run.}

\item{metrics}{Character vector of the support recovery metrics to plot. If
\code{evaluator_name} is provided, then this argument is optional, and the
default is to use all metrics found in the Evaluator's feature recover
results. Otherwise, this argument is required.}

\item{show}{Character vector with elements being one of "boxplot", "point",
"line", "bar", "errorbar", "ribbon" indicating what plot(s) to construct.}

\item{...}{Additional arguments to pass to \code{plot_eval_summary()}.}
}
\value{
A ggplot object or list of ggplot objects.
}
\description{
Plot raw or summarized feature recovery errors as a boxplot,
scatter plot, line plot, or bar plot with or without 1 SD error bars.
}
