% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer-lib-utils.R
\name{plot_fit_constructor}
\alias{plot_fit_constructor}
\title{Developer function for plotting results from particular replicate(s) in the
\code{Experiment} fit.}
\usage{
plot_fit_constructor(
  fit_results,
  vary_params = NULL,
  reps = 1,
  plot_fun,
  interactive = FALSE,
  ...
)
}
\arguments{
\item{fit_results}{A tibble, as returned by \code{fit_experiment()}.}

\item{vary_params}{A vector of parameter names that are varied across in the
\code{Experiment}.}

\item{reps}{Vector of replicates from which to plot results.}

\item{plot_fun}{The plotting function, which takes in the arguments
\code{fit_results}, \code{vary_params}, and possibly others passed from
\code{...}.}

\item{interactive}{Logical. If \code{TRUE}, returns interactive \code{plotly}
plots. If \code{FALSE}, returns static \code{ggplot} plots.}

\item{...}{Additional arguments to pass to \code{plot_fun()}.}
}
\value{
If \code{interactive = TRUE}, returns a \code{plotly} object or
list of \code{plotly} objects if there are multiple replicates, DGPs, or
Methods to plot. If \code{interactive = FALSE}, returns a \code{ggplot}
object or list of \code{ggplot} objects if there are multiple replicates,
DGPs, or Methods to plot.
}
\description{
A helper function for developing new \code{Visualizer} plotting
functions that plot results from particular replicate(s) in the
\code{Experiment} fit. This function will construct one plot for each
row in the \code{Experiment}'s \code{fit_results} from the specified
replicates.
}
\examples{
# generate example fit results data
fit_results <- tibble::tibble(
  .rep = rep(1:2, times = 2),
  .dgp_name = c("DGP1", "DGP1", "DGP2", "DGP2"),
  .method_name = c("Method"),
  # true response
  y = lapply(1:4, FUN = function(x) rnorm(100)),
  # predicted response
  predictions = lapply(1:4, FUN = function(x) rnorm(100))
)

# function to plot scatter plot of y vs predictions
plot_fun <- function(fit_results, vary_params = NULL) {
  plt <- fit_results \%>\%
    tidyr::unnest(c("y", "predictions")) \%>\%
    ggplot2::ggplot() +
    ggplot2::aes(x = y, y = predictions) +
    ggplot2::geom_point() +
    ggplot2::labs(title = sprintf("DGP: \%s | Method: \%s | Rep: \%s",
                                  fit_results$.dgp_name,
                                  fit_results$.method_name,
                                  fit_results$.rep))
  return(plt)
}

# returns the scatter plot for each (DGP, Method) combination from rep 1
plt <- plot_fit_constructor(fit_results, reps = 1, plot_fun = plot_fun)

}
