% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer-lib-utils.R
\name{plot_fit_results}
\alias{plot_fit_results}
\title{Developer function for plotting results from particular replicate(s) in the
\code{Experiment} fit.}
\usage{
plot_fit_results(
  fit_results,
  vary_params = NULL,
  reps = 1,
  plot_fun,
  interactive,
  ...
)
}
\arguments{
\item{fit_results}{A tibble, as returned by the \code{fit} method.}

\item{vary_params}{A vector of parameter names that are varied across in the
\code{Experiment}.}

\item{reps}{Vector of replicates from which to plot results.}

\item{plot_fun}{The plotting function, which takes in the arguments
\code{fit_results}, \code{vary_params}, and possibly others passed from
\code{...}.}

\item{interactive}{Logical. If \code{TRUE}, returns interactive \code{plotly}
plots. If \code{FALSE}, returns static \code{ggplot} plots.}

\item{...}{Additional arguments to pass to \code{plot_fun()}.}
}
\value{
If \code{interactive = TRUE}, returns a \code{plotly} object or
list of \code{plotly} objects if there are multiple replicates, DGPs, or
Methods to plot. If \code{interactive = FALSE}, returns a \code{ggplot}
object or list of \code{ggplot} objects if there are multiple replicates,
DGPs, or Methods to plot.
}
\description{
A helper function for developing new \code{Visualizer} plotting
functions that plot results from particular replicate(s) in the
\code{Experiment} fit. This function will construct one plot for each
row in the \code{Experiment}'s \code{fit_results} from the specified
replicates.
}
