% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer-lib-prediction-metrics.R
\name{plot_pred_err}
\alias{plot_pred_err}
\title{Plot prediction error according to various metrics.}
\usage{
plot_pred_err(
  fit_results,
  eval_results,
  evaluator_name = NULL,
  vary_params = NULL,
  metrics = NULL,
  group = NULL,
  show = c("boxplot", "point", "line", "bar", "errorbar"),
  ...
)
}
\arguments{
\item{fit_results}{A tibble, as returned by the \code{Experiment$fit()}
method.}

\item{eval_results}{A list of result tibbles, as returned by the
\code{Experiment$evaluate()} method.}

\item{evaluator_name}{(Optional) Name of evaluator containing results to
plot. If not provided, evaluation results will be automatically
computed by calling \code{eval_fun()}. If the evaluation results have
already been computed previously, should provide \code{evaluator_name} to
avoid duplicate computations.}

\item{vary_params}{Vector of parameter names that are varied in the
\code{Experiment}'s run.}

\item{metrics}{Prediction error metrics for which to plot results. If
\code{evaluator_name} is provided, then this argument is optional, and the
default is to use all metrics found in the Evaluator's prediction error
results. Otherwise, this argument is required.}

\item{group}{(Optional) vector of group ids (as a factor) to use for
assessing within-group prediction errors.}

\item{show}{Character vector with elements being one of "boxplot", "point",
"line", "bar", "errorbar", "ribbon" indicating what plot(s) to construct.}

\item{...}{Additional arguments to pass to \code{plot_eval_summary()}.}
}
\value{
A ggplot object or list of ggplot objects.
}
\description{
Plot raw or summarized prediction errors as a boxplot, scatter
plot, line plot, or bar plot with or without 1 SD error bars.
}
