% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer-lib-inference.R
\name{plot_testing_curve}
\alias{plot_testing_curve}
\title{Plot ROC/PR curves for feature rankings, ranked by p-values.}
\usage{
plot_testing_curve(
  fit_results,
  eval_results = NULL,
  evaluator_name = NULL,
  vary_params = NULL,
  curve = c("ROC", "PR"),
  show = c("line", "ribbon"),
  ...
)
}
\arguments{
\item{fit_results}{A tibble, as returned by the \code{fit} method.}

\item{eval_results}{A list of result tibbles, as returned by the
\code{evaluate} method.}

\item{evaluator_name}{Name of \code{Evaluator} containing results to plot.
To compute the evaluation summary results from scratch or if the evaluation
summary results have not yet been evaluated, set to \code{NULL}.}

\item{vary_params}{A vector of parameter names that are varied across in the
\code{Experiment}.}

\item{curve}{Either "ROC" or "PR" indicating whether to plot the ROC or
Precision-Recall curve.}

\item{show}{Character vector with elements being one of "boxplot", "point",
"line", "bar", "errorbar", "ribbon" indicating what plot layer(s) to
construct.}

\item{...}{Additional arguments to pass to \code{plot_eval_summary()}. This
includes arguments for plotting and for passing into
\code{summarize_testing_curve()}.}
}
\value{
If \code{interactive = TRUE}, returns a \code{plotly} object if
\code{plot_by} is \code{NULL} and a list of \code{plotly} objects if
\code{plot_by} is not \code{NULL}. If \code{interactive = FALSE}, returns
a \code{ggplot} object if \code{plot_by} is \code{NULL} and a list of
\code{ggplot} objects if \code{plot_by} is not \code{NULL}.
}
\description{
Plot ROC/PR curves for the feature rankings, ranked by their
p-values or some summary thereof across experimental replicates.
}
\examples{
# generate example fit_results data
fit_results <- tibble::tibble(
  .rep = rep(1:2, times = 2),
  .dgp_name = c("DGP1", "DGP1", "DGP2", "DGP2"),
  .method_name = c("Method"),
  feature_info = lapply(
    1:4,
    FUN = function(i) {
      tibble::tibble(
        # feature names
        feature = c("featureA", "featureB", "featureC"),  
        # true feature support
        true_support = c(TRUE, FALSE, TRUE),  
        # estimated p-values
        pval = 10^(sample(-3:0, 3, replace = TRUE))
      )
    }
  )
)

# generate example eval_results data
eval_results <- list(
  `ROC` = summarize_testing_curve(
    fit_results, 
    curve = "ROC",
    nested_data = "feature_info",
    truth_col = "true_support", 
    pval_col = "pval"
  ),
  `PR` = summarize_testing_curve(
    fit_results, 
    curve = "PR",
    nested_data = "feature_info",
    truth_col = "true_support", 
    pval_col = "pval"
  )
) 

# create summary ROC/PR plots using pre-computed evaluation results
roc_plt <- plot_testing_curve(fit_results = fit_results,
                              eval_results = eval_results,
                              evaluator_name = "ROC", curve = "ROC",
                              show = c("line", "ribbon"))
pr_plt <- plot_testing_curve(fit_results = fit_results,
                             eval_results = eval_results,
                             evaluator_name = "PR", curve = "PR",
                             show = c("line", "ribbon"))
# or alternatively, create the same plots without pre-computing evaluation results
roc_plt <- plot_testing_curve(fit_results, show = c("line", "ribbon"),
                              nested_data = "feature_info",
                              truth_col = "true_support",
                              pval_col = "pval",
                              curve = "ROC")
pr_plt <- plot_testing_curve(fit_results, show = c("line", "ribbon"),
                             nested_data = "feature_info",
                             truth_col = "true_support",
                             pval_col = "pval",
                             curve = "PR")

# can customize plot (see plot_eval_summary() for possible arguments)
roc_plt <- plot_testing_curve(fit_results = fit_results,
                              eval_results = eval_results,
                              evaluator_name = "ROC", curve = "ROC",
                              show = c("line", "ribbon"),
                              plot_by = ".dgp_name")

}
\seealso{
Other inference_funs: 
\code{\link{eval_reject_prob}()},
\code{\link{eval_testing_curve_funs}},
\code{\link{eval_testing_err_funs}},
\code{\link{plot_reject_prob}()},
\code{\link{plot_testing_err}()}
}
\concept{inference_funs}
