% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\docType{class}
\name{Plotter}
\alias{Plotter}
\title{\code{R6} class representing a plotter.}
\description{
A plotter (or plotting function) to \strong{plot} the performance of
methods and/or its evaluation metrics from the \code{Experiment} run.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the \code{Plotter}.}

\item{\code{plot_fun}}{The plotting function.}

\item{\code{plot_params}}{(Named) list of parameters to input into the
plotting function.}

\item{\code{rmd_options}}{List of options to control the aesthetics of the
\code{Plotter}'s displayed plot in the knitted R Markdown report.
Currently, possible options are "height" and "width" (in inches).}

\item{\code{show}}{If \code{TRUE} (default), show the resulting plot in the R
Markdown report; if \code{FALSE}, hide output in the R Markdown report.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Plotter$new()}}
\item \href{#method-plot}{\code{Plotter$plot()}}
\item \href{#method-clone}{\code{Plotter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Plotter}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plotter$new(plot_fun, name = NULL, rmd_options = list(), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plot_fun}}{The plotting function.}

\item{\code{name}}{(Optional) The name of the \code{Plotter}.}

\item{\code{rmd_options}}{(Optional) List of options to control the aesthetics
of the \code{Plotter}'s displayed plot in the knitted R Markdown
report. Currently, possible options are "height" and "width" (in
inches).}

\item{\code{...}}{Arguments to pass into \code{plot_fun()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
When plotting or running the \code{Experiment} (see
\code{Experiment$plot() or \code{Experiment$run()}}), the named
arguments \code{fit_results}, \code{eval_results}, and
\code{vary_params} are automatically passed into the \code{Plotter}
function \code{plot_fun()} and serve as placeholders for the
\code{Experiment$fit()} results, the \code{Experiment$evaluate()}
results, and the name of the varying parameter, respectively.
To plot the performance of a method(s) fit and/or its evaluation
metrics then, the \code{Plotter} function \code{plot_fun()} should
take in the named arguments \code{fit_results} and/or
\code{eval_results}. See \code{Experiment$fit()} or
\code{fit_experiment()} for details on the format of \code{fit_results}.
See \code{Experiment$evaluate()} or \code{evaluate_experiment()} for
details on the format of \code{eval_results}. If the \code{Plotter}
is used for \code{Experiments} with varying parameters,
\code{vary_params} should be used as a stand in for the name of this
varying parameter.
}

\subsection{Returns}{
A new \code{Plotter} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Plot the performance of methods and/or their evaluation
metrics from the \code{Experiment} using the \code{Plotter} and the
provided parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plotter$plot(fit_results = NULL, eval_results = NULL, vary_params = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fit_results}}{A tibble, typically returned by the
\code{Experiment$fit()} method.}

\item{\code{eval_results}}{A list of tibbles, typically returned by the
\code{Experiment$evaluate()} method.}

\item{\code{vary_params}}{Name of parameter/argument that was varied in the
\code{Experiment} (see \code{Experiment$get_vary_across()}).
Use \code{NULL} (default) if no \code{vary_across} component in
\code{Experiment} run.}

\item{\code{...}}{Not used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Result of \code{plot_fun()}, which should be a plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plotter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
