% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tables.R
\name{pretty_DT}
\alias{pretty_DT}
\title{Create pretty datatable with custom bolding options.}
\usage{
pretty_DT(
  X,
  digits = 3,
  sigfig = T,
  escape = F,
  rownames = TRUE,
  caption = "",
  na_disp = "NA",
  bold_function = NULL,
  bold_margin = NULL,
  bold_scheme = T,
  bold_color = NULL,
  options = list(),
  return_df = FALSE,
  ...
)
}
\arguments{
\item{X}{Data frame or data matrix to display in table}

\item{digits}{Number of digits to display for numeric values}

\item{sigfig}{Logical. If \code{TRUE}, \code{digits} refers to the number of
significant figures. If \code{FALSE}, \code{digits} refers to the number of
decimal places.}

\item{escape}{Logical. Whether or not to escape HTML entities in table. See
\code{\link[DT:datatable]{DT::datatable()}} for details.}

\item{rownames}{Logical. Whether or not to show rownames in table. See
\code{\link[DT:datatable]{DT::datatable()}} for details.}

\item{caption}{The table caption.}

\item{na_disp}{Character string to display if NA entry is found in table.}

\item{bold_function}{Optional function string or vector of function strings
to use for bolding entries, e.g. ". == max(.)" or ". >= 0.5".}

\item{bold_margin}{Specifies the margins of X that will be used to evaluate
\code{bold_function} across, i.e., 0 = across entire matrix, 1 = across
rows, and 2 = across columns. Required if \code{bold_function = TRUE}.}

\item{bold_scheme}{Scalar or vector of logicals, indicating whether or not
to apply \code{bold_function} to row if \code{bold_margin} = 1 and to
column if \code{bold_margin} = 1, 2. Default is to apply bolding to all
rows/columns.}

\item{bold_color}{Color of bolded text.}

\item{options}{See \code{options} argument in \code{\link[DT:datatable]{DT::datatable()}}.}

\item{return_df}{Logical. If \code{TRUE}, return data frame that was used
as input into \code{DT::datatable()} in addition to the datatable output.
If \code{FALSE}, only return the datatable output.}

\item{...}{Additional arguments to pass to \code{\link[DT:datatable]{DT::datatable()}}.}
}
\value{
If \code{return_df = FALSE}, returns a datatable object. Otherwise,
returns a list of two:
\describe{
\item{dt}{A datatable object.}
\item{df}{A data frame that was used as input into \code{DT::datatable()}.}
}
}
\description{
Create pretty datatable with custom bolding options.
}
\examples{
## Show iris data table
pretty_DT(iris, caption = "Iris Data Table")

## Bold max value of each numeric column of Iris data in red
pretty_DT(iris, caption = "Iris Data Table",
         bold_function = ". == max(.)", bold_margin = 2,
         bold_scheme = c(TRUE, TRUE, TRUE, TRUE, FALSE), bold_color = "red")
            
## Bold min value of each row in Iris data
pretty_DT(iris \%>\% dplyr::select(-Species), 
         sigfig = TRUE, caption = "Iris Data Table",
         na_disp = "NA", bold_function = ". == min(.)", bold_margin = 1,
         bold_scheme = TRUE, bold_color = "black")
}
