% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer-lib-ggthemes.R
\name{pretty_ggplot_color_fill_themes}
\alias{pretty_ggplot_color_fill_themes}
\alias{pretty_ggplot_color}
\alias{pretty_ggplot_fill}
\title{Customized pretty ggplot color and fill themes}
\usage{
pretty_ggplot_color(color, viridis = F, option = "plasma", drop = T, ...)

pretty_ggplot_fill(fill, viridis = F, option = "plasma", drop = T, ...)
}
\arguments{
\item{color}{Vector used for color aesthetic. Should match
\code{ggplot2::aes(color = ...)} argument.}

\item{viridis}{Logical. Whether or not to use \code{viridis} scheme if using
discrete color scheme.}

\item{option}{Argument indicating \code{viridis} palette name.}

\item{drop}{Logical; whether or not to drop factors with no observations.}

\item{...}{Other arguments to pass to \code{\link[ggplot2:scale_manual]{ggplot2::scale_color_manual()}},
\code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}}, \code{\link[viridis:scale_viridis]{viridis::scale_colour_viridis()}}, or
\code{\link[viridis:scale_viridis]{viridis::scale_fill_viridis()}}.}

\item{fill}{Vector used for fill aesthetic. Should match
\code{ggplot2::aes(fill = ...)} argument.}
}
\value{
A ggplot color or fill theme object.
}
\description{
Customized pretty ggplot color and fill themes
}
\examples{
require(ggplot2)
ggplot(iris) +
  aes(x = Sepal.Length, y = Sepal.Width, color = Species) +
  geom_point() +
  pretty_ggplot_theme() +
  pretty_ggplot_color(color = iris$Species)
ggplot(iris) +
  aes(x = Sepal.Length, fill = Species) +
  geom_density() +
  pretty_ggplot_theme() +
  pretty_ggplot_fill(fill = iris$Species)
  
}
