% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tables.R
\name{pretty_kable}
\alias{pretty_kable}
\title{Create pretty kable tables with custom bolding options.}
\usage{
pretty_kable(
  X,
  digits = 3,
  sigfig = T,
  align = "c",
  caption = "",
  format = c("html", "latex"),
  na_disp = "NA",
  bold_function = NULL,
  bold_margin = NULL,
  bold_scheme = T,
  bold_color = NULL,
  full_width = NULL,
  position = "center",
  font_size = NULL,
  fixed_thead = F,
  scroll = F,
  scroll_width = NULL,
  scroll_height = NULL,
  return_df = FALSE,
  ...
)
}
\arguments{
\item{X}{Data frame or data matrix to display in table}

\item{digits}{Number of digits to display for numeric values}

\item{sigfig}{Logical. If \code{TRUE}, \code{digits} refers to the number of
significant figures. If \code{FALSE}, \code{digits} refers to the number of
decimal places.}

\item{align}{A character vector indicating the column alignment, e.g., 'c'.
For further details, see \code{\link[knitr:kable]{knitr::kable()}}.}

\item{caption}{The table caption.}

\item{format}{One of "html" or "latex", indicating the output format.}

\item{na_disp}{Character string to display if NA entry is found in table.}

\item{bold_function}{Optional function string or vector of function strings
to use for bolding entries, e.g. ". == max(.)" or ". >= 0.5".}

\item{bold_margin}{Specifies the margins of X that will be used to evaluate
\code{bold_function} across, i.e., 0 = across entire matrix, 1 = across
rows, and 2 = across columns. Required if \code{bold_function = TRUE}.}

\item{bold_scheme}{Scalar or vector of logicals, indicating whether or not
to apply \code{bold_function} to row if \code{bold_margin} = 1 and to
column if \code{bold_margin} = 1, 2. Default is to apply bolding to all
rows/columns.}

\item{bold_color}{Color of bolded text.}

\item{full_width}{Logical. Whether or not table should have full width. See
\code{\link[kableExtra:kable_styling]{kableExtra::kable_styling()}} for details.}

\item{position}{Character string determining how to position table on page;
possible values inclue "left", "right", "center", "float_left",
"float_right". See \code{\link[kableExtra:kable_styling]{kableExtra::kable_styling()}} for details.}

\item{font_size}{A numeric input for table font size. See
\code{\link[kableExtra:kable_styling]{kableExtra::kable_styling()}} for details.}

\item{fixed_thead}{Logical. Whether or not table header should be fixed at
top. See \code{\link[kableExtra:kable_styling]{kableExtra::kable_styling()}} for details.}

\item{scroll}{Logical. If \code{TRUE}, add scroll box. Only used if
\code{format = "html"}.}

\item{scroll_width}{A character string indicating width of the scroll box,
e.g., "50px", "100\%". See \code{\link[kableExtra:scroll_box]{kableExtra::scroll_box()}} for details.}

\item{scroll_height}{A character string indicating height of the scroll box,
e.g. "100px". See \code{\link[kableExtra:scroll_box]{kableExtra::scroll_box()}} for details.}

\item{return_df}{Logical. If \code{TRUE}, return data frame that was used
as input into \code{knitr::kable()} in addition to the kable output.
If \code{FALSE}, only return the kable output.}

\item{...}{Additional arguments to pass to \code{\link[knitr:kable]{knitr::kable()}}.}
}
\value{
If \code{return_df = FALSE}, returns a kable object. Otherwise,
returns a list of two:
\describe{
\item{kable}{A kable object.}
\item{df}{A data frame that was used as input into \code{knitr::kable()}.}
}
}
\description{
Make pretty kable tables and enable easy bolding of cells in
kable (across rows, columns, or the entire table).
}
\examples{
## Show iris data table
pretty_kable(iris, align = "c", caption = "Iris Data Table")

## Bold max value of each numeric column of Iris data in red
pretty_kable(iris, caption = "Iris Data Table", scroll = TRUE,
            bold_function = ". == max(.)", bold_margin = 2,
            bold_scheme = c(TRUE, TRUE, TRUE, TRUE, FALSE), 
            bold_color = "red")
            
## Bold min value of each row in Iris data
pretty_kable(iris \%>\% dplyr::select(-Species), sigfig = TRUE, 
            caption = "Iris Data Table", format = "latex", 
            scroll = TRUE, na_disp = "NA",
            bold_function = ". == min(.)", bold_margin = 1,
            bold_scheme = TRUE, bold_color = "black")
}
