% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer-lib-ggthemes.R
\name{prettyGGplotColor}
\alias{prettyGGplotColor}
\title{Customized pretty ggplot color theme}
\usage{
prettyGGplotColor(color, viridis = F, option = "plasma", drop = T, ...)
}
\arguments{
\item{color}{Vector used for color aesthetic. Should match
\code{ggplot2::aes(color = ...)} argument.}

\item{viridis}{Logical. Whether or not to use viridis scheme if using
discrete color scheme}

\item{option}{Argument indicating viridis palette name}

\item{drop}{Logical; whether or not to drop factors with no observations.}

\item{...}{Other arguments to pass to scale_color_manual() or
scale_colour_viridis()}
}
\value{
A ggplot color theme object.
}
\description{
Customized pretty ggplot color theme
}
\examples{
require(ggplot2)
ggplot(iris) +
  aes(x = Sepal.Length, y = Sepal.Width, color = Species) +
  geom_point() +
  prettyGGplotTheme() +
  prettyGGplotColor(color = iris$Species)
  
}
