% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer-lib-ggthemes.R
\name{prettyGGplotFill}
\alias{prettyGGplotFill}
\title{Customized pretty ggplot fill theme}
\usage{
prettyGGplotFill(fill, viridis = F, option = "plasma", drop = T, ...)
}
\arguments{
\item{fill}{Vector used for fill aesthetic. Should match
\code{ggplot2::aes(fill = ...)} argument.}

\item{viridis}{Logical. Whether or not to use viridis scheme if using
discrete fill scheme}

\item{option}{Argument indicating viridis palette name}

\item{drop}{Logical; whether or not to drop factors with no observations.}

\item{...}{Other arguments to pass to scale_fill_manual() or
scale_fill_viridis()}
}
\value{
A ggplot color theme object.
}
\description{
Customized pretty ggplot fill theme
}
\examples{
require(ggplot2)
ggplot(iris) +
  aes(x = Sepal.Length, fill = Species) +
  geom_density() +
  prettyGGplotTheme() +
  prettyGGplotFill(fill = iris$Species)
  
}
