% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{remove_funs}
\alias{remove_funs}
\alias{remove_dgp}
\alias{remove_method}
\alias{remove_evaluator}
\alias{remove_visualizer}
\title{Helper functions for removing components of an \code{Experiment}.}
\usage{
remove_dgp(experiment, name, ...)

remove_method(experiment, name, ...)

remove_evaluator(experiment, name, ...)

remove_visualizer(experiment, name, ...)
}
\arguments{
\item{name}{A name to identify the object to be removed.}

\item{...}{Not used.}
}
\value{
The original \code{Experiment} object passed to \code{remove_*}.
}
\description{
Helper functions for removing \code{DGPs}, \code{Methods},
\code{Evaluators}, and \code{Visualizers} already added to an
\code{Experiment}.
}
