% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_source}
\alias{remove_source}
\title{Remove source code from functions.}
\usage{
remove_source(x, cols = NULL)
}
\arguments{
\item{x}{A tibble.}

\item{cols}{Character vector of column names to potentially remove source
code from.}
}
\value{
A tibble with the source code removed from functions in the
specified columns.
}
\description{
Remove the source code from functions so that caching works as
expected. This function is used internally to remove the source code from
functions in the \code{.dgp_params}, \code{.method_params}, and \code{vary_across}
components.
}
\keyword{internal}
