% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{remove_vary_across}
\alias{remove_vary_across}
\title{Remove variable parameters from an \code{Experiment}.}
\usage{
remove_vary_across(experiment)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}
}
\value{
The original \code{experiment} object passed to \code{remove_vary_across}.
}
\description{
Remove variable parameters from an \code{Experiment}.
}
