% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-helpers.R
\name{rename_funs}
\alias{rename_funs}
\alias{rename_dgps}
\alias{rename_methods}
\alias{rename_evaluators}
\alias{rename_visualizers}
\title{Helper functions for renaming components in an \code{Experiment}.}
\usage{
rename_dgps(experiment, ...)

rename_methods(experiment, ...)

rename_evaluators(experiment, ...)

rename_visualizers(experiment, ...)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{...}{Named character vector of names to rename. Use
\code{"new_name" = "old_name"} format.}
}
\value{
The original \code{Experiment} object passed to \verb{remove_*}.
}
\description{
Helper functions for renaming \code{DGPs},
\code{Methods}, \code{Evaluators}, and \code{Visualizers} from an
\code{Experiment}. These functions will also rename these objects in the
cached documentation and results files.
}
\examples{
\dontrun{
rename_dgps(experiment, "New DGP1" = "DGP1", "New DGP2" = "DGP2")
rename_methods(experiment, "New Method1" = "Method1")
rename_evaluators(experiment, "New Evaluator1" = "Evaluator1")
rename_visualizers(experiment, "New Visualizer1" = "Visualizer1")
}

}
