% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-lib-utils.R
\name{rescale_curve}
\alias{rescale_curve}
\title{Rescale ROC/PR curves onto the same x-axis grid}
\usage{
rescale_curve(curve_data, x_grid, xvar, yvar)
}
\arguments{
\item{curve_data}{A \code{data.frame} containing the x- and y-coordinates
that define the ROC/PR curve.}

\item{x_grid}{Vector of x-coordinates at which to evaluate ROC/PR curve}

\item{xvar}{Name of column in \code{curve_data} with the FPR values for an
ROC curve or the recall values for a PR curve.}

\item{yvar}{Name of column in \code{curve_data} with the TPR values for an
ROC curve or the precision values for a PR curve.}
}
\value{
A \code{data.frame} with the coordinates of the ROC/PR curve using
the new x-axis scale. This \code{data.frame} has two columns with names
given by those specified in \code{xvar} and \code{yvar}.
}
\description{
A helper function to map a ROC/PR curve with unique
coordinates given in a data.frame onto a new set of x-axis values (i.e.,
FPR for an ROC curve and recall for a PR curve).
}
\keyword{internal}
