% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{run_experiment}
\alias{run_experiment}
\title{Run an \code{Experiment}.}
\usage{
run_experiment(experiment, ...)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{...}{Passed to \code{experiment$run()}.}
}
\value{
A list of the simulation experiment's method results.
}
\description{
Run an \code{Experiment}.
}
