% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-tests.R
\name{run_tests}
\alias{run_tests}
\title{Run Tests}
\usage{
run_tests()
}
\description{
\code{run_tests()} runs the tests in the \verb{tests/} directory.
This function only works if the simulation study is set up as an R project
and if it is run when this R project is active.
}
\examples{
\dontrun{
# run tests in the tests/ directory
run_tests()}

}
