% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-lib.R
\name{rwd_dgp}
\alias{rwd_dgp}
\title{Read in real world data from X and y.}
\usage{
rwd_dgp(X, y, data_split, train_prop = 0.5, return_values = c("X", "y"))
}
\arguments{
\item{X}{Data matrix.}

\item{y}{Response vector.}

\item{data_split}{Logical; if \code{TRUE}, splits data into training and test
sets according to \code{train_prop}.}

\item{train_prop}{Proportion of data in training set if
\code{data_split = TRUE}.}

\item{return_values}{Character vector indicating what objects to return in
list. Elements in vector must be one of "X", "y", "support".}
}
\description{
Read in real world data from X and y.
}
