% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-helpers.R
\name{save_experiment}
\alias{save_experiment}
\title{Save an \code{Experiment}.}
\usage{
save_experiment(experiment)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}
}
\value{
The original \code{Experiment} object passed to
\code{save_experiment}.
}
\description{
Save an \code{Experiment} object to a .rds file under the
\code{Experiment}'s results directory (see \code{\link[=get_save_dir]{get_save_dir()}}).
}
\examples{
\dontrun{
save_experiment(experiment)}

}
