% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{set_save_dir}
\alias{set_save_dir}
\title{Set base directory to use for saving results and plots from an
\code{Experiment}.}
\usage{
set_save_dir(experiment, save_dir)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{save_dir}{A directory to use for saving \verb{experiment's} results.}
}
\value{
The original \code{experiment} object passed to \code{set_save_dir}.
}
\description{
Set base directory to use for saving results and plots from an
\code{Experiment}.
}
