% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-lib-utils.R
\name{shared_dgp_lib_args}
\alias{shared_dgp_lib_args}
\title{Arguments that are shared by multiple \code{DGP} library functions.}
\arguments{
\item{data_split}{Logical; if \code{TRUE}, splits data into training and test
sets according to \code{train_prop}.}

\item{err}{Function from which to generate simulate error vector. Default
is \code{NULL} which adds no error to the DGP.}

\item{intercept}{Scalar intercept term.}

\item{n}{Number of samples.}

\item{p}{Number of features.}

\item{return_support}{Logical specifying whether or not to return a vector
of the support column names. If \code{X} has no column names, then the
indices of the support are used.}

\item{return_values}{Character vector indicating what objects to return in
list. Elements in vector must be one of "X", "y", "support".}

\item{support}{Vector of feature indices in the true support of the DGP.}

\item{train_prop}{Proportion of data in training set if
\code{data_split = TRUE}.}

\item{X}{Data matrix or data frame.}

\item{y}{Response vector.}

\item{...}{Other arguments to pass to err() to generate the error vector.}
}
\value{
A list of the named objects that were requested in
\code{return_values}. Note that if \code{data_split = TRUE} and "X", "y"
are in \code{return_values}, then the returned list also contains slots for
"Xtest" and "ytest".
}
\description{
Arguments that are shared by multiple \code{DGP} library functions.
}
