% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simChef_example_funs}
\alias{simChef_example_funs}
\alias{linear_dgp}
\alias{linear_method}
\alias{rf_method}
\alias{prediction_error_evaluator}
\alias{prediction_error_plot}
\alias{example_experiment}
\alias{example_fit_results}
\alias{example_eval_results}
\alias{example_viz_results}
\title{simChef Example Components and Results Data}
\format{
\subsection{\code{linear_dgp}}{

An object of class \link{DGP}
}

\subsection{\code{linear_method}}{

An object of class \link{Method}
}

\subsection{\code{rf_method}}{

An object of class \link{Method}
}

\subsection{\code{prediction_error_evaluator}}{

An object of class \link{Evaluator}
}

\subsection{\code{prediction_error_plot}}{

An object of class \link{Visualizer}
}

\subsection{\code{example_experiment}}{

An object of class \link{Experiment}
}

\subsection{\code{example_fit_results}}{

A \code{tibble} with 20 rows and 5 columns containing the fitted results from
fitting all \code{Methods} across all \code{DGPs} in the \code{example_experiment} for 10
repetitions. The columns are:
\describe{
\item{.rep}{Replicate ID}
\item{.dgp_name}{Name of data-generating process}
\item{.method_name}{Name of method}
\item{predictions}{List of elements, each a 200-length vector of the
test predictions from the given DGP and method combination}
\item{y_test}{List of elements, each a 200-length vector of the true test
responses from the given DGP}
}
}

\subsection{\code{example_eval_results}}{

A named list of 1 element containing a \code{tibble} with the prediction error
results from evaluating the \code{Evaluators} in the \code{example_experiment}.
This \code{tibble} has 6 rows and 9 columns:
\describe{
\item{.dgp_name}{Name of data-generating process}
\item{.method_name}{Name of method}
\item{.metric}{Name of evaluation metric, either "rsq" (R-squared), "rmse"
(root mean squared error), or "mae" (mean absolute error)}
\item{mean_pred_err}{Mean prediction error across all 10 simulation
replicates}
\item{median_pred_err}{Median prediction error across all 10 simulation
replicates}
\item{min_pred_err}{Minimum prediction error across all 10 simulation
replicates}
\item{max_pred_err}{Maximum prediction error across all 10 simulation
replicates}
\item{sd_pred_err}{Standard deviation of prediction error across all 10
simulation replicates}
\item{raw_pred_err}{List of elements, each a vector of the raw prediction
errors from the 10 simulation replicates}
}
}

\subsection{\code{example_viz_results}}{

A named list of 1 element containing a \code{ggplot} object, visualizing the
prediction error results from the \code{example_experiment}.
}
}
\usage{
linear_dgp

linear_method

rf_method

prediction_error_evaluator

prediction_error_plot

example_experiment

example_fit_results

example_eval_results

example_viz_results
}
\description{
Collection of example \link{Experiment} components (\link{DGP}s, \link{Method}s,
\link{Evaluator}s, and \link{Visualizer}s) and results datasets (\code{fit_results},
\code{eval_results}, \code{viz_results}).
}
\section{Example \code{Experiment} Components}{


\code{linear_dgp}:
An example \link{DGP} which generates training and test data according to the
classical linear regression model: y = X*beta + noise.

\code{linear_method}:
An example \link{Method} which fits a linear regression model to the training data
and makes predictions on the test data.

\code{rf_method}:
An example \link{Method} which fits a random forest model to the training data
and makes predictions on the test data.

\code{prediction_error_evaluator}:
An example \link{Evaluator} which calculates the prediction error between the
predicted and true predicted values.

\code{prediction_error_plot}:
An example \link{Visualizer} which plots the prediction error.

\code{example_experiment}:
An example \link{Experiment} with the following components:
\itemize{
\item DGP: \code{linear_dgp}
\item Method: \code{linear_method}, \code{rf_method}
\item Evaluator: \code{prediction_error_evaluator}
\item Visualizer: \code{prediction_error_plot}
}
}

\section{Example \code{Experiment} Results Datasets}{


\code{example_fit_results}:
An example of the fitted results from an \link{Experiment}, computed via
\code{\link[=run_experiment]{run_experiment()}} (or more directly, \code{\link[=fit_experiment]{fit_experiment()}}) applied to
\code{example_experiment}.

\code{example_eval_results}:
An example of the evaluation results from an \link{Experiment}, computed via
\code{\link[=run_experiment]{run_experiment()}} (or more directly, \code{\link[=evaluate_experiment]{evaluate_experiment()}}) applied to
\code{example_experiment}.

\code{example_viz_results}:
An example of the visualization results from an \link{Experiment}, computed via
\code{\link[=run_experiment]{run_experiment()}} (or more directly, \code{\link[=visualize_experiment]{visualize_experiment()}}) applied to
\code{example_experiment}.
}

\examples{
data(linear_dgp)
data(linear_method)
data(rf_method)
data(prediction_error_evaluator)
data(prediction_error_plot)
data(example_experiment)
data(example_fit_results)
data(example_eval_results)
data(example_viz_results)

}
\keyword{datasets}
