% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rmd.R
\name{subchunkify}
\alias{subchunkify}
\title{Subchunkify R chunks in R Markdown.}
\usage{
subchunkify(
  g,
  i,
  fig_height = 12,
  fig_width = 10,
  caption = "''",
  add_class = NULL,
  other_args = ""
)
}
\arguments{
\item{g}{Plot or object to display in R Markdown output.}

\item{i}{Chunk id. Should be unique for each plot or subchunk.}

\item{fig_height}{Height of plot or output of subchunk (in inches). See
\code{fig.height} in the knitr chunk options.}

\item{fig_width}{Width of plot or output of subchunk (in inches). See
\code{fig.width} in the knitr chunk options.}

\item{caption}{Figure or subchunk caption. Should be surrounded by two sets
of quotes, e.g., "'This is a valid caption.'" See \code{fig.cap} in the
knitr chunk options.}

\item{add_class}{Vector of css classes to add to object.}

\item{other_args}{Other arguments to pass to knitr chunk header. Should be
surrounded by two sets of quotes, e.g., "results='asis'".}
}
\description{
Helper function that splits up a single written R chunk (in R
Markdown) into multiple subchunks. This can be used to enable plotting
multiple plots of different sizes and captions from within a single written
R code chunk in R Markdown.
}
\references{
Adapted from \href{http://michaeljw.com/blog/post/subchunkify/}{Michael James Williams}.
}
\keyword{internal}
