% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-lib-prediction-metrics.R
\name{summarize_auc_curve}
\alias{summarize_auc_curve}
\title{Summarize ROC/PR curves.}
\usage{
summarize_auc_curve(
  fit_results,
  vary_params = NULL,
  y = "y",
  yhat = "yhat",
  metric = c("ROC", "PR"),
  x_grid = seq(0, 1, by = 0.01),
  summary_funs = c("mean", "median", "min", "max", "sd", "raw"),
  custom_summary_funs = NULL,
  na.rm = F
)
}
\arguments{
\item{fit_results}{A tibble, as returned by the \code{Experiment$fit()}
method.}

\item{vary_params}{A vector of parameter names that are varied over in the
Experiment.}

\item{y}{Character string. Name of column in \code{fit_results} with the
observed response values.}

\item{yhat}{Character string. Name of column in \code{fit_results} with the
predicted response values.}

\item{metric}{A character string. Either "ROC" or "PR".}

\item{summary_funs}{Character vector specifying how to summarize
evaluation metrics. Elements of the vector must be one of "mean", "median",
"min", "max", "sd", "raw".}

\item{custom_summary_funs}{Named list of custom functions to summarize
results. Names in the list should correspond to the name of the summary
function. Values in the list should be a function that takes in one
argument, that being the values of the evaluated metrics. See examples.}

\item{na.rm}{Logical. Should missing values be removed in evaluating summary
functions?}

\item{grid}{Sequence of x-values to evaluate ROC or PR curve.}
}
\value{
A data frame with the following columns in addition to any arguments
in the \code{Experiment}'s \code{vary_params}:
\describe{
\item{dgp_name}{Name of DGP.}
\item{method_name}{Name of Method.}
\item{...}{Other columns corresponding to the results from the provided
\code{summary_funs} and \code{custom_summary_funs} functions.}
}
}
\description{
Summarize ROC/PR curves across experimental repetitions.
}
