% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-lib.R
\name{summarize_eval_results}
\alias{summarize_eval_results}
\title{Summarize evaluation results.}
\usage{
summarize_eval_results(
  eval_results,
  id,
  value = "value",
  summary_funs = c("mean", "median", "min", "max", "sd", "raw"),
  custom_summary_funs = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{eval_results}{A list of result tibbles, as returned by the
\code{Experiment$evaluate()} method.}

\item{id}{Character string. ID used for naming columns.}

\item{value}{Character string. Name of column in \code{eval_results} to
summarize.}

\item{summary_funs}{Character vector specifying how to summarize
evaluation metrics. Elements of the vector must be one of "mean", "median",
"min", "max", "sd", "raw".}

\item{custom_summary_funs}{Named list of custom functions to summarize
results. Names in the list should correspond to the name of the summary
function. Values in the list should be a function that takes in one
argument, that being the values of the evaluated metrics. See examples.}

\item{na.rm}{Logical. Should missing values be removed in evaluating summary
functions?}
}
\value{
A ggplot object or list of ggplot objects.
}
\description{
General helper function to summarize evaluation results in
a nice, organized table.
}
