% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-lib-feature-selection.R
\name{summarize_feature_importances}
\alias{summarize_feature_importances}
\title{Summarize feature importance results.}
\usage{
summarize_feature_importances(
  fit_results,
  vary_params = NULL,
  fi = "fi",
  feature_col = "feature",
  imp_col = "importance",
  summary_funs = c("mean", "median", "min", "max", "sd", "raw"),
  custom_summary_funs = NULL,
  na.rm = F
)
}
\arguments{
\item{fit_results}{A tibble, as returned by the \code{Experiment$fit()}
method.}

\item{vary_params}{A vector of parameter names that are varied over in the
Experiment.}

\item{fi}{Character string or vector. Name of column(s) in \code{fit_results}
with the feature importance score information.}

\item{feature_col}{Character string. Name of the feature column. Could be the
name of a column in \code{fi} or \code{fit_results}.}

\item{imp_col}{Character string. Name of the importance column. Could be the
name of a column in \code{fi} or \code{fit_results}.}

\item{summary_funs}{Character vector specifying how to summarize
evaluation metrics. Elements of the vector must be one of "mean", "median",
"min", "max", "sd", "raw".}

\item{custom_summary_funs}{Named list of custom functions to summarize
results. Names in the list should correspond to the name of the summary
function. Values in the list should be a function that takes in one
argument, that being the values of the evaluated metrics. See examples.}

\item{na.rm}{Logical. Should missing values be removed in evaluating summary
functions?}
}
\value{
A data frame with the following columns in addition to any arguments
in the \code{Experiment}'s \code{vary_params}:
\describe{
\item{dgp_name}{Name of DGP.}
\item{method_name}{Name of Method.}
\item{metric}{Name of prediction error metric.}
\item{...}{Other columns corresponding to the results from the provided
\code{summary_funs} and \code{custom_summary_funs} functions.}
}
}
\description{
Summarize feature importance results across experimental
repetitions.
}
