% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-lib-feature-selection.R
\name{summarize_feature_recovery}
\alias{summarize_feature_recovery}
\title{Summarize prediction error evaluation results.}
\usage{
summarize_feature_recovery(
  fit_results,
  vary_params = NULL,
  s = "s",
  shat = "shat",
  metrics = c("TP", "FP", "TPR", "FPR", "FDR", "P", "N", "AUROC", "AUPRC"),
  custom_metrics = NULL,
  summary_funs = c("mean", "median", "min", "max", "sd", "raw"),
  custom_summary_funs = NULL,
  na.rm = F
)
}
\arguments{
\item{fit_results}{A tibble, as returned by the \code{Experiment$fit()}
method.}

\item{vary_params}{A vector of parameter names that are varied over in the
Experiment.}

\item{s}{Character string. Name of column in \code{fit_results} with the
true feature support sets.}

\item{shat}{Character string. Name of column in \code{fit_results} with the
predicted feature support sets.}

\item{metrics}{Character vector of the support recovery metrics to compute.
Elements of the vector must be one of "TP", "FP", "TPR", "FPR", "FDR",
"P", "N", "AUROC", "AUPRC"}

\item{custom_metrics}{Named list of custom metric functions to compute. Names
in the list should correspond to the name of the metric. Values in the
list should be a function that takes in the arguments \code{s} and
\code{shat} only and returns the evaluated metric value.}

\item{summary_funs}{Character vector specifying how to summarize
evaluation metrics. Elements of the vector must be one of "mean", "median",
"min", "max", "sd", "raw".}

\item{custom_summary_funs}{Named list of custom functions to summarize
results. Names in the list should correspond to the name of the summary
function. Values in the list should be a function that takes in one
argument, that being the values of the evaluated metrics. See examples.}

\item{na.rm}{Logical. Should missing values be removed in evaluating summary
functions?}
}
\value{
A data frame with the following columns in addition to any arguments
in the \code{Experiment}'s \code{vary_params}:
\describe{
\item{dgp_name}{Name of DGP.}
\item{method_name}{Name of Method.}
\item{metric}{Name of prediction error metric.}
\item{...}{Other columns corresponding to the results from the provided
\code{summary_funs} and \code{custom_summary_funs} functions.}
}
}
\description{
Summarize prediction error evaluation results for a variety of
evaluation metrics across experimental repetitions.
}
