% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-lib-feature-selection.R
\name{summarize_feature_recovery_curve}
\alias{summarize_feature_recovery_curve}
\title{Summarize the ROC/PR curves for feature support recovery.}
\usage{
summarize_feature_recovery_curve(
  fit_results,
  vary_params = NULL,
  data_col,
  truth,
  estimate,
  metric = c("ROC", "PR"),
  options = list(),
  na_rm = FALSE,
  x_grid = seq(0, 1, by = 0.01),
  summary_funs = c("mean", "median", "min", "max", "sd", "raw"),
  custom_summary_funs = NULL
)
}
\arguments{
\item{fit_results}{A tibble, as returned by the \code{fit} method.}

\item{vary_params}{A vector of parameter names that are varied across in the
\code{Experiment}.}

\item{data_col}{Character string specifying the name of the list-type column
in \code{fit_results} with the feature support \code{data.frames} or
\code{tibbles} that contain the \code{truth} and \code{estimate} columns.
Each row in these \code{data.frames} or \code{tibbles} typically
corresponds to a feature or variable.}

\item{truth}{A character string identifying the column with the true
support. The column should be binary, where \code{TRUE} or \code{1} means
the feature (corresponding to that row) is in the support and \code{FALSE}
or \code{0} means the feature is not in the support.}

\item{estimate}{A character string identifying the column with the estimated
support, where a non-zero entry means that the feature (corresponding to
that row) is part of the estimated support and a zero entry means that
feature is not part of the estimated support. Moreover, a higher magnitude
indicates a more important feature.}

\item{metric}{Either "ROC" or "PR" indicating whether to evaluate the ROC or
Precision-Recall curve.}

\item{options}{A \code{list} of named options to pass to \code{\link[pROC:roc]{pROC::roc()}}
such as \code{smooth}. These options should not include \code{response},
\code{predictor}, \code{levels}, \code{quiet}, or \code{direction}.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA} values
should be stripped before the computation proceeds.}

\item{x_grid}{Vector of values between 0 and 1 at which to evaluate the ROC
or PR curve. If \code{metric = "ROC"}, the provided vector of values are
the FPR values at which to evaluate the TPR, and if \code{metric = "PR"},
the values are the recall values at which to evaluate the precision.}

\item{summary_funs}{Character vector specifying how to summarize
evaluation metrics. Must choose from a built-in library of summary
functions - elements of the vector must be one of "mean", "median",
"min", "max", "sd", "raw".}

\item{custom_summary_funs}{Named list of custom functions to summarize
results. Names in the list should correspond to the name of the summary
function. Values in the list should be a function that takes in one
argument, that being the values of the evaluated metrics.}
}
\value{
A grouped \code{tibble} containing both identifying information
and the feature recovery curve results aggregated over experimental
replicates. Specifically, the identifier columns include \code{dgp_name},
\code{method_name}, and any columns specified by \code{vary_params}. In
addition, there are results columns corresponding to the
requested statistics in \code{summary_funs} and \code{custom_summary_funs}.
If \code{metric = "ROC"}, these results columns include \code{FPR} and
others that end in the suffix "_TPR". If \code{metric = "PR"}, the results
columns include \code{recall} and others that end in the suffix
"_precision".
}
\description{
Summarize the ROC or PR curves for feature support recovery
across experimental repetitions.
}
\seealso{
Other feature_selection_funs: 
\code{\link{eval_feature_importances}()},
\code{\link{eval_feature_recovery_curve}()},
\code{\link{eval_feature_recovery}()},
\code{\link{plot_feature_imp}()},
\code{\link{plot_feature_recovery_curve}()},
\code{\link{plot_feature_recovery}()},
\code{\link{summarize_feature_importances}()},
\code{\link{summarize_feature_recovery}()}
}
\concept{feature_selection_funs}
