% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-lib-prediction-metrics.R
\name{summarize_pred_err}
\alias{summarize_pred_err}
\title{Summarize prediction error evaluation results.}
\usage{
summarize_pred_err(
  fit_results,
  vary_params = NULL,
  y = "y",
  yhat = "yhat",
  metrics = NULL,
  custom_metrics = NULL,
  summary_funs = c("mean", "median", "min", "max", "sd", "raw"),
  custom_summary_funs = NULL,
  group = NULL,
  na.rm = F
)
}
\arguments{
\item{fit_results}{A tibble, as returned by the \code{Experiment$fit()}
method.}

\item{vary_params}{A vector of parameter names that are varied over in the
Experiment.}

\item{y}{Character string. Name of column in \code{fit_results} with the
observed response values.}

\item{yhat}{Character string. Name of column in \code{fit_results} with the
predicted response values.}

\item{metrics}{Character vector of the prediction error metrics to compute.
Elements of the vector must be one of "RMSE", "MSE", "R2", "MAE",
"Corr", "ClassErr", "BalancedClassErr", "AUROC", "AUPRC".}

\item{custom_metrics}{Named list of custom metric functions to compute. Names
in the list should correspond to the name of the metric. Values in the
list should be a function that takes in the arguments \code{y} and
\code{yhat} only and returns the evaluated metric value.}

\item{summary_funs}{Character vector specifying how to summarize
evaluation metrics. Elements of the vector must be one of "mean", "median",
"min", "max", "sd", "raw".}

\item{custom_summary_funs}{Named list of custom functions to summarize
results. Names in the list should correspond to the name of the summary
function. Values in the list should be a function that takes in one
argument, that being the values of the evaluated metrics. See examples.}

\item{group}{(Optional) vector of group ids (as a factor) to use for
assessing within-group prediction errors.}

\item{na.rm}{Logical. Should missing values be removed?}
}
\value{
A data frame with the following columns in addition to any arguments
in the \code{Experiment}'s \code{vary_params}:
\describe{
\item{dgp_name}{Name of DGP.}
\item{method_name}{Name of Method.}
\item{group}{Name of group for within-group prediction errors. Note that the
group id "all" is used to denote the full data set without any grouping.
Column is only present if \code{group} is not \code{NULL}.}
\item{metric}{Name of prediction error metric.}
\item{...}{Other columns corresponding to the results from the provided
\code{summary_funs} and \code{custom_summary_funs} functions.}
}
}
\description{
Summarize prediction error evaluation results for a variety of
evaluation metrics across experimental repetitions.
}
