% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{update_funs}
\alias{update_funs}
\alias{update_dgp}
\alias{update_method}
\alias{update_evaluator}
\alias{update_plot}
\alias{update_vary_across}
\title{Helper functions for updating components of an \code{Experiment}.}
\usage{
update_dgp(experiment, dgp, name, ...)

update_method(experiment, method, name, ...)

update_evaluator(
  experiment,
  evaluator,
  name = NULL,
  dgp_name = NULL,
  method_name = NULL,
  ...
)

update_plot(experiment, plotter, name, ...)

update_vary_across(
  experiment,
  dgp = NULL,
  method = NULL,
  param_name,
  param_values
)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{dgp}{A \code{DGP} object.}

\item{name}{A name to identify the object to be updated.}

\item{...}{Not currently used.}

\item{method}{A \code{Method} object.}

\item{evaluator}{A \code{Evaluator} object.}

\item{dgp_name}{The name of a \code{DGP} object in \code{experiment} with which to
associate \code{evaluator}.}

\item{method_name}{The name of a \code{Method} object in \code{experiment} with
which to associate \code{evaluator}.}

\item{plotter}{A \code{Plotter} object.}
}
\value{
The original \code{experiment} object passed to \code{update_*}.
}
\description{
Helper functions for updating \code{DGPs}, \code{Methods},
\code{Evaluators}, and \code{Plotters} already added to an
\code{Experiment}.
}
