% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_templates.R
\name{use_dgp_template}
\alias{use_dgp_template}
\title{Function to create boilerplate code for creating DGPs.}
\usage{
use_dgp_template(ids = NULL)
}
\description{
Function to create boilerplate code for creating DGPs.
}
\keyword{internal}
