% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_templates.R
\name{use_evaluator_template}
\alias{use_evaluator_template}
\title{Function to create boilerplate code for creating Evaluators.}
\usage{
use_evaluator_template(
  ids = NULL,
  pred_nested_cols,
  pred_truth_col,
  pred_estimate_col,
  pred_prob_cols,
  feature_nested_cols,
  feature_col,
  feature_truth_col,
  feature_imp_col,
  feature_sel_col,
  feature_pval_col
)
}
\description{
Function to create boilerplate code for creating Evaluators.
}
\keyword{internal}
