% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_templates.R
\name{use_experiment_template}
\alias{use_experiment_template}
\title{Function to create boilerplate code for creating Experiments.}
\usage{
use_experiment_template(
  name = "Experiment",
  dgp_names = NULL,
  method_names = NULL,
  eval_names = NULL,
  viz_names = NULL
)
}
\description{
Function to create boilerplate code for creating Experiments.
}
\keyword{internal}
