% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-dir.R
\name{use_git_no_ask}
\alias{use_git_no_ask}
\title{Initialise a git repository}
\usage{
use_git_no_ask(message = "initial commit")
}
\arguments{
\item{message}{A \code{character} message to use for the initial commit.}
}
\description{
\code{use_git_no_ask()} is based on
\code{\link[usethis::use_git]{usethis::use_git}()}: it generates a git
repository. This function differs from \code{\link[usethis]{use_git}()},
however, in that users are not prompted to add untracked files to the
initial commit.
}
\keyword{internal}
