% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_templates.R
\name{use_init_docs_template}
\alias{use_init_docs_template}
\title{Function to create boilerplate code for creating the documentation template.}
\usage{
use_init_docs_template()
}
\description{
Function to create boilerplate code for creating the documentation template.
}
\keyword{internal}
