% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_templates.R
\name{use_library_template}
\alias{use_library_template}
\title{Function to create boilerplate code for loading in packages}
\usage{
use_library_template(cran_packages = NULL, github_packages = NULL)
}
\description{
Function to create boilerplate code for loading in packages
}
\keyword{internal}
