% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_templates.R
\name{use_method_template}
\alias{use_method_template}
\title{Function to create boilerplate code for creating Methods.}
\usage{
use_method_template(ids = NULL)
}
\description{
Function to create boilerplate code for creating Methods.
}
\keyword{internal}
