% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_templates.R
\name{use_variable_template}
\alias{use_variable_template}
\title{Function to create boilerplate code for assigning descriptive errors to
variables.}
\usage{
use_variable_template(id)
}
\description{
Function to create boilerplate code for assigning descriptive errors to
variables.
}
\keyword{internal}
