% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_templates.R
\name{use_visualizer_template}
\alias{use_visualizer_template}
\title{Function to create boilerplate code for creating Visualizers}
\usage{
use_visualizer_template(
  ids = NULL,
  pred_nested_data,
  pred_truth_col,
  pred_prob_cols,
  feature_nested_data,
  feature_col,
  feature_truth_col,
  feature_imp_col,
  feature_pval_col
)
}
\description{
Function to create boilerplate code for creating Visualizers
}
\keyword{internal}
