% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{visualize_experiment}
\alias{visualize_experiment}
\title{Visualize results of an \code{Experiment}.}
\usage{
visualize_experiment(experiment, ...)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{...}{Passed to \code{Experiment$visualize()}.}
}
\value{
A list of visualizations, one for each \code{Visualizer}.
}
\description{
Visualize the performance of methods and/or its evaluation metrics
using all \code{Visualizers} in the \code{Experiment} and return visualization results.
}
