% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signals.R
\name{warn}
\alias{warn}
\title{Signals a warning with default subclass "simChef_warning".}
\usage{
warn(
  message = NULL,
  class = "simChef_warning",
  call = rlang::caller_env(),
  ...
)
}
\arguments{
\item{message}{Message to include with the condition.}

\item{class}{Subclass passed to the appropriate \code{rlang} signal function.}

\item{call}{The function environment that is most relevant from the user's
perspective. Default is \code{rlang::caller_env()}, which gives the environment
attached to the function that called this signal function.}

\item{...}{Additional arguments to pass to the appropriate \code{rlang} signal
function.}
}
\description{
Signals a warning with default subclass "simChef_warning".
}
\keyword{internal}
