% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{add_funs}
\alias{add_funs}
\alias{add_dgp}
\alias{add_method}
\alias{add_evaluator}
\alias{add_plot}
\alias{add_vary_across}
\title{Helper functions for adding components to an \code{Experiment}.}
\usage{
add_dgp(experiment, dgp, name = NULL, ...)

add_method(experiment, method, name = NULL, ...)

add_evaluator(experiment, evaluator, name = NULL, ...)

add_plot(experiment, plotter, name = NULL, ...)

add_vary_across(
  experiment,
  dgp = NULL,
  method = NULL,
  param_name,
  param_values
)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{dgp}{A \code{DGP} object.}

\item{name}{A name to identify the object to be added.}

\item{...}{Not currently used.}

\item{method}{A \code{Method} object.}

\item{evaluator}{A \code{Evaluator} object.}

\item{plotter}{A \code{Plotter} object.}

\item{dgp_name}{The name of a \code{DGP} object in \code{experiment} with which to
associate \code{evaluator}.}

\item{method_name}{The name of a \code{Method} object in \code{experiment} with
which to associate \code{evaluator}.}
}
\value{
The original \code{experiment} object passed to \code{add_*}.
}
\description{
Helper functions for adding \code{DGPs}, \code{Methods},
\code{Evaluators}, and \code{Plotters} to an \code{Experiment}.
}
