% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-lib-utils.R
\name{add_na_counts}
\alias{add_na_counts}
\title{Developer function to add number of NAs to evaluator results}
\usage{
add_na_counts(out, data, value_col, na_rm, ...)
}
\arguments{
\item{out}{Evaluator results tibble to append number of NA results to.}

\item{data}{Data used to compute number of NAs.}

\item{value_col}{Character string, specifying the column used to compute
the number of NAs.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA} values
should be stripped before the computation proceeds.}

\item{...}{Additional name-value pairs to pass to dplyr::mutate() to append
columns.}
}
\value{
Tibble with additional rows containing the new metric "num_na" and
its corresponding ".estimate"
}
\description{
A helper function to append rows with number of NAs (per group,
if applicable) to evaluator results tibble.
}
