% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-lib-errors.R
\name{ar1_errors}
\alias{ar1_errors}
\title{Generate autoregressive Gaussian errors.}
\usage{
ar1_errors(n, rho)
}
\arguments{
\item{n}{Number of samples.}

\item{rho}{Correlation.}
}
\value{
A vector of simulated errors with length \code{n}.
}
\description{
Generate correlated Gaussian errors based on an
autoregressive(1) covariance structure.
}
\examples{
errs <- ar1_errors(n = 100, rho = 0.7)

}
