% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-lib-errors.R
\name{block_errors}
\alias{block_errors}
\title{Generate block-correlated Gaussian errors.}
\usage{
block_errors(n, n_blocks = 3, rho = 0.8)
}
\arguments{
\item{n}{Number of samples.}

\item{n_blocks}{Number of blocks.}

\item{rho}{Correlation. Must be a scalar or vector of length \code{k}.}
}
\value{
A vector of simulated errors with length \code{n}.
}
\description{
Generate correlated Gaussian errors based on a block-dependence
covariance structure with \code{n_blocks} (approximately) equally-sized
blocks.
}
\examples{
errs <- block_errors(n = 100, n_blocks = 3, rho = 0.7)

}
