% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_equal}
\alias{check_equal}
\title{Check equality of \code{Experiment} components.}
\usage{
check_equal(obj1, obj2)
}
\arguments{
\item{obj1}{An object of class \code{DGP}, \code{Method}, \code{Evaluator},
or \code{Plotter}}

\item{obj2}{An object of class \code{DGP}, \code{Method}, \code{Evaluator},
or \code{Plotter}}
}
\value{
Logical. Returns \code{TRUE} if both objects have the same function
and arguments and \code{FALSE} otherwise.
}
\description{
Check if any two \code{DGPs}, \code{Methods}, \code{Evaluators},
or \code{Plotters} are the same with respect to the function and inputted
arguments.
}
\keyword{internal}
