% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{create_dgp}
\alias{create_dgp}
\title{Create a new \code{DGP} (data-generating process)}
\usage{
create_dgp(.dgp_fun, .name = NULL, ...)
}
\arguments{
\item{.dgp_fun}{The user-defined data-generating process function.}

\item{.name}{(Optional) An optional name for the \code{DGP}, helpful for later
identification.}

\item{...}{User-defined default arguments to pass to \code{.dgp_fun()} when
\code{DGP$generate()} is called.}
}
\value{
A new \link{DGP} object.
}
\description{
Create a \link{DGP} which can \code{generate()} data in an \link{Experiment}.
}
\examples{
# create an example DGP function
dgp_fun <- function(n, beta, rho, sigma) {
  cov_mat <- matrix(c(1, rho, rho, 1), byrow = T, nrow = 2, ncol = 2)
  X <- MASS::mvrnorm(n = n, mu = rep(0, 2), Sigma = cov_mat)
  y <- X \%*\% beta + rnorm(n, sd = sigma)
  return(list(X = X, y = y))
}

# create DGP (with uncorrelated features)
dgp <- create_dgp(.dgp_fun = dgp_fun,
                  .name = "Linear Gaussian DGP",
                  # additional named parameters to pass to dgp_fun() by default
                  n = 50, beta = c(1, 0), rho = 0, sigma = 1)

print(dgp)

data_uncorr <- dgp$generate()
cor(data_uncorr$X)

data_corr <- dgp$generate(rho = 0.7)
cor(data_corr$X)

}
