% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{create_dgp}
\alias{create_dgp}
\title{Create a new \code{DGP} (data-generating process).}
\usage{
create_dgp(dgp_fun, name = NULL, ...)
}
\arguments{
\item{dgp_fun}{The data-generating process function.}

\item{name}{(Optional) The name of the \code{DGP}. The argument must be
specified by position or typed out in whole; no partial matching is allowed
for this argument.}

\item{...}{Arguments to pass into \code{dgp_fun()}.}
}
\value{
A new instance of \code{DGP}.
}
\description{
Create a new \code{DGP} (data-generating process).
}
