% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{create_doc_template}
\alias{create_doc_template}
\title{Create documentation template for the R Markdown results report.}
\usage{
create_doc_template(experiment, experiment_dirname)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{experiment_dirname}{A directory where results from an \code{Experiment}
were previously saved. Used if \code{experiment} was not provided.}
}
\value{
The original \code{Experiment} object passed to
\code{create_doc_template}.
}
\description{
Create documentation template (a series of .md files) to
fill out for the R Markdown results report. The documentation files can
be found in the \code{Experiment}'s results directory (see
\code{Experiment$get_save_dir()}) under docs/.
}
