% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_helpers.R
\name{create_doc_template}
\alias{create_doc_template}
\title{Create documentation template for the R Markdown results report.}
\usage{
create_doc_template(experiment, save_dir, ...)
}
\arguments{
\item{experiment}{An \code{Experiment} object. If provided,
\code{experiment$get_save_dir()} is used to find previously saved results.
If not provided, then \code{save_dir} is used instead.}

\item{save_dir}{An optional directory in which to find saved results. Not
used if \code{experiment} provided.}

\item{...}{Not used.}
}
\value{
The original \code{Experiment} object passed to
\code{create_doc_template}.
}
\description{
Create documentation template (a series of .md files) to
fill out for the R Markdown results report. The documentation files can
be found in the \code{Experiment}'s results directory (see
\code{Experiment$get_save_dir()}) under docs/.
}
