% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator.R
\name{create_evaluator}
\alias{create_evaluator}
\title{Create a new \code{Evaluator}.}
\usage{
create_evaluator(eval_fun, ...)
}
\arguments{
\item{eval_fun}{The evaluation function.}

\item{...}{Arguments to pass into \code{eval_fun()}.}
}
\value{
A new instance of \code{Evaluator}.
}
\description{
Create a new \code{Evaluator}.
}
