% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-helpers.R
\name{create_experiment}
\alias{create_experiment}
\title{Create a new \code{Experiment}.}
\usage{
create_experiment(
  name = "experiment",
  dgp_list = list(),
  method_list = list(),
  evaluator_list = list(),
  visualizer_list = list(),
  future.globals = TRUE,
  future.packages = NULL,
  clone_from = NULL,
  save_dir = NULL,
  ...
)
}
\arguments{
\item{name}{The name of the \code{Experiment}.}

\item{dgp_list}{An optional list of \code{DGP} objects.}

\item{method_list}{An optional list of \code{Method} objects.}

\item{evaluator_list}{An optional list of \code{Evaluator} objects.}

\item{visualizer_list}{An optional list of \code{Visualizer} objects.}

\item{future.globals}{Character vector of names in the global environment to
pass to parallel workers. Passed as the argument of the same name to
\code{future.apply::future_lapply} and related functions. To set for a
specific run of the experiment, use the same argument in
\code{Experiment$run}.}

\item{future.packages}{Character vector of packages required by parallel
workers. Passed as the argument of the same name to
\code{future.apply::future_lapply} and related functions. To set for a
specific run of the experiment, use the same argument in
\code{Experiment$run}.}

\item{clone_from}{An optional \code{Experiment} object to use as a base for
this one.}

\item{save_dir}{An optional directory in which to save the experiment's
results. If \code{NULL}, results are saved in the current working directory
in a directory called "results" with a sub-directory named after
\code{Experiment$name}.}

\item{...}{Not used.}
}
\value{
A new \code{Experiment} object.
}
\description{
Create a new \code{Experiment}.
}
