% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{create_method}
\alias{create_method}
\title{Create a new \code{Method}.}
\usage{
create_method(method_fun, name = NULL, ...)
}
\arguments{
\item{method_fun}{The method function.}

\item{name}{(Optional) The name of the \code{Method}.}

\item{...}{Arguments to pass into \code{method_fun()}.}
}
\value{
A new instance of \code{Method}.
}
\description{
Create a new \code{Method}.
}
