% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{create_rmd}
\alias{create_rmd}
\title{Create an R Markdown file summarizing the results of an \code{Experiment}.}
\usage{
create_rmd(experiment, experiment_dirname, open = TRUE)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{experiment_dirname}{A directory where results from an \code{Experiment}
were previously saved. Used if \code{experiment} was not provided.}

\item{open}{If \code{TRUE}, open the R Markdown file in a web browser.}
}
\value{
The original \code{experiment} object passed to \code{create_rmd}.
}
\description{
Create an R Markdown file summarizing the results of an \code{Experiment}.
}
