% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{create_rmd}
\alias{create_rmd}
\title{Render an R Markdown file summarizing the results of an \code{Experiment} or
set of \code{Experiments}.}
\usage{
create_rmd(
  experiment,
  save_dir,
  open = TRUE,
  title = NULL,
  author = "",
  verbose = 2,
  quiet = TRUE,
  pretty = TRUE,
  eval_order = NULL,
  viz_order = NULL,
  ...
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{create_rmd()} was renamed to \code{render_docs()} to create a more consistent
API.
}
\keyword{internal}
