% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_helpers.R
\name{create_rmd}
\alias{create_rmd}
\title{Create an R Markdown file summarizing the results of an \code{Experiment}.}
\usage{
create_rmd(
  experiment,
  experiment_dirname,
  open = TRUE,
  author = "",
  verbose = 2,
  quiet = TRUE,
  pretty = TRUE,
  ...
)
}
\arguments{
\item{experiment}{An \code{Experiment} object. If provided,
\code{experiment$get_save_dir()} is used to find previously saved results.
If not provided, then \code{save_dir} is used instead.}

\item{save_dir}{An optional directory in which to find saved results. Not
used if \code{experiment} provided.}

\item{open}{If \code{TRUE}, open the R Markdown-generated html file in a web
browser.}

\item{author}{Character string of author names to display in knitted R
Markdown document.}

\item{verbose}{Level of verboseness (0, 1, 2) when knitting R Markdown.
Default is 2.}

\item{quiet}{Default is \code{TRUE}. See \code{\link[rmarkdown:render]{rmarkdown::render()}} for
details.}

\item{pretty}{Logical. Specifies whether or not to use pretty R Markdown
results template or more barebones R Markdown results template. Default
\code{TRUE} uses the pretty template. Set to \code{FALSE} to start from
the barebones template, which can be helpful when using your own custom
R Markdown theme.}

\item{...}{Additional arguments to pass to \code{\link[rmarkdown:render]{rmarkdown::render()}}. Useful
for applying a custom R Markdown output theme.}
}
\value{
The original \code{Experiment} object passed to \code{create_rmd}.
}
\description{
Knits an R Markdown file summarizing the results of an
\code{Experiment}. Outputs an R Markdown-generated html file, saved in
the \code{Experiment}'s root results directory (see
\code{Experiment$get_save_dir()}).
}
