% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rmd.R
\name{create_rmd}
\alias{create_rmd}
\title{Create an R Markdown file summarizing the results of an \code{Experiment} or
set of \code{Experiments}.}
\usage{
create_rmd(
  experiment,
  save_dir,
  open = TRUE,
  title = NULL,
  author = "",
  verbose = 2,
  quiet = TRUE,
  pretty = TRUE,
  eval_order = NULL,
  viz_order = NULL,
  ...
)
}
\arguments{
\item{experiment}{An \code{Experiment} object. If provided,
\code{experiment$get_save_dir()} is used to find previously saved results.
If not provided, then \code{save_dir} is used instead.}

\item{save_dir}{An optional directory in which to find saved results. Not
used if \code{experiment} provided.}

\item{open}{If \code{TRUE}, open the R Markdown-generated html file in a
web browser.}

\item{title}{Character string. Title of the report. By default, this will be
the name of the \code{experiment} if \code{experiment} is provided.}

\item{author}{Character string of author names to display in knitted R
Markdown document.}

\item{verbose}{Level of verboseness (0, 1, 2) when knitting R Markdown.
Default is 2.}

\item{quiet}{Default is \code{TRUE}. See \code{\link[rmarkdown:render]{rmarkdown::render()}} for
details.}

\item{pretty}{Logical. Specifies whether or not to use pretty R Markdown
results template or more barebones R Markdown results template. Default
\code{TRUE} uses the pretty template. Set to \code{FALSE} to start from
the barebones template, which can be helpful when using your own custom
R Markdown theme.}

\item{eval_order}{Vector of \code{Evaluator} names in their desired order for
display. By default, the report will display the \code{Evaluator} results
in the order that they were computed.}

\item{viz_order}{Vector of \code{Visualizer} names in their desired order for
display. By default, the report will display the \code{Visualizer} results
in the order that they were computed.}

\item{...}{Additional arguments to pass to \code{\link[rmarkdown:render]{rmarkdown::render()}}. Useful
for applying a custom R Markdown output theme.}
}
\value{
The original \code{Experiment} object if provided. Otherwise,
returns \code{NULL}.
}
\description{
Knits an R Markdown file summarizing the results of an
\code{Experiment} or set of \code{Experiments}. Outputs an R
Markdown-generated html file. If \code{experiment} is provided, the results
are saved in the \code{Experiment}'s root results directory (see
\code{Experiment$get_save_dir()}). Otherwise, the root results directory is
taken to be that specified by \code{save_dir}. Note that
\code{create_rmd()} will process and include results from all
\code{Experiments} found \emph{under} the root directory.
}
