% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer.R
\name{create_visualizer}
\alias{create_visualizer}
\title{Create a new \code{Visualizer}.}
\usage{
create_visualizer(visualizer_fun, name = NULL, rmd_options = list(), ...)
}
\arguments{
\item{visualizer_fun}{The visualization function.}

\item{name}{(Optional) The name of the \code{Visualizer}.}

\item{rmd_options}{(Optional) List of options to control the aesthetics of
the \code{Visualizer}'s visualization in the knitted R Markdown report.
Currently, possible options are "height" and "width" (in inches).}

\item{...}{Arguments to pass into \code{visualizer_fun()}.}
}
\value{
A new instance of \code{Visualizer}.
}
\description{
Create a new \code{Visualizer}.
}
