% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\docType{class}
\name{DGP}
\alias{DGP}
\title{\code{R6} class representing a data-generating process.}
\description{
A data-generating process which will be used in the
\code{Experiment} to \strong{generate} data.
}
\section{Public fields}{


\describe{
\item{\code{name}}{The name of the \code{DGP}.}
\item{\code{dgp_fun}}{The data-generating process function.}
\item{\code{dgp_params}}{(Named) list of parameters to input into the data-generating process function.}
}
}

\section{Public methods}{


\itemize{
\item{\code{DGP$new()}} See \link{create_dgp}.
}
}

