% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{do_call_handler}
\alias{do_call_handler}
\title{Call a function with given parameters and if an error, warning, or message
occurs during the call print information about the call to the console.}
\usage{
do_call_handler(name, fun, params = list(), verbose = 1)
}
\arguments{
\item{name}{a name to give some context for the call}

\item{fun}{a function to call}

\item{params}{arguments to pass to fun}
}
\value{
the results of calling fun with params
}
\description{
Call a function with given parameters and if an error, warning, or message
occurs during the call print information about the call to the console.
}
\keyword{internal}
