% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-lib-feature-selection.R
\name{eval_feature_importances}
\alias{eval_feature_importances}
\title{Show feature importance scores.}
\usage{
eval_feature_importances(fit_results, vary_params = NULL, fi = "imp")
}
\arguments{
\item{fit_results}{A tibble, as returned by the \code{Experiment$fit()}
method.}

\item{vary_params}{A vector of parameter names that are varied over in the
Experiment.}

\item{fi}{Character string or vector. Name of column(s) in \code{fit_results}
with the feature importance score information.}
}
\value{
TODO
}
\description{
Show feature importance scores.
}
