% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-lib-prediction-metrics.R
\name{eval_pred_curve_funs}
\alias{eval_pred_curve_funs}
\alias{eval_pred_curve}
\alias{summarize_pred_curve}
\title{Evaluate and/or summarize ROC or PR curves.}
\usage{
eval_pred_curve(
  fit_results,
  vary_params = NULL,
  nested_data = NULL,
  truth_col,
  prob_cols,
  curve = c("ROC", "PR"),
  groups = NULL,
  options = list(),
  na_rm = FALSE
)

summarize_pred_curve(
  fit_results,
  vary_params = NULL,
  nested_data = NULL,
  truth_col,
  prob_cols,
  curve = c("ROC", "PR"),
  groups = NULL,
  options = list(),
  na_rm = FALSE,
  x_grid = seq(0, 1, by = 0.01),
  summary_funs = c("mean", "median", "min", "max", "sd", "raw"),
  custom_summary_funs = NULL
)
}
\arguments{
\item{fit_results}{A tibble, as returned by the \code{fit} method.}

\item{vary_params}{A vector of parameter names that are varied across in the
\code{Experiment}.}

\item{nested_data}{(Optional) Character string. If specified, should be the
name of the column in \code{fit_results} containing columns that must be
unnested before evaluating results. Default is \code{NULL}, meaning no
columns in \code{fit_results} need to be unnested prior to computation.}

\item{truth_col}{A character string identifying the column with the true
responses. The column should be numeric for a regression problem and a
factor for a classification problem.}

\item{prob_cols}{A character string or vector identifying the column(s)
containing class probabilities. If the \code{truth_col} column is binary,
only 1 column name should be provided. Otherwise, the length of the
\code{prob_cols} should be equal to the number of factor levels of
the \code{truth_col} column. This argument is not used when evaluating
numeric metrics.}

\item{curve}{Either "ROC" or "PR" indicating whether to evaluate the ROC or
Precision-Recall curve.}

\item{groups}{(Optional) vector of group IDs to group observations by before
evaluating prediction errors. This is useful for assessing within-group
prediction errors. Note: the (unstratified) prediction errors, aggregated
across the full data set, are computed in addition to these stratified
within-group errors.}

\item{options}{A list of named options to pass to \code{pROC::roc()} such as
\code{smooth}. These options should not include \code{response},
\code{predictor}, \code{levels}, \code{quiet}, or \code{direction}. This
argument is only used when computing the ROC and is ignored otherwise.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA} values
should be stripped before the computation proceeds.}

\item{x_grid}{Vector of values between 0 and 1 at which to evaluate the ROC
or PR curve. If \code{curve = "ROC"}, the provided vector of values are
the FPR values at which to evaluate the TPR, and if \code{curve = "PR"},
the values are the recall values at which to evaluate the precision.}

\item{summary_funs}{Character vector specifying how to summarize
evaluation metrics. Must choose from a built-in library of summary
functions - elements of the vector must be one of "mean", "median",
"min", "max", "sd", "raw".}

\item{custom_summary_funs}{Named list of custom functions to summarize
results. Names in the list should correspond to the name of the summary
function. Values in the list should be a function that takes in one
argument, that being the values of the evaluated metrics.}
}
\value{
The output of \code{eval_pred_curve()} is a \code{tibble} with the following
columns:
\describe{
\item{rep}{Replicate ID.}
\item{dgp_name}{Name of DGP.}
\item{method_name}{Name of Method.}
\item{.group}{If \code{groups} is not \code{NULL}, this column specifies the
name of the group under evaluation. Otherwise, this column is not returned.}
\item{curve_estimate}{A list of tibbles with x and y coordinate values for
the ROC/PR curve for the given experimental replicate. If
\code{curve = "ROC"}, the \code{tibble} has the columns \code{.threshold},
\code{FPR}, and \code{TPR} for the threshold, false positive rate, and true
positive rate, respectively. If \code{curve = "PR"}, the \code{tibble} has
the columns \code{.threshold}, \code{recall}, and \code{precision}.}
}
as well as any columns specified by \code{vary_params}.

The output of \code{summarize_pred_curve()} is a grouped \code{tibble}
containing both identifying information and the prediction curve results
aggregated over experimental replicates. Specifically, the identifier columns
include \code{dgp_name}, \code{method_name}, and any columns specified by
\code{vary_params}. In addition, there are results columns corresponding to
the requested statistics in \code{summary_funs} and
\code{custom_summary_funs}. If \code{curve = "ROC"}, these results columns
include \code{FPR} and others that end in the suffix "_TPR". If
\code{curve = "PR"}, the results columns include \code{recall} and others
that end in the suffix "_precision".
}
\description{
Evaluate the ROC or PR curves, given the true responses and the
predicted probabilities for each class. \code{eval_pred_curve()} evaluates
the ROC or PR curve for each experimental replicate separately.
\code{summarize_pred_curve()} summarizes the ROC or PR curve across
experimental replicates.
}
\seealso{
Other prediction_error_funs: 
\code{\link{eval_pred_err_funs}},
\code{\link{plot_pred_curve}()},
\code{\link{plot_pred_err}()}
}
\concept{prediction_error_funs}
