% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-lib-inference.R
\name{eval_reject_prob}
\alias{eval_reject_prob}
\title{Evaluate the rejection probability of a hypothesis test.}
\usage{
eval_reject_prob(
  fit_results,
  vary_params = NULL,
  nested_cols = NULL,
  feature_col = NULL,
  pval_col,
  group_cols = NULL,
  alphas = NULL,
  na_rm = FALSE
)
}
\arguments{
\item{fit_results}{A tibble, as returned by \code{\link[=fit_experiment]{fit_experiment()}}.}

\item{vary_params}{A vector of \code{DGP} or \code{Method} parameter names that are
varied across in the \code{Experiment}.}

\item{nested_cols}{(Optional) A character string or vector specifying the
name of the column(s) in \code{fit_results} that need to be
unnested before evaluating results. Default is \code{NULL}, meaning no
columns in \code{fit_results} need to be unnested prior to computation.}

\item{feature_col}{A character string identifying the column in
\code{fit_results} with the feature names or IDs.}

\item{pval_col}{A character string identifying the column in
\code{fit_results} with the estimated p-values data. Each element in
this column should be an array of length \code{p}, where \code{p} is the
number of features and the feature order aligns with that of
\code{truth_col}.}

\item{group_cols}{(Optional) A character string or vector specifying the
column(s) to group rows by before evaluating metrics.
This is useful for assessing within-group metrics.}

\item{alphas}{(Optional) Vector of significance levels at which to evaluate
the rejection probability. By default, \code{alphas} is \code{NULL}, which
evaluates the full empirical cumulative distribution of the p-values, i.e.,
the rejection probability is evaluated at all possible significance levels.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA} values
should be stripped before the computation proceeds.}
}
\value{
A grouped \code{tibble} containing both identifying information
and the rejection probability results aggregated over experimental
replicates. Specifically, the identifier columns include \code{.dgp_name},
\code{.method_name}, any columns specified by \code{group_cols} and
\code{vary_params}, and the feature names given in \code{feature_col} if
applicable. In addition, there are results columns \code{.alpha} and
\code{reject_prob}, which respectively give the significance level and the
estimated rejection probabilities (averaged across experimental
replicates).
}
\description{
Evaluate the probability of rejecting the null hypothesis
across various levels of significance (possibly for multiple hypothesis
tests, one for each feature).
}
\examples{
# generate example fit_results data for a feature selection problem
fit_results <- tibble::tibble(
  .rep = rep(1:2, times = 2),
  .dgp_name = c("DGP1", "DGP1", "DGP2", "DGP2"),
  .method_name = c("Method"),
  feature_info = lapply(
    1:4,
    FUN = function(i) {
      tibble::tibble(
        # feature names
        feature = c("featureA", "featureB", "featureC"),
        # true feature support
        true_support = c(TRUE, FALSE, TRUE),
        # estimated p-values
        pval = 10^(sample(-3:0, 3, replace = TRUE))
      )
    }
  )
)

# evaluate rejection probabilities for each feature across all possible values of alpha
eval_results <- eval_reject_prob(
  fit_results,
  nested_cols = "feature_info",
  feature_col = "feature",
  pval_col = "pval"
)

# evaluate rejection probability for each feature at specific values of alpha
eval_results <- eval_reject_prob(
  fit_results,
  nested_cols = "feature_info",
  feature_col = "feature",
  pval_col = "pval",
  alphas = c(0.05, 0.1)
)

}
\seealso{
Other inference_funs: 
\code{\link{eval_testing_curve_funs}},
\code{\link{eval_testing_err_funs}},
\code{\link{plot_reject_prob}()},
\code{\link{plot_testing_curve}()},
\code{\link{plot_testing_err}()}
}
\concept{inference_funs}
