% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_helpers.R
\name{evaluate_experiment}
\alias{evaluate_experiment}
\title{Evaluate an \code{Experiment}.}
\usage{
evaluate_experiment(
  experiment,
  fit_results,
  use_cached = FALSE,
  save = FALSE,
  verbose = 1,
  ...
)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{fit_results}{A tibble, as returned by the \code{fit} method.}

\item{use_cached}{Logical. If \code{TRUE}, find and return previously saved
results. If cached results cannot be found, continue as if
\code{use_cached} was \code{FALSE}.}

\item{save}{If \code{TRUE}, save outputs to disk.}

\item{verbose}{Level of verbosity. Default is 1, which prints out messages
after major checkpoints in the experiment. If 0, no messages are printed.}

\item{...}{Not used.}
}
\value{
A list of evaluation result tibbles, one for each
\code{Evaluator}.
}
\description{
Evaluate the performance of method(s) across all
\code{Evaluators} in the \code{Experiment} and return results.
}
