% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator.R
\docType{class}
\name{Evaluator}
\alias{Evaluator}
\title{\code{R6} class representing an evaluator.}
\description{
An evaluator (or evaluation function) to \strong{evaluate} the
performance of methods in the \code{Experiment}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the \code{Evaluator}.}

\item{\code{eval_fun}}{The evaluation function.}

\item{\code{eval_params}}{(Named) list of parameters to input into the
evaluation function.}

\item{\code{rmd_options}}{List of options to control the aesthetics of the
displayed \code{Evaluator}'s results table in the knitted R Markdown
report. See \code{\link[=prettyDT]{prettyDT()}} for possible options.}

\item{\code{show}}{If \code{TRUE} (default), show \code{Evaluator}'s results as
a table in the R Markdown report; if \code{FALSE}, hide output in the
R Markdown report.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Evaluator$new()}}
\item \href{#method-evaluate}{\code{Evaluator$evaluate()}}
\item \href{#method-clone}{\code{Evaluator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Evaluator}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evaluator$new(eval_fun, name = NULL, rmd_options = list(), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{eval_fun}}{The evaluation function.}

\item{\code{name}}{(Optional) The name of the \code{Evaluator}.}

\item{\code{rmd_options}}{(Optional) List of options to control the aesthetics
of the displayed \code{Evaluator}'s results table in the knitted R
Markdown report. See \code{\link[=prettyDT]{prettyDT()}} for possible options.}

\item{\code{...}}{Arguments to pass into \code{eval_fun()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
When evaluating or running the \code{Experiment} (see
\code{Experiment$evaluate() or \code{Experiment$run()}}), the named
arguments \code{fit_results} and \code{vary_params} are automatically
passed into the \code{Evaluator} function \code{eval_fun()} and serve
as placeholders for the \code{Experiment$fit()} results (i.e., the
results from the method fits) and the name of the varying parameter,
respectively. To evaluate the performance of a method(s) fit then,
the \code{Evaluator} function \code{eval_fun()} should almost always
take in the named argument \code{fit_results}. See
\code{Experiment$fit()} or \code{fit_experiment()} for details on the
format of \code{fit_results}. If the \code{Evaluator}
is used for \code{Experiments} with varying parameters,
\code{vary_params} should be used as a stand in for the name of this
varying parameter.
}

\subsection{Returns}{
A new \code{Evaluator} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-evaluate}{}}}
\subsection{Method \code{evaluate()}}{
Evaluate the performance of method(s) in the
\code{Experiment} using the \code{Evaluator} and its given parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evaluator$evaluate(fit_results, vary_params = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fit_results}}{A tibble, typically returned by the
\code{Experiment$fit()} method.}

\item{\code{vary_params}}{Name of parameters/arguments that were varied in the
\code{Experiment} (see \code{Experiment$get_vary_across()}).
Use \code{NULL} (default) if no \code{vary_across} component in
\code{Experiment} run.}

\item{\code{...}}{Not used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Result of \code{eval_fun()}, coerced into a tibble.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evaluator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
