% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\docType{class}
\name{Experiment}
\alias{Experiment}
\title{\code{R6} class representing a simulation experiment.}
\description{
A simulation experiment with any number of \code{DGPs}, \code{Methods},
\code{Evaluators}, and \code{Plotters}.
}
\details{
When run, an \code{Experiment} seamlessly combines \code{DGPs} and \code{Methods},
computing results in parallel. Those results can then be evaluated using
\code{Evaluators} and plotted using \code{Plotters}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the \code{Experiment}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Experiment$new()}}
\item \href{#method-run}{\code{Experiment$run()}}
\item \href{#method-generate_data}{\code{Experiment$generate_data()}}
\item \href{#method-fit}{\code{Experiment$fit()}}
\item \href{#method-evaluate}{\code{Experiment$evaluate()}}
\item \href{#method-plot}{\code{Experiment$plot()}}
\item \href{#method-create_doc_template}{\code{Experiment$create_doc_template()}}
\item \href{#method-create_rmd}{\code{Experiment$create_rmd()}}
\item \href{#method-add_dgp}{\code{Experiment$add_dgp()}}
\item \href{#method-update_dgp}{\code{Experiment$update_dgp()}}
\item \href{#method-get_dgps}{\code{Experiment$get_dgps()}}
\item \href{#method-add_method}{\code{Experiment$add_method()}}
\item \href{#method-update_method}{\code{Experiment$update_method()}}
\item \href{#method-get_methods}{\code{Experiment$get_methods()}}
\item \href{#method-add_evaluator}{\code{Experiment$add_evaluator()}}
\item \href{#method-update_evaluator}{\code{Experiment$update_evaluator()}}
\item \href{#method-get_evaluators}{\code{Experiment$get_evaluators()}}
\item \href{#method-add_plot}{\code{Experiment$add_plot()}}
\item \href{#method-update_plot}{\code{Experiment$update_plot()}}
\item \href{#method-get_plots}{\code{Experiment$get_plots()}}
\item \href{#method-add_vary_across}{\code{Experiment$add_vary_across()}}
\item \href{#method-update_vary_across}{\code{Experiment$update_vary_across()}}
\item \href{#method-remove_vary_across}{\code{Experiment$remove_vary_across()}}
\item \href{#method-get_vary_across}{\code{Experiment$get_vary_across()}}
\item \href{#method-get_save_dir}{\code{Experiment$get_save_dir()}}
\item \href{#method-set_save_dir}{\code{Experiment$set_save_dir()}}
\item \href{#method-clone}{\code{Experiment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Experiment}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$new(
  name = "experiment",
  dgp_list = list(),
  method_list = list(),
  evaluator_list = list(),
  plotter_list = list(),
  future.globals = NULL,
  future.packages = NULL,
  clone_from = NULL,
  save_dir = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the \code{Experiment}.}

\item{\code{dgp_list}}{An optional list of \code{DGP} objects.}

\item{\code{method_list}}{An optional list of \code{Method} objects.}

\item{\code{evaluator_list}}{An optional list of \code{Evaluator} objects.}

\item{\code{plotter_list}}{An option list of \code{Plotter} objects.}

\item{\code{future.globals}}{Passed as the argument of the same name to
code{future.apply::future_lapply} and related functions. See
\link{future.apply}{future_apply}. To set for a specific run of the
experiment, use the same argument in \code{Experiment$run}.}

\item{\code{future.packages}}{Passed as the argument of the same name to
code{future.apply::future_lapply} and related functions. See
\link{future.apply}{future_apply}. To set for a specific run of the
experiment, use the same argument in \code{Experiment$run}.}

\item{\code{save_dir}}{An optional directory in which to save the experiment's
results.}

\item{\code{...}}{Not used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Experiment} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
Run the entire simulation experiment pipeline.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$run(
  n_reps = 1,
  parallel_strategy = c("reps", "dgps", "methods"),
  future.globals = NULL,
  future.packages = NULL,
  use_cached = FALSE,
  save = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_reps}}{The number of replicates of the \code{Experiment} for this
run.}

\item{\code{parallel_strategy}}{A vector with some combination of "reps",
"dgps", or "methods". Determines how computation will be distributed
across available resources.}

\item{\code{future.globals}}{Passed as the argument of the same name to
code{future.apply::future_lapply} and related functions. See
\link{future.apply}{future_apply}. To set for all runs of the
experiment, use the same argument during initialization.}

\item{\code{future.packages}}{Passed as the argument of the same name to
code{future.apply::future_lapply} and related functions. See
\link{future.apply}{future_apply}. To set for all runs of the
experiment, use the same argument during initialization.}

\item{\code{use_cached}}{If \code{TRUE}, find and return previously saved
results.}

\item{\code{save}}{If \code{TRUE}, save results to disk.}

\item{\code{...}}{Not used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of results from the simulation experiment.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_data"></a>}}
\if{latex}{\out{\hypertarget{method-generate_data}{}}}
\subsection{Method \code{generate_data()}}{
Generate data from each of the \code{DGPs} in the
\code{Experiment}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$generate_data(n_reps = 1, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_reps}}{The number of datasets to generate per \code{DGP}.}

\item{\code{...}}{Passed to the generate.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of length equal to the number of \code{DGPs} in the
\code{Experiment}. Each element is a list of datasets of length
\code{n_reps}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Return fit results from the \code{Methods} in the
\code{Experiment}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$fit(
  n_reps = 1,
  parallel_strategy = "reps",
  future.globals = NULL,
  future.packages = NULL,
  use_cached = FALSE,
  save = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_reps}}{The number of replicates to run.}

\item{\code{parallel_strategy}}{A vector with some combination of "reps",
"dgps", or "methods". Determines how computation will be distributed
across available resources.}

\item{\code{future.globals}}{Character vector of names in the global environment
to pass to parallel workers. Passed as the argument of the same name to
code{future.apply::future_lapply} and related functions. To set for all
runs of the experiment, use the same argument during initialization.}

\item{\code{future.packages}}{Character vector of packages required by parallel
workers. Passed as the argument of the same name to
code{future.apply::future_lapply} and related functions.
To set for all runs of the experiment, use the same argument during
initialization.}

\item{\code{use_cached}}{If \code{TRUE}, find and return previously saved
results.}

\item{\code{save}}{If \code{TRUE}, save results to disk.}

\item{\code{...}}{Not used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of results from the simulation experiment.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-evaluate}{}}}
\subsection{Method \code{evaluate()}}{
Evaluate the \code{Experiment}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$evaluate(fit_results, use_cached = FALSE, save = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fit_results}}{A list of results, as returned by the \code{run}
method.}

\item{\code{use_cached}}{If \code{TRUE}, find and return previously saved
evaluation results.}

\item{\code{save}}{If \code{TRUE}, save evaluation results to disk.}

\item{\code{...}}{Not used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of evaluation results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Plot the simulation experiment's method results or
evaluation results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$plot(
  fit_results = NULL,
  eval_results = NULL,
  use_cached = FALSE,
  save = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fit_results}}{A list of results, as returned by the \code{fit}
method.}

\item{\code{eval_results}}{A list of results, as returned by the \code{evaluate}
method.}

\item{\code{use_cached}}{If \code{TRUE}, find and return previously saved
results.}

\item{\code{save}}{If \code{TRUE}, save plots to disk.}

\item{\code{...}}{Not used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of plots.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_doc_template"></a>}}
\if{latex}{\out{\hypertarget{method-create_doc_template}{}}}
\subsection{Method \code{create_doc_template()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$create_doc_template(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_rmd"></a>}}
\if{latex}{\out{\hypertarget{method-create_rmd}{}}}
\subsection{Method \code{create_rmd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$create_rmd(open = TRUE, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_dgp"></a>}}
\if{latex}{\out{\hypertarget{method-add_dgp}{}}}
\subsection{Method \code{add_dgp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$add_dgp(dgp, name = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_dgp"></a>}}
\if{latex}{\out{\hypertarget{method-update_dgp}{}}}
\subsection{Method \code{update_dgp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$update_dgp(dgp, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dgps"></a>}}
\if{latex}{\out{\hypertarget{method-get_dgps}{}}}
\subsection{Method \code{get_dgps()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_dgps()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_method"></a>}}
\if{latex}{\out{\hypertarget{method-add_method}{}}}
\subsection{Method \code{add_method()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$add_method(method, name = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_method"></a>}}
\if{latex}{\out{\hypertarget{method-update_method}{}}}
\subsection{Method \code{update_method()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$update_method(method, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_methods"></a>}}
\if{latex}{\out{\hypertarget{method-get_methods}{}}}
\subsection{Method \code{get_methods()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_methods()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_evaluator"></a>}}
\if{latex}{\out{\hypertarget{method-add_evaluator}{}}}
\subsection{Method \code{add_evaluator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$add_evaluator(evaluator, name = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_evaluator"></a>}}
\if{latex}{\out{\hypertarget{method-update_evaluator}{}}}
\subsection{Method \code{update_evaluator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$update_evaluator(evaluator, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_evaluators"></a>}}
\if{latex}{\out{\hypertarget{method-get_evaluators}{}}}
\subsection{Method \code{get_evaluators()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_evaluators()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_plot"></a>}}
\if{latex}{\out{\hypertarget{method-add_plot}{}}}
\subsection{Method \code{add_plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$add_plot(plotter, name = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_plot"></a>}}
\if{latex}{\out{\hypertarget{method-update_plot}{}}}
\subsection{Method \code{update_plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$update_plot(plotter, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_plots"></a>}}
\if{latex}{\out{\hypertarget{method-get_plots}{}}}
\subsection{Method \code{get_plots()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_plots()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_vary_across"></a>}}
\if{latex}{\out{\hypertarget{method-add_vary_across}{}}}
\subsection{Method \code{add_vary_across()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$add_vary_across(dgp = NULL, method = NULL, param_name, param_values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_vary_across"></a>}}
\if{latex}{\out{\hypertarget{method-update_vary_across}{}}}
\subsection{Method \code{update_vary_across()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$update_vary_across(
  dgp = NULL,
  method = NULL,
  param_name,
  param_values
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_vary_across"></a>}}
\if{latex}{\out{\hypertarget{method-remove_vary_across}{}}}
\subsection{Method \code{remove_vary_across()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$remove_vary_across()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_vary_across"></a>}}
\if{latex}{\out{\hypertarget{method-get_vary_across}{}}}
\subsection{Method \code{get_vary_across()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_vary_across()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_save_dir"></a>}}
\if{latex}{\out{\hypertarget{method-get_save_dir}{}}}
\subsection{Method \code{get_save_dir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_save_dir()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_save_dir"></a>}}
\if{latex}{\out{\hypertarget{method-set_save_dir}{}}}
\subsection{Method \code{set_save_dir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$set_save_dir(save_dir)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
