% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\docType{class}
\name{Experiment}
\alias{Experiment}
\title{\code{R6} class representing a simulation experiment.}
\description{
A simulation experiment with any number of \code{DGPs},
\code{Methods}, \code{Evaluators}, and \code{Visualizers}.
}
\details{
When run, an \code{Experiment} seamlessly combines \code{DGPs} and
\code{Methods}, computing results in parallel. Those results can then be
evaluated using \code{Evaluators} and visualized using \code{Visualizers}.
}
\section{Public fields}{


\describe{
\item{\code{name}}{The name of the \code{Experiment}.}
}
}

\section{Public methods}{


\itemize{
\item{\code{Experiment$new()}} See \link{create_experiment}.
\item{\code{Experiment$run()}} See \link{run_experiment}.
\item{\code{Experiment$generate_data()}} See \link{generate_data}.
\item{\code{Experiment$fit()}} See \link{fit_experiment}.
\item{\code{Experiment$evaluate()}} See \link{evaluate_experiment}.
\item{\code{Experiment$visualize()}} See \link{visualize_experiment}.
\item{\code{Experiment$create_doc_template()}} See \link{create_doc_template}.
\item{\code{Experiment$create_rmd()}} See \link{create_rmd}.
\item{\code{Experiment$add_dgp()}} See \link{add_dgp}.
\item{\code{Experiment$update_dgp()}} See \link{update_dgp}.
\item{\code{Experiment$remove_dgp()}} See \link{remove_dgp}.
\item{\code{Experiment$get_dgps()}} See \link{get_dgps}.
\item{\code{Experiment$add_method()}} See \link{add_method}.
\item{\code{Experiment$update_method()}} See \link{update_method}.
\item{\code{Experiment$remove_method()}} See \link{remove_method}.
\item{\code{Experiment$get_methods()}} See \link{get_methods}.
\item{\code{Experiment$add_evaluator()}} See \link{add_evaluator}.
\item{\code{Experiment$update_evaluator()}} See \link{update_evaluator}.
\item{\code{Experiment$remove_evaluator()}} See \link{remove_evaluator}.
\item{\code{Experiment$get_evaluators()}} See \link{get_evaluators}.
\item{\code{Experiment$add_visualizer()}} See \link{add_visualizer}.
\item{\code{Experiment$update_visualizer()}} See \link{update_visualizer}.
\item{\code{Experiment$remove_visualizer()}} See \link{remove_visualizer}.
\item{\code{Experiment$get_visualizers()}} See \link{get_visualizers}.
\item{\code{Experiment$add_vary_across()}} See \link{add_vary_across}.
\item{\code{Experiment$update_vary_across()}} See \link{update_vary_across}.
\item{\code{Experiment$remove_vary_across()}} See \link{remove_vary_across}.
\item{\code{Experiment$get_vary_across()}} See \link{get_vary_across}.
\item{\code{Experiment$clear_cache()}} See \link{clear_cache}.
\item{\code{Experiment$get_cache()}} See \link{get_cache}.
\item{\code{Experiment$set_rmd_options()}} See \link{set_rmd_options}.
\item{\code{Experiment$set_save_dir()}} See \link{set_save_dir}.
\item{\code{Experiment$save()}} See \link{save_experiment}.
}
}

