% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-helpers.R
\name{export_visualizers}
\alias{export_visualizers}
\title{Export cached \code{Visualizer} results to image.}
\usage{
export_visualizers(experiment, ...)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{...}{Additional arguments to pass to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}}
}
\value{
The original \code{Experiment} object passed to
\code{export_visualizers}.
}
\description{
Export all cached \code{Visualizer} results from an
\code{Experiment} to images in viz_results/ under the \code{Experiment}'s
results directory (see \code{Experiment$get_save_dir()}).
}
