% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-lib-utils.R
\name{finalize_estimator_internal_constructor}
\alias{finalize_estimator_internal_constructor}
\title{Helper function for constructing \code{finalize_estimator_internal} for
custom \code{yardstick} metrics that are restricted to binary estimators.}
\usage{
finalize_estimator_internal_constructor(metric_dispatcher, x, estimator)
}
\description{
Helper function for constructing \code{finalize_estimator_internal} for
custom \code{yardstick} metrics that are restricted to binary estimators.
}
\keyword{internal}
