% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{fit_experiment}
\alias{fit_experiment}
\title{Fit an \code{Experiment}.}
\usage{
fit_experiment(experiment, n_reps = 1, ...)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{n_reps}{The number of replicates of the \code{Experiment} for this run.}

\item{...}{Passed to \code{Experiment$fit()}.}
}
\value{
A tibble containing the results from fitting all \code{Methods}
across all \code{DGPs} for \code{n_reps} repetitions. In addition to
results columns, has columns named 'rep', 'dgp_name', 'method_name', and the
\code{vary_across} parameter names if applicable.
}
\description{
Fit \code{Methods} in the \code{Experiment} across all
\code{DGPs} for \code{n_reps} repetitions and return results from fits.
}
