% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_helpers.R
\name{fit_experiment}
\alias{fit_experiment}
\title{Fit an \code{Experiment}.}
\usage{
fit_experiment(
  experiment,
  n_reps = 1,
  parallel_strategy = c("reps"),
  future.globals = NULL,
  future.packages = NULL,
  future.seed = TRUE,
  use_cached = FALSE,
  save = FALSE,
  verbose = 1,
  ...
)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{n_reps}{The number of replicates of the \code{Experiment} for this run.}

\item{parallel_strategy}{A vector with some combination of "reps", "dgps", or
"methods". Determines how computation will be distributed across available
resources. Default is "reps".}

\item{future.globals}{Character vector of names in the global environment to
pass to parallel workers. Passed as the argument of the same name to
\code{future.apply::future_lapply} and related functions. To set for all
runs of the experiment, use the same argument during initialization.}

\item{future.packages}{Character vector of packages required by parallel
workers. Passed as the argument of the same name to
\code{future.apply::future_lapply} and related functions. To set for all
runs of the experiment, use the same argument during initialization.}

\item{future.seed}{Passed as the argument of the same name in
\code{future.apply::future_apply}.}

\item{use_cached}{Logical. If \code{TRUE}, find and return previously saved
results. If cached results cannot be found, continue as if
\code{use_cached} was \code{FALSE}.}

\item{save}{If \code{TRUE}, save outputs to disk.}

\item{verbose}{Level of verbosity. Default is 1, which prints out messages
after major checkpoints in the experiment. If 0, no messages are printed.}

\item{...}{Not used.}
}
\value{
A tibble containing the results from fitting all \code{Methods}
across all \code{DGPs} for \code{n_reps} repetitions. In addition to
results columns, has columns named '.rep', '.dgp_name', '.method_name', and the
\code{vary_across} parameter names if applicable.
}
\description{
Fit \code{Methods} in the \code{Experiment} across all
\code{DGPs} for \code{n_reps} repetitions and return results from fits.
}
