% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-lib-utils.R
\name{generate_coef}
\alias{generate_coef}
\title{Helper function to generate a coefficient vector.}
\usage{
generate_coef(.betas = NULL, .p = 1, .s = .p, .betas_name = "betas", ...)
}
\arguments{
\item{.betas}{Coefficient vector or function to generate the coefficients. If
a scalar is provided, the coefficient vector is a constant vector. If
\code{NULL} (default), \code{.s} non-zero entries in the coefficient vector
are equal to 0.5. If a function, must take one or more of \code{n},
\code{.s}, or \code{.p}, and can optionally take \code{.betas_name}, along
with other user-defined args from \code{...}) and must return a numeric
vector of length 1, \code{.s}, or \code{.p}. If the function take \code{n}
but not \code{.s} or \code{.p}, then \code{.s} is passed as \code{n}.}

\item{.p}{Number of features.}

\item{.s}{Sparsity level. Coefficients corresponding to features after the
\code{s}th position (i.e., positions i = \code{s} + 1, ..., \code{p}) are
set to 0.}

\item{.betas_name}{Name of coefficient variable to use in error message.}

\item{...}{Additional user arguments to pass to \code{.betas} when
\code{.betas} is a function.}
}
\value{
A vector of length \code{.p}.
}
\description{
Generate a coefficient vector with the specified dimensions and
sparsity level.
}
\examples{
# generate constant beta = 0.5 of dimension 10
beta <- generate_coef(.p = 10)

# generate beta ~ N(1, 2) of dimension 10
beta <- generate_coef(rnorm, .p = 10, mean = 1, sd = 2)

# generate beta = [1, 1, 0, 0, 0]
beta <- generate_coef(.betas = 1, .p = 5, .s = 2)

# generate beta = [1, 2, 3]
beta <- generate_coef(.betas = 1:3, .p = 3)

# use a function to generate beta
beta_fun <- function(.s, df) {
  return(rt(n = .s, df = df))
}
beta <- generate_coef(.betas = beta_fun, .p = 10, .s = 3, df = 10)

# we can do the same without wrapping rt
beta <- generate_coef(.betas = rt, .p = 10, .s = 3, df = 10)

}
