% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{generate_data}
\alias{generate_data}
\title{Evaluate an \code{Experiment}}
\usage{
generate_data(experiment, n_reps = 1, ...)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{n_reps}{The number of datasets to generate per \code{DGP}.}

\item{...}{Passed to \code{experiment$generate()}.}
}
\value{
A list of the simulation experiment's evaluation results.
}
\description{
Evaluate an \code{Experiment}
}
