% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{generate_data}
\alias{generate_data}
\title{Generate data from each \code{DGP} in the \code{Experiment}.}
\usage{
generate_data(experiment, n_reps = 1, ...)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{n_reps}{The number of datasets to generate per \code{DGP}.}

\item{...}{Passed to \code{experiment$generate_data()}.}
}
\value{
A list of length equal to the number of \code{DGPs} in the
\code{Experiment}. Each element in the list is a list of \code{n_reps}
datasets generated by the given \code{DGP}.
}
\description{
Generate data from each \code{DGP} in the \code{Experiment}.
}
