% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_helpers.R
\name{generate_data}
\alias{generate_data}
\title{Generate data from each \code{DGP} in the \code{Experiment}.}
\usage{
generate_data(experiment, n_reps = 1, ...)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{n_reps}{The number of datasets to generate per \code{DGP}.}

\item{...}{Not used.}
}
\value{
A list of length equal to the number of \code{DGPs} in the
\code{Experiment}. If the \code{Experiment} does not have a
\code{vary_across} component, then each element in the list is a list
of \code{n_reps} datasets generated by the given \code{DGP}. If the
\code{Experiment} does have a \code{vary_across} component, then each
element in the outermost list is a list of lists. The second layer of
lists corresponds to a specific parameter setting within the
\code{vary_across} scheme, and the innermost layer of lists is of
length \code{n_reps} with the dataset replicates, generated by the
\code{DGP}.
}
\description{
Generate data from each \code{DGP} in the \code{Experiment}.
}
