% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-lib-y.R
\name{generate_y_linear}
\alias{generate_y_linear}
\title{Simulate linear response data.}
\usage{
generate_y_linear(
  X,
  U,
  betas,
  betas_unobs,
  err = NULL,
  return_support = FALSE,
  ...
)
}
\arguments{
\item{X}{Design data matrix of observed variables.}

\item{U}{Design data matrix of unobserved (omitted) variables.}

\item{betas}{Coefficient vector for observed design matrix.}

\item{betas_unobs}{Coefficient vector for unobserved design matrix.}

\item{err}{Function from which to generate simulate error vector. Default
is \code{NULL} which adds no error to the response y.}

\item{return_support}{Logical specifying whether or not to return a vector
of the support column names. If \code{X} has no column names, then the
indices of the support are used.}

\item{...}{Other arguments to pass to err() to generate the error vector.}
}
\value{
A response vector of length nrow(X).
}
\description{
Generate linear response data with a specified error
distribution given the observed and unobserved design matrices.
}
