% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-lib-y.R
\name{generate_y_logistic}
\alias{generate_y_logistic}
\title{Simulate (binary) logistic response data.}
\usage{
generate_y_logistic(X, betas = 0, intercept = 0, return_support = FALSE, ...)
}
\arguments{
\item{X}{Design data matrix of observed variables.}

\item{betas}{Coefficient vector for observed design matrix.}

\item{intercept}{Scalar intercept term.}

\item{return_support}{Logical specifying whether or not to return a vector
of the support column names. If \code{X} has no column names, then the
indices of the support are used.}

\item{...}{Not used.}
}
\value{
If \code{return_support = TRUE}, returns a list of two:
\describe{
\item{y}{A response vector of length \code{nrow(X)}.}
\item{support}{A vector of feature indices in the true support of the DGP.}
}

If \code{return_support = FALSE}, returns only the response vector \code{y}.
}
\description{
Generate (binary) logistic response data given the observed
design matrices.
}
\examples{
X <- generate_X_gaussian(n = 100, p = 2)

# generate the response from: log(p / (1 - p)) = 3*x_1 - x_2
# where p = P(y  = 1 | x)
y <- generate_y_logistic(X = X, betas = c(3, -1))

}
