% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-lib-y.R
\name{generate_y_lss}
\alias{generate_y_lss}
\title{Generate locally spiky smooth (LSS) response data.}
\usage{
generate_y_lss(
  X,
  k,
  s,
  thresholds = 1,
  signs = 1,
  betas = 1,
  intercept = 0,
  overlap = FALSE,
  err = NULL,
  return_support = FALSE,
  ...
)
}
\arguments{
\item{X}{Data matrix or data frame.}

\item{k}{Order of the interactions.}

\item{s}{Number of interactions in the LSS model or a matrix of the support
indices with each interaction taking a row in this matrix and ncol = k.}

\item{thresholds}{A scalar or a s x k matrix of the thresholds for each term
in the LSS model.}

\item{signs}{A scalar or a s x k matrix of the sign of each interaction
(1 means > while -1 means <).}

\item{betas}{Scalar or parameter vector for interaction terms.}

\item{intercept}{Scalar intercept term.}

\item{overlap}{If TRUE, simulate support indices with replacement; if FALSE,
simulate support indices without replacement (so no overlap)}

\item{err}{Function from which to generate simulate error vector. Default
is \code{NULL} which adds no error to the DGP.}

\item{return_support}{Logical specifying whether or not to return a vector
of the support column names. If \code{X} has no column names, then the
indices of the support are used.}

\item{...}{Other arguments to pass to err() to generate the error vector.}
}
\value{
If \code{return_support = TRUE}, returns a list of two:
\describe{
\item{y}{A response vector of length \code{nrow(X)}.}
\item{support}{A vector of signed feature indices in the true (interaction)
support of the DGP. For example, "1+_2-" means that the interaction
between high values of feature 1 and low values of feature 2 appears in the
underlying DGP.}
}

If \code{return_support = FALSE}, returns only the response vector \code{y}.
}
\description{
Generate LSS response data with a specified error
distribution given the observed data matrices.
}
