% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_col_to_chr}
\alias{list_col_to_chr}
\title{Convert a list column to a readable character column.}
\usage{
list_col_to_chr(list_col, name = NULL, verbatim = FALSE)
}
\arguments{
\item{list_col}{A list-type column to be converted to character.}

\item{name}{Name of column. Used as a prefix in the returned character
strings. Default is \code{NULL}, which adds no prefix.}

\item{verbatim}{If \code{TRUE}, paste list contents together into a character
vector. If \code{FALSE} (default), map items in list to unique identifiers
(i.e., 1, 2, 3, ...).}
}
\value{
A character vector of the same length as \code{list_col}.
}
\description{
Convert a list-type column (typically in a tibble) to a
character-type column. Often useful for plotting along this column.
}
