% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-lib.R
\name{logistic_gaussian_dgp}
\alias{logistic_gaussian_dgp}
\title{Generate independent Gaussian covariates and (binary) logistic response data.}
\usage{
logistic_gaussian_dgp(
  n,
  p,
  s = p,
  betas = NULL,
  betas_sd = 1,
  intercept = 0,
  data_split = FALSE,
  train_prop = 0.5,
  return_values = c("X", "y", "support"),
  ...
)
}
\arguments{
\item{n}{Number of samples.}

\item{p}{Number of features.}

\item{s}{Sparsity level of features. Coefficients corresponding to features
after the \code{s} position (i.e., positions i = \code{s} + 1, ...,
\code{p}) are set to 0.}

\item{betas}{Coefficient vector for observed design matrix. If a scalar is provided, the coefficient vector is constant. If \code{NULL} (default), entries in the coefficient vector are drawn iid from N(0, \code{betas_sd}^2). Can also be a function that generates the coefficient vector; see \code{generate_coef()}.}

\item{betas_sd}{(Optional) SD of normal distribution from which to draw \code{betas}. Only used if \code{betas} argument is \code{NULL} or is a function in which case \code{betas_sd} is optionally passed to the function as \code{sd}; see \code{generate_coef()}.}

\item{intercept}{Scalar intercept term.}

\item{data_split}{Logical; if \code{TRUE}, splits data into training and test
sets according to \code{train_prop}.}

\item{train_prop}{Proportion of data in training set if \code{data_split =
TRUE}.}

\item{return_values}{Character vector indicating what objects to return in
list. Elements in vector must be one of "X", "y", "support".}

\item{...}{Not used.}
}
\value{
A list of the named objects that were requested in
\code{return_values}. See brief descriptions below.
\describe{
\item{X}{A \code{data.frame}.}
\item{y}{A response vector of length \code{nrow(X)}.}
\item{support}{A vector of feature indices indicating all features used in
the true support of the DGP.}
}
Note that if \code{data_split = TRUE} and "X", "y"
are in \code{return_values}, then the returned list also contains slots for
"Xtest" and "ytest".
}
\description{
Generate independent normally-distributed covariates and
logistic response data.
}
\details{
Data is generated via: \deqn{log(p / (1 - p)) = intercept +
betas \%*\% X,} where p = P(y = 1 | X), X is a standard Gaussian random matrix,
and the true underlying support of this data is the first s features in X
(unless specified otherwise by \code{betas}).
}
\examples{
# generate data from: log(p / (1 - p)) = betas_1 * x_1 + betas_2 * x_2, where
# betas_1, betas_2 ~ N(0, 1) and X ~ N(0, I_10)
sim_data <- logistic_gaussian_dgp(n = 100, p = 10, s = 2, betas_sd = 1)

}
