% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-lib.R
\name{lss_gaussian_dgp}
\alias{lss_gaussian_dgp}
\title{Generate independent Gaussian covariates and LSS response data.}
\usage{
lss_gaussian_dgp(
  n,
  p,
  k,
  s,
  thresholds = 0,
  signs = 1,
  betas = 1,
  intercept = 0,
  overlap = FALSE,
  err = NULL,
  data_split = FALSE,
  train_prop = 0.5,
  return_values = c("X", "y", "support"),
  ...
)
}
\arguments{
\item{n}{Number of samples.}

\item{p}{Number of features.}

\item{k}{Order of the interactions.}

\item{s}{Number of interactions in the LSS model or a matrix of the support
indices with each interaction taking a row in this matrix and ncol = k.}

\item{thresholds}{A scalar or a s x k matrix of the thresholds for each term
in the LSS model.}

\item{signs}{A scalar or a s x k matrix of the sign of each interaction (1
means > while -1 means <).}

\item{betas}{Scalar, vector, or function to generate coefficients
corresponding to interaction terms. See \\code{generate_coef()}.}

\item{intercept}{Scalar intercept term.}

\item{overlap}{If TRUE, simulate support indices with replacement; if FALSE,
simulate support indices without replacement (so no overlap)}

\item{err}{Function from which to generate simulated error vector. Default is
\code{NULL} which adds no error to the DGP.}

\item{data_split}{Logical; if \code{TRUE}, splits data into training and test
sets according to \code{train_prop}.}

\item{train_prop}{Proportion of data in training set if \code{data_split =
TRUE}.}

\item{return_values}{Character vector indicating what objects to return in
list. Elements in vector must be one of "X", "y", "support", "int_support".}

\item{...}{Other arguments to pass to err() to generate the error vector.}
}
\value{
A list of the named objects that were requested in
\code{return_values}. See brief descriptions below.
\describe{
\item{X}{A \code{data.frame}.}
\item{y}{A response vector of length \code{nrow(X)}.}
\item{support}{A vector of feature indices indicating all features used in
the true support of the DGP.}
\item{int_support}{A vector of signed feature indices in the true
(interaction) support of the DGP. For example, "1+_2-" means that the
interaction between high values of feature 1 and low values of feature 2
appears in the underlying DGP.}
}
Note that if \code{data_split = TRUE} and "X", "y"
are in \code{return_values}, then the returned list also contains slots for
"Xtest" and "ytest".
}
\description{
Generate independent normally-distributed covariates and LSS
response data with a specified error distribution.
}
\details{
Data is generated via: \deqn{y = intercept + sum_{i = 1}^{s} beta_i prod_{j = 1}^{k}1(X_{S_j} lessgtr thresholds_ij) + err(...),} where X is a
standard Gaussian random matrix. If \code{overlap = TRUE}, then the true
interaction support is randomly chosen from the p features in \code{X}. If
\code{overlap = FALSE}, then the true interaction support is sequentially
taken from the first \code{s*k} features in \code{X}.

For more details on the LSS model, see Behr, Merle, et al. "Provable Boolean Interaction Recovery from Tree Ensemble obtained via Random Forests." arXiv preprint arXiv:2102.11800 (2021).
}
\examples{
# generate data from: y = 1(X_1 > 0, X_2 > 0) + 1(X_3 > 0, X_4 > 0), where
# X is a 100 x 10 standard Gaussian random matrix
sim_data <- lss_gaussian_dgp(n = 100, p = 10, k = 2, s = 2,
                             thresholds = 0, signs = 1, betas = 1)

# generate data from: y = 3 * 1(X_1 < 0) - 1(X_2 > 1) + N(0, 1), where
# X is a 100 x 10 standard Gaussian random matrix
sim_data <- lss_gaussian_dgp(n = 100, p = 10, k = 1, s = 2,
                             thresholds = matrix(0:1, nrow = 2),
                             signs = matrix(c(-1, 1), nrow = 2),
                             betas = c(3, -1),
                             err = rnorm)

}
