% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\docType{class}
\name{Method}
\alias{Method}
\title{\code{R6} class representing a method.}
\description{
A method to be \strong{fit} and assessed in the \code{Experiment}.
}
\section{Public fields}{


\describe{
\item{\code{name}}{The name of the \code{Method}.}
\item{\code{method_fun}}{The data-generating process function.}
\item{\code{method_params}}{(Named) list of parameters to input into the data-generating process function.}
}
}

\section{Public methods}{


\itemize{
\item{\code{Method$new()}} See \link{create_method}.
\item{\code{Method$fit()}} See \link{fit_method}.
}
}

