% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer-lib.R
\name{plot_eval_summary}
\alias{plot_eval_summary}
\title{Plot summary of evaluation results.}
\usage{
plot_eval_summary(
  fit_results,
  eval_results,
  evaluator_name = NULL,
  eval_out = NULL,
  eval_id = "",
  eval_fun = paste0("summarize_", eval_id),
  vary_params = NULL,
  show = c("boxplot", "point", "line", "bar", "errorbar", "ribbon"),
  x_str = NULL,
  y_str = NULL,
  err_sd_str = NULL,
  color_str = NULL,
  facet_wrap_formula = NULL,
  facet_grid_formula = NULL,
  plot_by = NULL,
  add_ggplot_layers = NULL,
  boxplot_args = NULL,
  point_args = NULL,
  line_args = NULL,
  bar_args = NULL,
  errorbar_args = NULL,
  ribbon_args = NULL,
  facet_args = NULL,
  interactive = FALSE,
  ...
)
}
\arguments{
\item{fit_results}{A tibble, as returned by the \code{Experiment$fit()}
method.}

\item{eval_results}{A list of result tibbles, as returned by the
\code{Experiment$evaluate()} method.}

\item{evaluator_name}{(Optional) Name of evaluator containing results to
plot. If not provided, evaluation results will be automatically
computed by calling \code{eval_fun()}. If the evaluation results have
already been computed previously, should provide \code{evaluator_name} to
avoid duplicate computations.}

\item{eval_out}{(Optional) Tibble containing evaluation results to plot. If
provided, \code{eval_results} and \code{evaluator_name} are ignored, and
the tibble given by \code{eval_out} is used for plotting instead. This
argument is useful if you need to overwrite the tibble given by
\code{eval_results[[evaluator_name]]} with something else.}

\item{eval_id}{Character string. ID used for naming columns in
\code{summarize_eval_results()}. Should be the same as the \code{id}
input argument in \code{summarize_eval_results()}.}

\item{eval_fun}{Function used to compute evaluation results summary. This
function is only used (and required) if necessary results have not already
been computed in \code{eval_results[[evaluator_name]]}.}

\item{vary_params}{Vector of parameter names that are varied in the
\code{Experiment}'s run.}

\item{show}{Character vector with elements being one of "boxplot", "point",
"line", "bar", "errorbar", "ribbon" indicating what plot(s) to construct.}

\item{x_str}{(Optional) Name of column in evaluation results data frame to
plot on the x-axis.}

\item{y_str}{(Optional) Name of column in evaluation results data frame to
use for plotting on the y-axis if \code{show} is anything but "boxplot".}

\item{err_sd_str}{(Optional) Name of column in evaluation results data frame
containing the standard deviations of \code{y_str}.}

\item{color_str}{(Optional) Name of column in evaluation results data frame
to use for the color aesthetic when plotting.}

\item{facet_wrap_formula}{(Optional) Formula for \code{ggplot2::facet_wrap()}
if need be.}

\item{facet_grid_formula}{(Optional) Formula for \code{ggplot2::facet_grid()}
if need be.}

\item{plot_by}{(Optional) Name of column in evaluation results data frame
to use for subsetting and creating different plots for each unique value.}

\item{add_ggplot_layers}{Additional layers to add to a ggplot object via
\code{+}.}

\item{boxplot_args}{(Optional) Additional arguments to pass to
\code{ggplot2::geom_boxplot()}.}

\item{point_args}{(Optional) Additional arguments to pass to
\code{ggplot2::geom_point()}.}

\item{line_args}{(Optional) Additional arguments to pass to
\code{ggplot2::geom_line()}.}

\item{bar_args}{(Optional) Additional arguments to pass to
\code{ggplot2::geom_bar()}.}

\item{errorbar_args}{(Optional) Additional arguments to pass to
\code{ggplot2::geom_errorbar()}.}

\item{ribbon_args}{(Optional) Additional arguments to pass to
\code{ggplot2::geom_ribbon()}.}

\item{facet_args}{(Optional) Additional arguments to pass to
\code{ggplot2::facet_grid()} or \code{ggplot2::facet_wrap()}.}

\item{interactive}{Logical. Whether or not to return interactive
\code{plotly} plots instead of static \code{ggplot} plots.}

\item{...}{Additional arguments to pass to \code{eval_fun()}.}
}
\value{
A ggplot object or list of ggplot objects.
}
\description{
General helper function to plot summary of evaluation results
as a boxplot, scatter plot, line plot, or bar plot with or without 1 SD
error bars.
}
