% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer-lib-utils.R
\name{plot_eval_summary}
\alias{plot_eval_summary}
\title{Developer function for plotting summary of evaluation results.}
\usage{
plot_eval_summary(
  fit_results,
  eval_tib = NULL,
  eval_id = NULL,
  eval_fun = paste0("summarize_", eval_id),
  vary_params = NULL,
  show = c("boxplot", "point", "line", "bar", "errorbar", "ribbon"),
  x_str = NULL,
  y_str = NULL,
  err_sd_str = NULL,
  color_str = NULL,
  linetype_str = NULL,
  facet_wrap_formula = NULL,
  facet_grid_formula = NULL,
  plot_by = NULL,
  add_ggplot_layers = NULL,
  boxplot_args = NULL,
  point_args = NULL,
  line_args = NULL,
  bar_args = NULL,
  errorbar_args = NULL,
  ribbon_args = NULL,
  facet_args = NULL,
  interactive = FALSE,
  ...
)
}
\arguments{
\item{fit_results}{A tibble, as returned by the \code{fit} method.}

\item{eval_tib}{(Optional) \code{Tibble} (typically from the output of
\code{eval_summary_constructor}) containing the summarized evaluation
results to plot. If not provided, the evaluation results will be
automatically computed by calling \code{eval_fun()}. If the summarized
evaluation results have already been computed previously, \code{eval_tib}
should be specified to avoid duplicate computations.}

\item{eval_id}{Character string. ID used as the suffix for naming columns in
\code{eval_summary_constructor()}. Should be the same as the \code{eval_id}
argument in \code{eval_summary_constructor()}.}

\item{eval_fun}{Function used to compute evaluation results summary. This
function is only used (and required) if necessary results have not already
been computed in \code{eval_tib}.}

\item{vary_params}{A vector of parameter names that are varied across in the
\code{Experiment}.}

\item{show}{Character vector with elements being one of "boxplot", "point",
"line", "bar", "errorbar", "ribbon" indicating what plot layer(s) to
construct.}

\item{x_str}{(Optional) Name of column in \code{eval_tib} to plot on the
x-axis.}

\item{y_str}{(Optional) Name of column in \code{eval_tib} to plot on the
y-axis if \code{show} is anything but "boxplot".}

\item{err_sd_str}{(Optional) Name of column in \code{eval_tib} containing the
standard deviations of \code{y_str}. Used for plotting the errorbar and
ribbon ggplot layers.}

\item{color_str}{(Optional) Name of column in \code{eval_tib} to use for the
color and fill aesthetics when plotting.}

\item{linetype_str}{(Optional) Name of column in \code{eval_tib} to use for
the linetype aesthetic when plotting. Used only when \code{show = "line"}.}

\item{facet_wrap_formula}{(Optional) Formula for \code{ggplot2::facet_wrap()}
if need be.}

\item{facet_grid_formula}{(Optional) Formula for \code{ggplot2::facet_grid()}
if need be.}

\item{plot_by}{(Optional) Name of column in \code{eval_tib} to use for
subsetting data and creating different plots for each unique value.}

\item{add_ggplot_layers}{Additional layers to add to a ggplot object via
\code{+}.}

\item{boxplot_args}{(Optional) Additional arguments to pass into
\code{ggplot2::geom_boxplot()}.}

\item{point_args}{(Optional) Additional arguments to pass into
\code{ggplot2::geom_point()}.}

\item{line_args}{(Optional) Additional arguments to pass into
\code{ggplot2::geom_line()}.}

\item{bar_args}{(Optional) Additional arguments to pass into
\code{ggplot2::geom_bar()}.}

\item{errorbar_args}{(Optional) Additional arguments to pass into
\code{ggplot2::geom_errorbar()}.}

\item{ribbon_args}{(Optional) Additional arguments to pass into
\code{ggplot2::geom_ribbon()}.}

\item{facet_args}{(Optional) Additional arguments to pass into
\code{ggplot2::facet_grid()} or \code{ggplot2::facet_wrap()}.}

\item{interactive}{Logical. If \code{TRUE}, returns interactive \code{plotly}
plots. If \code{FALSE}, returns static \code{ggplot} plots.}

\item{...}{Additional arguments to pass to \code{eval_fun()}. This is only
used if necessary results have not already been computed in
\code{eval_tib}.}
}
\value{
If \code{interactive = TRUE}, returns a \code{plotly} object if
\code{plot_by} is \code{NULL} and a list of \code{plotly} objects if
\code{plot_by} is not \code{NULL}. If \code{interactive = FALSE}, returns
a \code{ggplot} object if \code{plot_by} is \code{NULL} and a list of
\code{ggplot} objects if \code{plot_by} is not \code{NULL}.
}
\description{
A helper function for developing new \code{Visualizer} plotting
functions that plot the summarized evaluation results as a boxplot,
scatter plot, line plot, or bar plot with or without 1 SD error
bars/ribbons. This function accepts either (1) a pre-computed tibble
containing the summarized evaluation results or (2) the Evaluator function
and its corresponding function arguments for computing the evaluation
results within this function call.
}
