% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{plot_experiment}
\alias{plot_experiment}
\title{Plot an \code{Experiment}'s method or evaluation results.}
\usage{
plot_experiment(experiment, ...)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{...}{Passed to \code{experiment$plot()}.}
}
\value{
A list of the simulation experiment's plots.
}
\description{
Plot an \code{Experiment}'s method or evaluation results.
}
