% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer-lib-feature-selection.R
\name{plot_feature_imp}
\alias{plot_feature_imp}
\title{Plot feature importances.}
\usage{
plot_feature_imp(
  fit_results,
  eval_results,
  evaluator_name = NULL,
  vary_params = NULL,
  max_features = NULL,
  feature,
  show = c("errorbar", "bar"),
  ...
)
}
\arguments{
\item{fit_results}{A tibble, as returned by the \code{fit} method.}

\item{eval_results}{A list of result tibbles, as returned by the
\code{evaluate} method.}

\item{evaluator_name}{Name of \code{Evaluator} containing results to plot.
To compute the evaluation summary results from scratch or if the evaluation
summary results have not yet been evaluated, set to \code{NULL}.}

\item{vary_params}{A vector of parameter names that are varied across in the
\code{Experiment}.}

\item{max_features}{Maximum number of features to plot.}

\item{feature}{A character string identifying the column with the feature
names.}

\item{show}{Character vector with elements being one of "boxplot", "point",
"line", "bar", "errorbar", "ribbon" indicating what plot layer(s) to
construct.}

\item{...}{Additional arguments to pass to \code{plot_eval_summary()}. This
includes arguments for plotting and for passing into
\code{summarize_feature_importances()}.}
}
\value{
If \code{interactive = TRUE}, returns a \code{plotly} object if
\code{plot_by} is \code{NULL} and a list of \code{plotly} objects if
\code{plot_by} is not \code{NULL}. If \code{interactive = FALSE}, returns
a \code{ggplot} object if \code{plot_by} is \code{NULL} and a list of
\code{ggplot} objects if \code{plot_by} is not \code{NULL}.
}
\description{
Plot the raw or summarized feature importances as a boxplot,
scatter plot, line plot, or bar plot with or without 1 SD error bars.
}
\seealso{
Other feature_selection_funs: 
\code{\link{eval_feature_importances}()},
\code{\link{eval_feature_recovery_curve}()},
\code{\link{eval_feature_recovery}()},
\code{\link{plot_feature_recovery_curve}()},
\code{\link{plot_feature_recovery}()},
\code{\link{summarize_feature_importances}()},
\code{\link{summarize_feature_recovery_curve}()},
\code{\link{summarize_feature_recovery}()}
}
\concept{feature_selection_funs}
