% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer-lib-prediction-metrics.R
\name{plot_pred_curve}
\alias{plot_pred_curve}
\title{Plot ROC/PR curves.}
\usage{
plot_pred_curve(
  fit_results,
  eval_results,
  evaluator_name = NULL,
  vary_params = NULL,
  metric = c("ROC", "PR"),
  show = c("line", "ribbon"),
  ...
)
}
\arguments{
\item{fit_results}{A tibble, as returned by the \code{fit} method.}

\item{eval_results}{A list of result tibbles, as returned by the
\code{evaluate} method.}

\item{evaluator_name}{Name of \code{Evaluator} containing results to plot.
To compute the evaluation summary results from scratch or if the evaluation
summary results have not yet been evaluated, set to \code{NULL}.}

\item{vary_params}{A vector of parameter names that are varied across in the
\code{Experiment}.}

\item{metric}{Either "ROC" or "PR" indicating whether to plot the ROC or
Precision-Recall curve.}

\item{show}{Character vector with elements being one of "boxplot", "point",
"line", "bar", "errorbar", "ribbon" indicating what plot layer(s) to
construct.}

\item{...}{Additional arguments to pass to \code{plot_eval_summary()}. This
includes arguments for plotting and for passing into
\code{summarize_pred_curve()}.}
}
\value{
If \code{interactive = TRUE}, returns a \code{plotly} object if
\code{plot_by} is \code{NULL} and a list of \code{plotly} objects if
\code{plot_by} is not \code{NULL}. If \code{interactive = FALSE}, returns
a \code{ggplot} object if \code{plot_by} is \code{NULL} and a list of
\code{ggplot} objects if \code{plot_by} is not \code{NULL}.
}
\description{
Plot ROC/PR curves or some summary thereof across experimental
replicates.
}
\seealso{
Other prediction_error_funs: 
\code{\link{eval_pred_curve}()},
\code{\link{eval_pred_err}()},
\code{\link{plot_pred_err}()},
\code{\link{summarize_pred_curve}()},
\code{\link{summarize_pred_err}()}
}
\concept{prediction_error_funs}
