% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer-lib-prediction.R
\name{plot_pred_err}
\alias{plot_pred_err}
\title{Plot prediction error according to various metrics.}
\usage{
plot_pred_err(
  fit_results,
  eval_results = NULL,
  evaluator_name = NULL,
  vary_params = NULL,
  metrics = NULL,
  show = c("point", "line"),
  ...
)
}
\arguments{
\item{fit_results}{A tibble, as returned by the \code{fit} method.}

\item{eval_results}{A list of result tibbles, as returned by the
\code{evaluate} method.}

\item{evaluator_name}{Name of \code{Evaluator} containing results to plot.
To compute the evaluation summary results from scratch or if the evaluation
summary results have not yet been evaluated, set to \code{NULL}.}

\item{vary_params}{A vector of parameter names that are varied across in the
\code{Experiment}.}

\item{metrics}{A \code{metric_set} object indicating the metrics to plot.
See \code{\link[yardstick:metric_set]{yardstick::metric_set()}} for more details. Default \code{NULL} will
use the default metrics in \code{\link[yardstick:metrics]{yardstick::metrics()}}.}

\item{show}{Character vector with elements being one of "boxplot", "point",
"line", "bar", "errorbar", "ribbon" indicating what plot layer(s) to
construct.}

\item{...}{Additional arguments to pass to \code{plot_eval_summary()}. This
includes arguments for plotting and for passing into
\code{summarize_pred_err()}.}
}
\value{
If \code{interactive = TRUE}, returns a \code{plotly} object if
\code{plot_by} is \code{NULL} and a list of \code{plotly} objects if
\code{plot_by} is not \code{NULL}. If \code{interactive = FALSE}, returns
a \code{ggplot} object if \code{plot_by} is \code{NULL} and a list of
\code{ggplot} objects if \code{plot_by} is not \code{NULL}.
}
\description{
Plot the raw or summarized prediction errors as a boxplot,
scatter plot, line plot, or bar plot with or without 1 SD error bars.
}
\examples{
# generate example fit_results data
fit_results <- tibble::tibble(
  .rep = rep(1:2, times = 2),
  .dgp_name = c("DGP1", "DGP1", "DGP2", "DGP2"),
  .method_name = c("Method"),
  # true response
  y = lapply(1:4, FUN = function(x) rnorm(100)),
  # predicted response
  predictions = lapply(1:4, FUN = function(x) rnorm(100))
)

# generate example eval_results data
eval_results <- list(
  `Prediction Errors` = summarize_pred_err(
    fit_results, truth_col = "y", estimate_col = "predictions"
  )
)

# create errorbar plot using pre-computed evaluation results
plt <- plot_pred_err(fit_results = fit_results, eval_results = eval_results,
                     evaluator_name = "Prediction Errors",
                     show = c("point", "errorbar"))
# or alternatively, create the same plot without pre-computing evaluation results
plt <- plot_pred_err(fit_results, show = c("point", "errorbar"),
                     truth_col = "y", estimate_col = "predictions")

# can customize plot (see plot_eval_summary() for possible arguments)
plt <- plot_pred_err(fit_results = fit_results, eval_results = eval_results,
                     evaluator_name = "Prediction Errors",
                     show = c("point", "errorbar"), 
                     color_str = NULL,
                     facet_formula = .method_name ~ .metric,
                     facet_type = "grid")

}
\seealso{
Other prediction_error_funs: 
\code{\link{eval_pred_curve_funs}},
\code{\link{eval_pred_err_funs}},
\code{\link{plot_pred_curve}()}
}
\concept{prediction_error_funs}
