% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{run_experiment}
\alias{run_experiment}
\title{Run the full \code{Experiment} pipeline.}
\usage{
run_experiment(experiment, n_reps = 1, ...)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{n_reps}{The number of replicates of the \code{Experiment} for this run.}

\item{...}{Passed to \code{experiment$run()}.}
}
\value{
A list of the simulation experiment's method results.
}
\description{
Run the full \code{Experiment} pipeline.
}
