% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_helpers.R
\name{run_experiment}
\alias{run_experiment}
\title{Run the full \code{Experiment} pipeline (fitting, evaluating, and visualizing).}
\usage{
run_experiment(
  experiment,
  n_reps = 1,
  parallel_strategy = c("reps"),
  future.globals = NULL,
  future.packages = NULL,
  future.seed = TRUE,
  use_cached = FALSE,
  save = FALSE,
  verbose = 1,
  ...
)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{n_reps}{The number of replicates of the \code{Experiment} for this run.}

\item{parallel_strategy}{A vector with some combination of "reps", "dgps", or
"methods". Determines how computation will be distributed across available
resources. Default is "reps".}

\item{future.globals}{Character vector of names in the global environment to
pass to parallel workers. Passed as the argument of the same name to
\code{future.apply::future_lapply} and related functions. To set for all
runs of the experiment, use the same argument during initialization.}

\item{future.packages}{Character vector of packages required by parallel
workers. Passed as the argument of the same name to
\code{future.apply::future_lapply} and related functions. To set for all
runs of the experiment, use the same argument during initialization.}

\item{future.seed}{Passed as the argument of the same name in
\code{future.apply::future_apply}.}

\item{use_cached}{Logical. If \code{TRUE}, find and return previously saved
results. If cached results cannot be found, continue as if
\code{use_cached} was \code{FALSE}.}

\item{save}{If \code{TRUE}, save outputs to disk.}

\item{verbose}{Level of verbosity. Default is 1, which prints out messages
after major checkpoints in the experiment. If 2, prints additional
debugging information for warnings and messages from user-defined functions
(in addition to error debugging information). If 0, no messages are printed
other than user-defined function error debugging information.}

\item{...}{Not used.}
}
\value{
A list of results from the simulation experiment.
\describe{
\item{fit_results}{A tibble containing results from the \code{fit}
method. In addition to results columns, has columns named '.rep', '.dgp_name',
'.method_name', and the \code{vary_across} parameter names if applicable.}
\item{eval_results}{A list of tibbles containing results from the
\code{evaluate} method, which evaluates each \code{Evaluator} in
the \code{Experiment}. Length of list is equivalent to the number of
\code{Evaluators}.}
\item{viz_results}{A list of tibbles containing results from the
\code{visualize} method, which visualizes each \code{Visualizer} in
the \code{Experiment}. Length of list is equivalent to the number of
\code{Visualizers}.}
}
}
\description{
Run the full \code{Experiment} pipeline (fitting, evaluating, and visualizing).
}
