% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-helpers.R
\name{set_rmd_options}
\alias{set_rmd_options}
\title{Set R Markdown options for \code{Evaluator} and \code{Visualizer} outputs in
summary report.}
\usage{
set_rmd_options(
  experiment,
  field_name = c("evaluator", "visualizer"),
  name,
  show = NULL,
  ...
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{set_rmd_options()} was renamed to \code{set_doc_options()} to create a more
consistent API.
}
\keyword{internal}
