% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{set_save_dir}
\alias{set_save_dir}
\title{Set results directory for an \code{Experiment}.}
\usage{
set_save_dir(experiment, save_dir)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{save_dir}{The directory in which the \code{Experiment}'s results
will be saved.}
}
\value{
The original \code{Experiment} object with the updated saving
directory.
}
\description{
Set the directory in which the \code{Experiment}'s results and
plots are saved.
}
