% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-lib-utils.R
\name{shared_dgp_lib_args}
\alias{shared_dgp_lib_args}
\title{Arguments that are shared by multiple \code{DGP} library functions.}
\arguments{
\item{betas}{Coefficient vector for observed design matrix. If a scalar is provided, the coefficient vector is constant. If \code{NULL} (default), entries in the coefficient vector are drawn iid from N(0, \code{betas_sd}^2). Can also be a function that generates the coefficient vector; see \code{generate_coef()}.}

\item{betas_sd}{(Optional) SD of normal distribution from which to draw \code{betas}. Only used if \code{betas} argument is \code{NULL} or is a function in which case \code{betas_sd} is optionally passed to the function as \code{sd}; see \code{generate_coef()}.}

\item{betas_unobs}{Coefficient vector for unobserved design matrix. If a scalar is provided, the coefficient vector is constant. If \code{NULL} (default), entries in the coefficient vector are drawn iid from N(0, \code{betas_unobs_sd}^2). Can also be a function that generates the coefficient vector; see \code{generate_coef()}.}

\item{betas_unobs_sd}{(Optional) SD of normal distribution from which to draw \code{betas_unobs}. Only used if \code{betas_unobs} argument is \code{NULL} or is a function in which case \code{betas_unobs_sd} is optionally passed to the function as \code{sd}; see \code{generate_coef()}.}

\item{betas_corr}{Coefficient vector for correlated features. If a scalar is provided, the coefficient vector is constant. If \code{NULL} (default), entries in the coefficient vector are drawn iid from N(0, \code{betas_corr_sd}^2). Can also be a function that generates the coefficient vector; see \code{generate_coef()}.}

\item{betas_corr_sd}{(Optional) SD of normal distribution from which to draw \code{betas_corr}. Only used if \code{betas_corr} argument is \code{NULL} or is a function in which case \code{betas_corr_sd} is optionally passed to the function as \code{sd}; see \code{generate_coef()}.}

\item{betas_uncorr}{Coefficient vector for uncorrelated features. If a scalar is provided, the coefficient vector is constant. If \code{NULL} (default), entries in the coefficient vector are drawn iid from N(0, \code{betas_uncorr_sd}^2). Can also be a function that generates the coefficient vector; see \code{generate_coef()}.}

\item{betas_uncorr_sd}{(Optional) SD of normal distribution from which to draw \code{betas_uncorr}. Only used if \code{betas_uncorr} argument is \code{NULL} or is a function in which case \code{betas_uncorr_sd} is optionally passed to the function as \code{sd}; see \code{generate_coef()}.}

\item{data_split}{Logical; if \code{TRUE}, splits data into training and test
sets according to \code{train_prop}.}

\item{err}{Function from which to generate simulated error vector. Default is
\code{NULL} which adds no error to the DGP.}

\item{intercept}{Scalar intercept term.}

\item{n, .n}{Number of samples.}

\item{p, .p}{Number of features.}

\item{return_support}{Logical specifying whether or not to return a vector of
the support column names. If \code{X} has no column names, then the indices
of the support are used.}

\item{return_values}{Character vector indicating what objects to return in
list. Elements in vector must be one of "X", "y", "support".}

\item{support}{Vector of feature indices in the true support of the DGP.}

\item{train_prop}{Proportion of data in training set if \code{data_split =
TRUE}.}

\item{X}{Data matrix or data frame.}

\item{y}{Response vector.}

\item{...}{Other arguments to pass to err() to generate the error vector.}
}
\value{
A list of the named objects that were requested in
\code{return_values}. See brief descriptions below.
\describe{
\item{X}{A \code{data.frame}.}
\item{y}{A response vector of length \code{nrow(X)}.}
\item{support}{A vector of feature indices indicating all features used in
the true support of the DGP.}
}
Note that if \code{data_split = TRUE} and "X", "y"
are in \code{return_values}, then the returned list also contains slots for
"Xtest" and "ytest".
}
\description{
Arguments that are shared by multiple \code{DGP} library functions.
}
\keyword{internal}
