% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-lib-utils.R
\name{shared_eval_lib_args}
\alias{shared_eval_lib_args}
\title{Arguments that are shared by multiple \code{Evaluator} library functions.}
\arguments{
\item{custom_summary_funs}{Named list of custom functions to summarize
results. Names in the list should correspond to the name of the summary
function. Values in the list should be a function that takes in one
argument, that being the values of the evaluated metrics.}

\item{metric}{Either "ROC" or "PR" indicating whether to evaluate the ROC or
Precision-Recall curve.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA} values
should be stripped before the computation proceeds.}

\item{summary_funs}{Character vector specifying how to summarize
evaluation metrics. Must choose from a built-in library of summary
functions - elements of the vector must be one of "mean", "median",
"min", "max", "sd", "raw".}
}
\description{
Arguments that are shared by multiple \code{Evaluator} library functions.
}
