% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_helpers.R
\name{shared_experiment_helpers_args}
\alias{shared_experiment_helpers_args}
\title{Arguments that are shared by multiple Experiment helper funs.}
\arguments{
\item{checkpoint_n_reps}{The number of experiment replicates to compute
before saving results to disk. If 0 (the default), no checkpoints are
saved.}

\item{dgp}{A \code{DGP} object.}

\item{evaluator}{An \code{Evaluator} object.}

\item{eval_results}{A list of result tibbles, as returned by the
\code{evaluate} method.}

\item{experiment}{An \code{Experiment} object.}

\item{fit_results}{A tibble, as returned by the \code{fit} method.}

\item{future.globals}{Character vector of names in the global environment to
pass to parallel workers. Passed as the argument of the same name to
\code{future.apply::future_lapply} and related functions. To set for all
runs of the experiment, use the same argument during initialization.}

\item{future.packages}{Character vector of packages required by parallel
workers. Passed as the argument of the same name to
\code{future.apply::future_lapply} and related functions. To set for all
runs of the experiment, use the same argument during initialization.}

\item{future.seed}{Passed as the argument of the same name in
\code{future.apply::future_apply}.}

\item{method}{A \code{Method} object.}

\item{n_reps}{The number of replicates of the \code{Experiment} for this run.}

\item{parallel_strategy}{A vector with some combination of "reps", "dgps", or
"methods". Determines how computation will be distributed across available
resources. Default is "reps".}

\item{save}{If \code{TRUE}, save outputs to disk.}

\item{use_cached}{Logical. If \code{TRUE}, find and return previously saved
results. If cached results cannot be found, continue as if
\code{use_cached} was \code{FALSE}.}

\item{vary_params}{A vector of parameter names that are varied across in the
\code{Experiment}.}

\item{verbose}{Level of verbosity. Default is 1, which prints out messages
after major checkpoints in the experiment. If 2, prints additional
debugging information for warnings and messages from user-defined functions
(in addition to error debugging information). If 0, no messages are printed
other than user-defined function error debugging information.}

\item{visualizer}{A \code{Visualizer} object.}

\item{...}{Not used.}
}
\description{
Arguments that are shared by multiple Experiment helper funs.
}
