% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer-lib-utils.R
\name{shared_viz_lib_args}
\alias{shared_viz_lib_args}
\title{Arguments that are shared by multiple \code{Visualizer} library functions.}
\arguments{
\item{evaluator_name}{Name of \code{Evaluator} containing results to plot.
To compute the evaluation summary results from scratch or if the evaluation
summary results have not yet been evaluated, set to \code{NULL}.}

\item{interactive}{Logical. If \code{TRUE}, returns interactive \code{plotly}
plots. If \code{FALSE}, returns static \code{ggplot} plots.}

\item{curve}{Either "ROC" or "PR" indicating whether to plot the ROC or
Precision-Recall curve.}

\item{show}{Character vector with elements being one of "boxplot", "point",
"line", "bar", "errorbar", "ribbon" indicating what plot layer(s) to
construct.}
}
\description{
Arguments that are shared by multiple \code{Visualizer} library functions.
}
