% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simplify_tibble}
\alias{simplify_tibble}
\title{Simplify tibble.}
\usage{
simplify_tibble(tbl, cols = NULL, empty_as_na = TRUE)
}
\arguments{
\item{tbl}{\code{tibble::tibble} to simplify.}

\item{cols}{Character vector of column names to simplify. If NULL (default),
all columns are eligible to be simplified.}

\item{empty_as_na}{If TRUE (default), 0-length values will be treated as NA.}
}
\value{
A tibble that has been "simplified".
}
\description{
Simplify or unlist list columns in tibble if each element in
the list is a scalar value.
}
\keyword{internal}
