% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{simplify_tibble}
\alias{simplify_tibble}
\title{Simplify tibble.}
\usage{
simplify_tibble(tib, omit_cols = NULL)
}
\arguments{
\item{tib}{Tibble to simplify.}

\item{omit_cols}{Character vector of column names to omit when simplifying.}
}
\value{
A tibble that has been "simplified".
}
\description{
Simplify or unlist list columns in tibble if each element in
the list is a scalar value.
}
\keyword{internal}
