% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{update_funs}
\alias{update_funs}
\alias{update_dgp}
\alias{update_method}
\alias{update_evaluator}
\alias{update_visualizer}
\title{Helper functions for updating components of an \code{Experiment}.}
\usage{
update_dgp(experiment, dgp, name, ...)

update_method(experiment, method, name, ...)

update_evaluator(experiment, evaluator, name, ...)

update_visualizer(experiment, visualizer, name, ...)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{dgp}{A \code{DGP} object.}

\item{name}{A name to identify the object to be updated.}

\item{...}{Not used.}

\item{method}{A \code{Method} object.}

\item{evaluator}{An \code{Evaluator} object.}

\item{visualizer}{A \code{Visualizer} object.}
}
\value{
The original \code{Experiment} object passed to \code{update_*}.
}
\description{
Helper functions for updating \code{DGPs}, \code{Methods},
\code{Evaluators}, and \code{Visualizers} already added to an
\code{Experiment}.
}
