% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{vary_across}
\alias{vary_across}
\alias{add_vary_across}
\alias{update_vary_across}
\alias{remove_vary_across}
\title{Varying across parameters in an \code{Experiment}.}
\usage{
add_vary_across(experiment, ...)

update_vary_across(experiment, ...)

remove_vary_across(experiment, ...)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{...}{Any number of named arguments where names match an argument in the
user-specified \code{DGP} or \code{Method} function and values are vectors
(for scalar parameters) or lists (for arbitrary parameters).}

\item{dgp}{Name of \code{DGP} to vary in the \code{Experiment}. Can also be a
\code{DGP} object that matches one in the \code{Experiment}. Must provide
either the \code{dgp} or \code{method} argument.}

\item{method}{Name of \code{Method} to vary in the \code{Experiment}. Can
also be a \code{Method} object that matches one in the \code{Experiment}.
Must provide either the \code{dgp} or \code{method} argument.}

\item{param_names}{A character vector of parameter names to remove. If
not provided, the entire set of \code{vary_across} parameters will be
removed for the specified \code{DGP}/\code{Method}.}
}
\value{
The original \code{Experiment} object passed to
\code{*_vary_across()}.
}
\description{
Helper functions for adding, updating, or removing a
\code{vary_across} component in an \code{Experiment}. When a
\code{vary_across} component is added and the \code{Experiment} is run, the
\code{Experiment} is systematically varied across values of the specified
parameter in the \code{DGP} or \code{Method} while all other parameters are
held constant at their baseline value.
}
