% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_helpers.R
\name{visualize_experiment}
\alias{visualize_experiment}
\title{Visualize results of an \code{Experiment}.}
\usage{
visualize_experiment(
  experiment,
  fit_results,
  eval_results = NULL,
  use_cached = FALSE,
  save = FALSE,
  verbose = 1,
  ...
)
}
\arguments{
\item{experiment}{An \code{Experiment} object.}

\item{fit_results}{A tibble, as returned by the \code{fit} method.}

\item{eval_results}{A list of result tibbles, as returned by the
\code{evaluate} method.}

\item{use_cached}{Logical. If \code{TRUE}, find and return previously saved
results. If cached results cannot be found, continue as if
\code{use_cached} was \code{FALSE}.}

\item{save}{If \code{TRUE}, save outputs to disk.}

\item{verbose}{Level of verbosity. Default is 1, which prints out messages
after major checkpoints in the experiment. If 0, no messages are printed.}

\item{...}{Not used.}
}
\value{
A list of visualizations, one for each \code{Visualizer}.
}
\description{
Visualize the performance of methods and/or its evaluation metrics
using all \code{Visualizers} in the \code{Experiment} and return visualization results.
}
