% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer.R
\docType{class}
\name{Visualizer}
\alias{Visualizer}
\title{\code{R6} class representing a visualizer.}
\description{
A visualizer (or visualization function) to \strong{visualize} the performance of
methods and/or its evaluation metrics from the \code{Experiment} run.
}
\section{Public fields}{


\describe{
\item{\code{name}}{The name of the \code{Visualizer}.}
\item{\code{viz_fun}}{The visualizer function.}
\item{\code{viz_params}}{(Named) list of parameters to input into the visualizer function.}
\item{\code{doc_options}}{List of options to control the aesthetics of the
displayed \code{Visualizer}'s visualizations in the knitted R Markdown
report. Currently, possible options are "height" and "width" (in inches).}
\item{\code{doc_show}}{If \code{TRUE} (default), show the resulting visualization
in the R Markdown report; if \code{FALSE}, hide output in the
R Markdown report.}
}
}

\section{Public methods}{


\itemize{
\item{\code{Visualizer$new()}} See \link{create_visualizer}.
\item{\code{Visualizer$visualize()}} See \link{visualize_visualizer}.
}
}

