% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer.R
\docType{class}
\name{Visualizer}
\alias{Visualizer}
\title{\code{R6} class representing a visualizer.}
\description{
A visualizer (or visualization function) to \strong{visualize} the performance of
methods and/or its evaluation metrics from the \code{Experiment} run.
}
\section{Public fields}{


\describe{
\item{\code{name}}{The name of the \code{Visualizer}.}
\item{\code{viz_fun}}{The visualizer function.}
\item{\code{viz_params}}{(Named) list of parameters to input into the visualizer function.}
\item{\code{doc_options}}{List of options to control the aesthetics of the
displayed \code{Visualizer}'s visualizations in the knitted R Markdown
report. Currently, possible options are "height" and "width" (in inches).}
\item{\code{doc_show}}{If \code{TRUE} (default), show the resulting visualization
in the R Markdown report; if \code{FALSE}, hide output in the
R Markdown report.}
}
}

\section{Public methods}{


\itemize{
\item{\code{Visualizer$new()}} See \link{create_visualizer}.
\item{\code{Visualizer$visualize()}} See \link{visualize_visualizer}.
}
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Visualizer-new}{\code{Visualizer$new()}}
\item \href{#method-Visualizer-visualize}{\code{Visualizer$visualize()}}
\item \href{#method-Visualizer-print}{\code{Visualizer$print()}}
\item \href{#method-Visualizer-clone}{\code{Visualizer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer-new"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer$new(
  .viz_fun,
  .name = NULL,
  .doc_options = list(),
  .doc_show = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer-visualize"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer-visualize}{}}}
\subsection{Method \code{visualize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer$visualize(
  fit_results = NULL,
  eval_results = NULL,
  vary_params = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer-print"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
